Compute the reliability of a double difference score

This function uses Lord's (1963) algorithm to quickly compute the reliability of a double mean difference score. Missing values are dealt with using multiple imputation.

aat_intercategorical_alpha(
  ds,
  subjvar,
  stimvar,
  pullvar,
  targetvar,
  rtvar,
  iters = 5
)

Arguments

ds
the data.frame to use

subjvar
Name of the subject-identifying variable

stimvar
Name of the stimulus-identifying variable

pullvar
Name of the movement-direction identifying variable

targetvar
Name of the stimulus-category identifying variable

rtvar
Name of the reaction-time identifying variable

iters
If there are missing values (which is almost inevitable) then multiple imputation will be used to complete the covariance matrix - this option sets the number of multiple imputations to be used.

Value

Returns an aat_intercategorical_alpha object containing the alpha value as well as the dataset and covariance matrix with replaced missing values.

Details

When the study in question uses double median difference scores or D-scores, aat_splithalf() is recommended instead; and if only single mean difference scores are used, aat_alpha suffices.

References

Lord, F.Y. (1963), "Elementary Models for Measuring Change", in Problems in Measuring Change, C.W. Harris, ed.. Madison. Wisconsin: University of Wisconsin.

Examples

#We generate a dataset with 16 stimuli in each category
ds<-aat_simulate(biasfx_jitter=40,nstims=16)
ds$stim<-paste0(ds$stim,"-",ds$is_target)

#If the intercategorical alpha and
# bootstrapped splithalf measure something similar,
# then the outcomes should be close to each other.
aat_intercategorical_alpha(ds=ds,subjvar="subj",stimvar="stim",pullvar="is_pull",
                           targetvar="is_target",rtvar="rt")

#> $alpha
#> [1] 0.2060208
#> 
#> $data
#>                      0           1           0           1           0
#> subject.1   110.918423  195.317062  103.995682  147.864832  141.864675
#> subject-2   -60.687866  -23.815907 -111.648138  196.539904 -110.177260
#> subject-3    58.622327   83.516515   15.678835  -69.162408   -7.664995
#> subject-4    81.183755 -223.245164   70.709943   88.664057  -11.756566
#> subject-5    21.543436   77.930130  -72.055170   54.874604   97.394080
#> subject-6  -139.510213   24.366470   -4.430791  274.570314  -13.425277
#> subject-7   -91.083984   17.553005   84.218950  -85.477097  -53.816057
#> subject-8    90.976305    3.525609  145.268105  -90.617264   27.052034
#> subject-9   -26.306042  -98.267190  104.085542  204.235068   31.991329
#> subject-10 -468.749525  191.424381  -29.996362  114.243824  238.338054
#> subject-11  -21.312048  205.513731    2.733338   63.471373   24.162597
#> subject-12   45.049440  204.784999  -75.790595    9.973226  145.213635
#> subject-13 -348.628217  247.761351  124.853676  -66.672751   23.998475
#> subject-14  122.074543   92.416221  -40.354762    8.814242  219.301383
#> subject-15   23.465884  -49.154495 -139.680688  161.234772   97.262898
#> subject-16  215.304117   49.829647  100.270930  244.053172  108.248926
#> subject-17   21.724548  268.859138  -29.036678  162.257104  133.583906
#> subject-18 -183.366969  167.492258  245.122994  225.674389  342.667752
#> subject-19  -34.512448  -36.433350 -159.107466 -277.855665  -61.252897
#> subject-20  -43.396316  227.750563  -49.216492  -80.516787   -2.009588
#> subject-21   24.612695  -85.836847   74.244705   17.112737  -99.570580
#> subject-22  151.636626   68.530947  150.876613   -9.538657   69.206529
#> subject-23  167.801012  -88.328796  183.382212  156.883497  -30.529157
#> subject-24   77.935780 -266.493300  -84.009587  109.840737 -309.283400
#> subject-25  181.859122 -373.747322 -218.579182  -14.965862 -148.012542
#> subject-26   70.907010  -97.019434  157.376311  306.426978 -138.135819
#> subject-27  -98.904585 -104.634001  122.350611   51.458616  143.094248
#> subject-28   62.470039 -109.997033  -41.886343   54.567772   86.530829
#> subject-29   31.176307  -28.099518 -156.072605  152.206138  -43.009794
#> subject-30 -277.164544  220.083970 -115.800137    6.436435  -27.281483
#> subject-31  -31.151320   97.006227  187.236474  378.255253  -41.421472
#> subject-32   50.632994  227.466678  143.305145  111.524753  -34.847061
#> subject-33    4.058040  268.641938  -33.168272   79.677235    7.808101
#> subject-34  113.164591  -30.451425  257.319435  171.034304  207.241751
#> subject-35  -95.781041 -175.466803   31.055499   53.842981  -35.493480
#> subject-36   -4.731014  440.205847  213.527614  353.741028  259.578718
#> subject-37 -104.867198  149.845311 -148.956444  110.408272   16.292785
#> subject-38   87.511653 -189.169574   77.789410   89.873144   -9.488058
#> subject-39  201.693899  202.381972   62.761398  197.526645  225.637054
#> subject-40   27.885812   43.201339 -129.257151 -150.692833   17.755476
#>                      1           0             1           0           1
#> subject.1   294.088814  160.959582  2.307377e+02    9.380715  216.134509
#> subject-2  -235.686805  -81.854497  1.357041e+02  163.244119  -85.684233
#> subject-3    72.469436   41.106790  1.395660e+02   38.782575  -57.054578
#> subject-4   -76.000687   31.700563 -1.869285e+01  -13.078126  169.739389
#> subject-5   -34.139540  475.274665  5.843245e+01   17.282467  -20.659874
#> subject-6   273.381248  -72.856802  1.069424e+02   -3.897784   52.543611
#> subject-7    47.947467 -378.081713  2.495450e+02  -75.192932  -48.685362
#> subject-8   -49.639711  124.960898  1.615680e+00   16.960441  128.056606
#> subject-9   105.967605  -60.690890 -6.594308e+01   51.403125 -108.133035
#> subject-10  -91.684655 -480.704166  1.458767e+02 -112.036165  -61.418019
#> subject-11  123.966921  148.465546  2.119459e+02   38.424922  226.578130
#> subject-12  123.325051    4.922200  8.738222e+00  158.253018   49.354639
#> subject-13 -140.638843 -236.484925  1.111568e+02  -40.645278 -106.060173
#> subject-14  -85.722567   64.567403  9.185706e+01  -29.094772  119.210640
#> subject-15 -169.797929  -81.546559 -1.762829e+02   72.011089  -23.171421
#> subject-16  -28.544283  -21.937787  1.489721e+02  -49.184549  147.261011
#> subject-17  123.425651   77.771683  5.063145e+01 -126.290635   75.645013
#> subject-18 -245.977449  242.796400  8.549915e+01   67.325992   -5.588940
#> subject-19 -142.094805 -151.619022 -1.042746e+02 -232.950733 -248.046079
#> subject-20   80.382542 -139.029665 -5.992526e-03   -5.904470   62.657906
#> subject-21  201.123607   -8.719236  1.539620e+02   55.263056   46.311548
#> subject-22  248.832656 -177.749983  3.575668e+02  129.215011  277.737062
#> subject-23   77.173513   23.413644  2.884810e+01  235.502591 -269.129909
#> subject-24 -169.909612   87.267550  1.700803e+02 -288.517619  205.128783
#> subject-25  293.213900  399.636925 -3.410715e+01   65.814636 -296.656801
#> subject-26   -6.061398  186.654621  5.794225e+01  191.391452  275.067062
#> subject-27   -2.394674  -31.807679  7.945472e+01  -47.068080  -60.493702
#> subject-28  -24.974463  -32.557501  8.006570e+01   67.622260   32.250722
#> subject-29  -80.733050  -20.336153 -2.126715e+01  -65.510126   68.256509
#> subject-30  -71.410630  -67.918700 -1.768569e+02 -245.012776  -19.068682
#> subject-31   70.274678  -46.907068  3.185009e+02   94.977031   13.371378
#> subject-32  181.429712  -75.636310  9.605728e+01   16.589219  153.805991
#> subject-33  -73.859054   55.546600  1.452523e+02  128.323854   92.485373
#> subject-34   -3.956324  180.123697  2.075160e+02   95.838653  364.623618
#> subject-35  -66.050394   31.910401  2.908981e+02   78.786529   -7.441801
#> subject-36 -283.752997   21.395332  5.156096e+01   55.674966  509.538167
#> subject-37    7.641756   32.215790  1.351524e+02  146.332471  -24.888196
#> subject-38  -76.706089  228.538858  4.773080e+01   60.210676  114.779720
#> subject-39   97.615173  178.340687  1.644223e+02   76.227559  336.476090
#> subject-40 -180.696256 -172.660366 -8.058063e+01 -214.047252   16.456600
#>                       0           1           0           1           0
#> subject.1    71.1243473  297.239544  101.596895   17.062604   23.331713
#> subject-2   398.8668974 -118.182870  227.427495  -34.274529   -9.869068
#> subject-3   100.5758804  116.077096   67.150605  300.722077  142.084691
#> subject-4   337.3116585  390.826880   49.675478   15.216944 -188.466246
#> subject-5    -5.9306036   68.544550   -7.505926  -12.479337  211.136846
#> subject-6  -121.7737303   51.048246  107.508322   -7.078662  -51.034166
#> subject-7   233.6705645  188.175530 -139.759316 -177.514657   70.551064
#> subject-8    85.2199499  137.222845  -85.845368  124.117129  111.794911
#> subject-9   -18.2448645  104.241762   61.958729  395.977917 -102.950809
#> subject-10   93.9268928  184.550478 -145.765183   54.514896 -198.869343
#> subject-11  280.7844893   39.986665   92.205949   55.568354   44.946131
#> subject-12  180.8909416    7.893388  115.328675 -113.277204  -90.599693
#> subject-13  -36.9379564 -278.091726 -161.038300   -8.569767  -26.332156
#> subject-14  190.8314987  126.581772   77.072320   75.106249   51.786441
#> subject-15   24.9885718 -238.053720  -63.290486  -17.171088  301.037903
#> subject-16   67.1641062  257.070640   94.607656  123.854835 -140.917572
#> subject-17  -20.3428419  -54.468871  102.070837  158.543764   89.179868
#> subject-18  -29.5258787  216.977599 -229.481299  -23.592832 -246.482778
#> subject-19  -87.4214387  -77.570935  -98.751187 -131.184568 -347.482394
#> subject-20 -136.1801606  116.425824   59.054565  131.077129  249.150118
#> subject-21   89.3853486  196.168142  -74.830151 -195.831595  -56.646001
#> subject-22 -145.9163166    8.211214  220.281661  -10.269464  -83.210293
#> subject-23   27.8979112  -49.673929   67.880000  -44.044382  202.920793
#> subject-24  -24.3145354   69.755309 -311.372116  113.305349  -13.794715
#> subject-25   -4.9377458  -57.603367 -152.595266  138.823114 -400.573969
#> subject-26 -317.5932324  467.596825  -28.510177  408.928356  403.586478
#> subject-27   42.3037340  116.575870  157.870824   16.925975  159.592312
#> subject-28   64.3596260   -1.442041  106.322467   58.687254   -9.023463
#> subject-29   -4.3739713  -76.552642 -145.541959 -179.290015   25.296081
#> subject-30   -0.3551699 -145.955355   22.803764   45.167071  -28.540650
#> subject-31  150.1324439   29.508117   57.291734 -221.672746 -219.496898
#> subject-32  268.8216444   94.162985   56.415957  135.907255   31.097415
#> subject-33   80.1601968   38.890795  115.433101  126.414894 -100.542957
#> subject-34   90.9625480  169.894228 -115.262011  -98.973550   69.699604
#> subject-35  -10.9337037 -165.069395  214.677211  300.304727  -81.130774
#> subject-36  187.7487666  125.427248   57.331634  162.335805  -91.192166
#> subject-37 -160.2894199  188.221931  -73.374705   63.353531   63.171844
#> subject-38 -306.5652968  118.493564 -124.161932  183.591480   -6.903360
#> subject-39   10.0198766  125.161460  208.295781   84.052283  258.951106
#> subject-40  319.7346226  121.587528   38.547776  159.795274  -54.721636
#>                      1           0           1            0           1
#> subject.1   255.680793  114.036323  154.848780   71.4124001  264.949284
#> subject-2   -55.415517  162.012191   15.823300  -55.5361615  134.053342
#> subject-3   123.250974   53.415004  201.208511   -6.5188374   56.754829
#> subject-4    55.304383  384.302055  218.092733   54.0336457    1.175887
#> subject-5   -62.326078   37.627308  108.672399   25.8057294  270.900299
#> subject-6   118.573230  147.985463  -29.090257   -4.3776363    5.502364
#> subject-7   100.045623  335.073537   13.867565   45.7032777   67.006929
#> subject-8    12.062585   16.412490  173.576415  104.4625029   96.400183
#> subject-9    27.810134  131.320996  257.860404   64.4001825   56.160545
#> subject-10 -247.514486 -473.714086  -34.438228  133.6136461  -58.893479
#> subject-11  147.993940   90.764453  153.524407  155.4629643  115.266015
#> subject-12    5.542078  199.314351  240.500643  206.3633200 -175.989005
#> subject-13 -261.678584 -117.676907  -10.602628  -32.6442155  -90.665697
#> subject-14   97.840795  299.926848   93.188077  108.6208963  183.667689
#> subject-15  266.057811  -35.287421  -86.739753  248.7977416 -169.603433
#> subject-16  -28.599652   34.510043   41.755483   77.2709858  197.635721
#> subject-17  123.045120   22.355382  101.333358  136.2557162   97.896731
#> subject-18 -217.646846  258.509176   39.411097  105.4094787  458.894603
#> subject-19 -128.878255  -54.115036 -212.225579  -34.4176730   42.484741
#> subject-20  116.454530  -28.787886   79.883960  113.1855206 -101.178349
#> subject-21  -27.104956  122.491248   -3.296950   -0.7876181   78.003152
#> subject-22  -35.473765   12.686197  227.727045  -18.2556665  479.859163
#> subject-23  -45.799841   93.394858   48.304808 -208.7196428  267.721284
#> subject-24  101.143367  153.012163   -8.435503   29.3973785  -82.216133
#> subject-25  -89.254081 -339.965874  -29.025630   68.5111727 -167.565769
#> subject-26  133.231726   31.103106  216.101376   44.5248267  457.136843
#> subject-27  -32.228711  -10.273250  -31.945555   63.3570186  -82.557933
#> subject-28  -33.021850 -100.565360  184.505813  -75.3166663  -19.822751
#> subject-29   76.526488   52.144567  -99.384506  -40.1615735    8.442003
#> subject-30   66.856589 -129.860219   89.496620  137.2259005   94.003955
#> subject-31   -7.296423  -95.167704  -99.269551  -62.6082930  -11.050520
#> subject-32   -1.058402 -113.409479  108.849483  132.8413802  193.013336
#> subject-33  137.928593   22.413140   91.559972 -125.8710327  191.359182
#> subject-34  -33.637566   97.702818 -104.082518  -38.7232909 -105.796184
#> subject-35  115.856104  -19.263342  166.571759  120.7786242  110.373429
#> subject-36  230.326406   -9.299752  -41.549877 -329.1048309  -65.786975
#> subject-37  297.703921   33.098965  -75.617972  -60.9942381  -51.032403
#> subject-38  247.466340 -248.064624   52.954448  192.1285066   45.024173
#> subject-39  311.735044   83.865723   73.487790  -51.2441306  289.246019
#> subject-40   68.142362  -17.624309  -65.556716  -84.4844955  -81.350040
#>                      0           1           0           1           0
#> subject.1     7.182083  221.019096   23.919693  135.448965   60.708149
#> subject-2  -205.451451  -13.840506   18.911183   21.457948 -150.028376
#> subject-3   -59.300446   99.882782    1.947481   25.477567 -107.040039
#> subject-4   -64.859712  -79.010344  -10.491176   11.295753  160.505858
#> subject-5    65.459504  148.574784   48.784233  175.447934  130.178581
#> subject-6   189.094160   10.844423 -153.452285  147.194265  -55.242408
#> subject-7   147.146827  116.626144   11.604423  101.128481  309.653109
#> subject-8     1.540064  -21.125847  189.700797  113.932025   86.172064
#> subject-9  -170.970750  -15.179896   65.323497   78.776533  -29.120855
#> subject-10 -123.189780  105.272105 -126.577471  -87.576731 -326.814697
#> subject-11  173.709937  188.304433  -21.883466  101.102173  141.002775
#> subject-12   62.524528  -61.991793  155.521951   34.503033  105.705931
#> subject-13 -115.470881  -33.533728  160.172100  -72.276115 -141.093836
#> subject-14  237.007661   70.194667  212.553839   40.716571   77.872992
#> subject-15 -352.698504   27.947063  -88.231456 -135.449333  110.014827
#> subject-16   99.411309   80.442597   23.425977  103.772310  252.144752
#> subject-17  168.435947   12.220505  104.210640  266.606499  105.487325
#> subject-18  -71.538399  246.016683  -23.317327  136.897373   22.436982
#> subject-19  107.708155   -4.099327 -227.924696 -211.277964 -176.833635
#> subject-20  -60.063874 -199.324179  277.198356   48.193693  -41.030547
#> subject-21  166.790585    9.251649  122.712643  276.819271   13.752301
#> subject-22 -122.953206  151.677518   35.445804  233.777982  123.229401
#> subject-23  105.053813  -24.543742   27.010126  156.388248   61.244708
#> subject-24  -23.808465    4.500397   63.913844 -151.079865   10.746645
#> subject-25  -31.691323   94.595302  -43.489196   12.517240  -93.676063
#> subject-26   93.330248  184.677248   83.834542  161.355462 -287.931617
#> subject-27   47.455014  133.150290  -57.317815   -6.085645  148.891731
#> subject-28  -80.361938  166.743883 -124.071131   87.787056 -144.012958
#> subject-29 -138.174865 -135.082044 -216.158833   43.740020   17.189789
#> subject-30   12.019868   35.715911   78.427392 -132.063114    8.401982
#> subject-31  -65.182907 -580.836326  125.888308  -27.811842    6.283315
#> subject-32   40.309240  189.264767   21.217022   -2.205163 -166.168444
#> subject-33   41.411538  -13.831702   56.812941  117.055512   52.891513
#> subject-34   63.450070  -13.191783 -127.989139  275.621194   19.118673
#> subject-35 -108.336987   47.132045   -5.794030  212.477955  -24.833328
#> subject-36  186.249882  204.219408  -46.151355  211.554137 -156.789578
#> subject-37    3.344067    4.085776  177.527077  112.437494  -22.776015
#> subject-38   96.823007  122.869471   48.749212  -28.676063    9.284460
#> subject-39  155.565977  149.914744  283.327127  109.736153  180.049322
#> subject-40  -42.077319   80.145923 -138.695467   -2.549672   57.923555
#>                      1           0             1           0           1
#> subject.1   182.548543   34.723949  253.04076248   95.364092  181.465226
#> subject-2   165.880454  271.003696  436.09952273 -383.212291  100.616276
#> subject-3    40.110601 -233.501144  -32.16525671   59.032446  -13.182013
#> subject-4    -4.594161  -10.949871    0.05480911   46.927812  -24.410595
#> subject-5    80.615494   79.492349   18.60954991  118.568570  166.754478
#> subject-6   -17.429325  152.263754   16.29813165   44.814915   11.265652
#> subject-7   -13.991996  -91.649875   27.77583957   21.946211 -131.835124
#> subject-8   246.011099   11.530418  100.62543510 -147.939041  155.133947
#> subject-9   -78.379290   96.352654   22.89027119  -17.638477  157.594942
#> subject-10    5.323513 -317.357224 -311.03212268 -343.090461  -32.201132
#> subject-11  -17.795803   58.196541   80.72057218   77.013296  129.492119
#> subject-12  109.822441   -5.243147  -37.51797473  117.557204  100.143048
#> subject-13 -140.540824  -36.528186 -180.64545731  128.108935 -128.926017
#> subject-14  337.637277  132.923337   63.61047875   61.752913   98.950641
#> subject-15   88.132474    1.278310  254.48039684    8.288192  242.833301
#> subject-16   44.093278  -48.945206  137.69764318  146.156717   75.952640
#> subject-17   -6.076608   69.059542   60.90774981   -6.530740  216.682736
#> subject-18   95.531371  -20.606779 -108.58469295  -33.423526 -357.792144
#> subject-19  -63.259551  -81.940196 -198.34615113   52.276991   39.247182
#> subject-20    7.022293  -33.186601 -202.97752390  209.026301  202.237208
#> subject-21   65.221723 -106.721682  298.45188448  208.980086 -156.015119
#> subject-22  420.275146 -113.159073 -218.01955336  276.768227  133.992227
#> subject-23  -28.949674   49.859492  135.00695618  208.546203   33.579233
#> subject-24  300.397778  160.622754   23.28186704  -58.768716   70.264260
#> subject-25 -194.037818   14.577663   19.71773736  -11.196020 -159.440148
#> subject-26  172.269476  123.351323  252.27845024  -93.368502   62.930563
#> subject-27   13.090499   54.425094  169.04133967  -76.534064   30.197594
#> subject-28   74.333677 -111.416167    5.99360924   71.195423   55.733918
#> subject-29   45.530332  -66.460499  -40.57925173  -46.391754   33.705297
#> subject-30   50.256914   34.321590    2.08525608  -33.462791   72.766605
#> subject-31  101.237604  -64.112082  -28.57745190  -13.978176  274.366622
#> subject-32  202.612726    8.410654   61.31031325   80.158132  137.050682
#> subject-33  231.455902  -30.939975   99.45656177  189.978224   49.500328
#> subject-34   74.588267  192.629103  -72.01468214  106.367832    2.080129
#> subject-35  -82.871097   14.834563  123.24914861  109.866427  232.008073
#> subject-36  105.022579  302.911424   41.85555118   23.286107  143.242448
#> subject-37  161.393893  -77.449533   54.42333639  144.011287  121.345563
#> subject-38   25.915026  -15.508490   65.38106738  -48.496395   64.818660
#> subject-39  307.984464   57.128629   95.39887494   21.935043  133.406015
#> subject-40  -27.715483  -22.031305  246.56760016 -184.195119  143.338124
#>                      0            1
#> subject.1   141.728555  186.1425878
#> subject-2  -361.219535  129.1059346
#> subject-3    -5.773695  120.0774317
#> subject-4   316.602318   84.2261759
#> subject-5    64.088534  215.6143072
#> subject-6   -67.330600  292.9965117
#> subject-7   -21.038504  -76.5082178
#> subject-8    56.955213   90.0072390
#> subject-9    56.592652   -8.1627807
#> subject-10 -250.339390 -250.3215953
#> subject-11   88.899457  147.8577005
#> subject-12  162.141215  173.0753070
#> subject-13   63.851688  -99.6657083
#> subject-14  283.447593   89.4201970
#> subject-15    7.060882 -235.1416498
#> subject-16  113.402019   60.6700067
#> subject-17  165.000187  -11.4226601
#> subject-18  177.083977  -99.6420217
#> subject-19  -89.892764  -54.0381418
#> subject-20   22.295042    5.4678894
#> subject-21  169.216940   44.2382358
#> subject-22  137.221196   69.9854573
#> subject-23 -302.349191   29.0968179
#> subject-24 -158.415591   24.4927059
#> subject-25  160.999509  258.9893101
#> subject-26  465.687105  109.4989886
#> subject-27  -80.279645  -40.4975842
#> subject-28  -10.846569    0.1676731
#> subject-29  -76.651946  -41.2137006
#> subject-30  -62.267726  -48.8149016
#> subject-31   72.122856   58.3874549
#> subject-32  -12.584028  140.3604358
#> subject-33   49.456368   89.8009103
#> subject-34 -125.272690  100.2115142
#> subject-35  177.661739    2.2396626
#> subject-36   25.114410  159.8183465
#> subject-37   66.510740   25.2135286
#> subject-38  -22.879764  -25.6387781
#> subject-39  198.366219  456.3441580
#> subject-40  -96.423729    3.1531975
#> 
#> $covmat
#>            0          1          0           1          0          1          0
#> 0 19974.4873 -7820.6389  1531.6840  1477.77568 -2056.8766  6416.5430 13374.3307
#> 1 -7820.6389 29562.0092  3459.6803  2633.13474 11895.6948 -1258.6195 -6680.5408
#> 0  1531.6840  3459.6803 15468.0032  6950.95696  5344.0835  1352.7847  1021.8724
#> 1  1477.7757  2633.1347  6950.9570 18713.10308  3610.7111  -135.7963  5589.3579
#> 0 -2056.8766 11895.6948  5344.0835  3610.71111 16098.5482 -2539.9611   713.2626
#> 1  6416.5430 -1258.6195  1352.7847  -135.79635 -2539.9611 21915.8389  3349.1521
#> 0 13374.3307 -6680.5408  1021.8724  5589.35794   713.2626  3349.1521 32114.4933
#> 1   937.0307  2298.7015  6704.0628  3846.07160   683.3731  5283.9841 -1001.0240
#> 0  4945.8518  -208.8261  5255.2770  6577.07382  2117.6499  4356.7554  6464.5528
#> 1  7453.1061  9554.2440  8857.5083  9165.73609  6300.2323  -542.5680  5670.4893
#> 0   595.1348  4449.6602  -230.0661 -1286.05858  2122.7792 -3410.6093 -5985.4849
#> 1  5293.3986  -765.7089  7744.9865  4876.97015  3035.8805  3064.2669  4841.4730
#> 0  4389.2083  5396.7956  1204.8177  2593.98358  3195.7935  5628.2807  -985.5564
#> 1  2323.3847 -1585.1992   706.0443  1740.90956 -1189.0190  -640.3830  5591.3275
#> 0  4557.0056  3408.8648  2289.4407  2420.97429   522.0161   926.1221  3967.8876
#> 1  7353.3319  2575.0774 -2236.7326  4174.37014  -850.4274  2526.0154  5703.5270
#> 0  6685.1279  -478.0947  4135.7024  1919.22914   958.6554  -575.5405  3472.3144
#> 1  3829.9231   595.1666  3136.6692   -51.00116   630.9229  5417.7933  4168.2056
#> 0 -1943.8804 -2610.0958 -2717.7872 -2840.96205   429.6665  2968.5560   975.3678
#> 1  5407.8837  4238.7827  8554.0795  4840.31767  2944.9503  3234.4343  7882.2136
#> 0  4344.9577  4974.6366  3273.2175   354.67364  1974.9121  4792.7340  5722.2980
#> 1  1438.5711  1721.5188  1610.6369 -1564.97179  5578.5140  -104.7663  5376.4884
#> 0  2246.2685  5865.9269  2380.4668   221.44005   149.8431  3337.9335  3702.0212
#> 1  5295.2378  3010.6538  6810.1885  5184.60175  4284.4498  6224.9830  6988.6125
#> 0  7568.0369  -557.6212  1388.2532 -1430.79501  2803.0655  3284.8248  3972.6728
#> 1  6463.2900  5210.5169  2609.1675  3264.25024  2321.8724  -323.0654  2828.3551
#> 0  4173.7636  -227.8905  2165.9255  6882.56382  -146.2694 -2962.2353  8678.3589
#> 1  7725.9827 -5685.9464  -795.6095  5189.06652 -4299.8777  -451.4388  7431.6948
#> 0  6447.9361  3303.5845  3209.9744 -2345.47326   594.2085  8697.0907  3947.2095
#> 1  4478.1279  4082.6571 -2142.4066  2666.83865  -448.1373  1330.4388   151.1570
#> 0  6473.3692   939.6401  3652.3355  2099.34224  2942.4533  6039.1210 11254.5809
#> 1 10071.4809  1511.8113  1118.1553  3700.50991  -373.8356  9078.1105 12945.4367
#>             1          0         1           0           1          0
#> 0   937.03068  4945.8518  7453.106   595.13476  5293.39862  4389.2083
#> 1  2298.70154  -208.8261  9554.244  4449.66015  -765.70890  5396.7956
#> 0  6704.06283  5255.2770  8857.508  -230.06613  7744.98648  1204.8177
#> 1  3846.07160  6577.0738  9165.736 -1286.05858  4876.97015  2593.9836
#> 0   683.37306  2117.6499  6300.232  2122.77916  3035.88045  3195.7935
#> 1  5283.98413  4356.7554  -542.568 -3410.60934  3064.26694  5628.2807
#> 0 -1001.02403  6464.5528  5670.489 -5985.48489  4841.47298  -985.5564
#> 1 14795.53756  4195.0477  6844.734  1325.34129  3768.21986  3486.0865
#> 0  4195.04773 13954.7810  1808.304 -1434.02790  1862.40398  5376.3847
#> 1  6844.73395  1808.3042 27747.178  1294.58072 11421.29705  3762.9804
#> 0  1325.34129 -1434.0279  1294.581 24710.23392    15.56059  5005.7898
#> 1  3768.21986  1862.4040 11421.297    15.56059 23649.56126 -1379.6152
#> 0  3486.08651  5376.3847  3762.980  5005.78978 -1379.61519 16784.7552
#> 1 -2435.15403   930.9427  3582.867 -6036.85723  4589.70401  3569.5757
#> 0  -824.78103  4615.4436  7361.471 -5151.31240  2679.96128  3835.4627
#> 1    51.46366  2012.8140 10603.811 -2639.03518  3621.27383  4753.0133
#> 0  2731.61418  2246.2669  6389.526  7991.07631  7797.13478  3231.3207
#> 1  1715.30283  4298.1665  4214.305   347.81733  4417.04059  6562.0765
#> 0 -2310.52052 -2393.4951 -2168.888 -2028.05784  -238.45130 -1736.7600
#> 1  6647.98671  5657.9231  6256.118 -5142.11790  8102.55980  5637.1679
#> 0  2490.42753 -1014.9846  6501.088  -374.74199  6917.58105  1469.3092
#> 1    78.73081  -138.4489  5413.332 -1304.70034  5279.20143   421.6984
#> 0  2500.59861  3505.4948  3389.132 -2295.19444  2328.88371  2726.7824
#> 1  6485.90671  7230.3557  8071.425 -1060.23577  5682.61162  5044.2821
#> 0  2163.05484  -707.2377  3188.822  4378.11296  1635.29168  3183.2035
#> 1  5339.93276  2660.6699 12651.451   178.84558  5164.82297  3248.4566
#> 0 -1322.82641  1720.7304  7253.349  1963.63436  -526.79770  2700.9344
#> 1  -964.39176  4352.0040  1894.203  6265.63397  2387.37763  5924.6969
#> 0  3751.14094  4362.8000  2793.933 -5124.48150  -635.59021  3172.2039
#> 1   124.28103   609.9066  5688.959   425.56383 -3253.89172  8584.6509
#> 0  1796.81027  4134.4516  9092.097 -5121.27409  9834.89222  2478.5257
#> 1  3707.67583  4981.2787  8093.620  2187.64879  4531.23659  7283.2146
#>            1          0           1           0           1          0
#> 0  2323.3847  4557.0056  7353.33185  6685.12794  3829.92313 -1943.8804
#> 1 -1585.1992  3408.8648  2575.07738  -478.09468   595.16660 -2610.0958
#> 0   706.0443  2289.4407 -2236.73263  4135.70243  3136.66923 -2717.7872
#> 1  1740.9096  2420.9743  4174.37014  1919.22914   -51.00116 -2840.9621
#> 0 -1189.0190   522.0161  -850.42740   958.65540   630.92285   429.6665
#> 1  -640.3830   926.1221  2526.01536  -575.54047  5417.79330  2968.5560
#> 0  5591.3275  3967.8876  5703.52698  3472.31444  4168.20555   975.3678
#> 1 -2435.1540  -824.7810    51.46366  2731.61418  1715.30283 -2310.5205
#> 0   930.9427  4615.4436  2012.81395  2246.26690  4298.16648 -2393.4951
#> 1  3582.8671  7361.4707 10603.81080  6389.52615  4214.30508 -2168.8876
#> 0 -6036.8572 -5151.3124 -2639.03518  7991.07631   347.81733 -2028.0578
#> 1  4589.7040  2679.9613  3621.27383  7797.13478  4417.04059  -238.4513
#> 0  3569.5757  3835.4627  4753.01326  3231.32066  6562.07647 -1736.7600
#> 1 21438.1855  5121.7327  5903.42704 -4671.26380  8361.71347  1999.3300
#> 0  5121.7327 28248.0901 10750.23759  5018.47659  4206.92405  1389.5002
#> 1  5903.4270 10750.2376 18945.46888  4745.27551  2018.14661   153.8343
#> 0 -4671.2638  5018.4766  4745.27551 26996.23213  4823.00912 -1470.4146
#> 1  8361.7135  4206.9241  2018.14661  4823.00912 12964.45930  3695.0653
#> 0  1999.3300  1389.5002   153.83429 -1470.41465  3695.06533 12696.8600
#> 1  3808.0215  5274.7985  -338.55322  7337.29173  8113.53595 -1066.6805
#> 0  -879.6985  1738.1298  3654.91181  4835.98590  -111.22064 -2326.1184
#> 1  7487.3838  3596.6703  1705.89210   -60.03277  4336.80201  1487.1944
#> 0  3109.6421  6987.4514  4037.66655  3502.90957  5507.80174  2075.6743
#> 1  1447.0301  4256.5202  2530.91052  6224.15653  4339.62180 -3526.0746
#> 0 -5725.5822  3597.6444  5207.92459 11773.07861  2047.22469  2798.4749
#> 1  -447.2724  6364.1110  5459.98220  5874.42959  3458.29209 -1043.8841
#> 0  2505.0877  4147.0095  4288.20238  6858.74556  -245.34489 -2809.0858
#> 1  3129.7406  8691.5909  8601.34889  6665.36418  1193.25836  -723.3936
#> 0 -2511.1049  1809.9617  2775.05366  5022.87584  3058.75106 -1944.4172
#> 1  4511.5524  7449.1541  8973.52036 -2208.86264  2373.66033   865.3667
#> 0  6559.9938  1403.9510  5278.07474  6757.10515  8665.23612  4754.4596
#> 1  2163.2311  2074.7074  5785.35663  5159.34960  4032.31471 -3736.5245
#>            1          0           1          0           1          0
#> 0  5407.8837  4344.9577  1438.57114  2246.2685  5295.23783  7568.0369
#> 1  4238.7827  4974.6366  1721.51879  5865.9269  3010.65383  -557.6212
#> 0  8554.0795  3273.2175  1610.63690  2380.4668  6810.18846  1388.2532
#> 1  4840.3177   354.6736 -1564.97179   221.4400  5184.60175 -1430.7950
#> 0  2944.9503  1974.9121  5578.51405   149.8431  4284.44976  2803.0655
#> 1  3234.4343  4792.7340  -104.76635  3337.9335  6224.98302  3284.8248
#> 0  7882.2136  5722.2980  5376.48837  3702.0212  6988.61250  3972.6728
#> 1  6647.9867  2490.4275    78.73081  2500.5986  6485.90671  2163.0548
#> 0  5657.9231 -1014.9846  -138.44892  3505.4948  7230.35570  -707.2377
#> 1  6256.1182  6501.0876  5413.33201  3389.1318  8071.42513  3188.8222
#> 0 -5142.1179  -374.7420 -1304.70034 -2295.1944 -1060.23577  4378.1130
#> 1  8102.5598  6917.5811  5279.20143  2328.8837  5682.61162  1635.2917
#> 0  5637.1679  1469.3092   421.69844  2726.7824  5044.28213  3183.2035
#> 1  3808.0215  -879.6985  7487.38375  3109.6421  1447.03006 -5725.5822
#> 0  5274.7985  1738.1298  3596.67030  6987.4514  4256.52017  3597.6444
#> 1  -338.5532  3654.9118  1705.89210  4037.6665  2530.91052  5207.9246
#> 0  7337.2917  4835.9859   -60.03277  3502.9096  6224.15653 11773.0786
#> 1  8113.5359  -111.2206  4336.80201  5507.8017  4339.62180  2047.2247
#> 0 -1066.6805 -2326.1184  1487.19439  2075.6743 -3526.07459  2798.4749
#> 1 28490.0164  4342.5576  9374.08513  3970.0207  8356.72195  2630.8845
#> 0  4342.5576 15987.8276  4163.51260  2938.7123  5495.62878  3745.8118
#> 1  9374.0851  4163.5126 20155.93175 -3348.1828  3646.05601  -560.6701
#> 0  3970.0207  2938.7123 -3348.18282 14747.5163  2249.40230  3700.4896
#> 1  8356.7220  5495.6288  3646.05601  2249.4023 14084.28752  3614.6123
#> 0  2630.8845  3745.8118  -560.67006  3700.4896  3614.61227 18339.4614
#> 1 10244.7470  1530.9274  2421.54384  6066.2986  2328.68031  2654.5519
#> 0   900.0852  3899.2670  1526.84803  1215.1125  3218.15440   738.9111
#> 1  3438.9816  1389.9435  3421.50037  1365.3016  3585.76053  3991.3800
#> 0  3726.3436  4672.3942  -599.73482  4153.3910  6212.78469  6128.8150
#> 1 -1311.6086 -1428.5843 -4756.89138  3260.1648    88.45584  1671.6502
#> 0  9025.5924  4679.8555  2962.01010  8030.5822  6313.64685  4121.6829
#> 1  5190.0387  7763.5784  2344.43293  4387.3436  6010.37898  2451.9255
#>            1          0          1          0            1            0
#> 0  6463.2900  4173.7636  7725.9827  6447.9361  4478.127935  6473.369214
#> 1  5210.5169  -227.8905 -5685.9464  3303.5845  4082.657108   939.640109
#> 0  2609.1675  2165.9255  -795.6095  3209.9744 -2142.406612  3652.335478
#> 1  3264.2502  6882.5638  5189.0665 -2345.4733  2666.838648  2099.342240
#> 0  2321.8724  -146.2694 -4299.8777   594.2085  -448.137347  2942.453274
#> 1  -323.0654 -2962.2353  -451.4388  8697.0907  1330.438806  6039.121033
#> 0  2828.3551  8678.3589  7431.6948  3947.2095   151.156950 11254.580901
#> 1  5339.9328 -1322.8264  -964.3918  3751.1409   124.281026  1796.810266
#> 0  2660.6699  1720.7304  4352.0040  4362.8000   609.906642  4134.451592
#> 1 12651.4506  7253.3490  1894.2028  2793.9335  5688.958747  9092.096759
#> 0   178.8456  1963.6344  6265.6340 -5124.4815   425.563833 -5121.274086
#> 1  5164.8230  -526.7977  2387.3776  -635.5902 -3253.891724  9834.892223
#> 0  3248.4566  2700.9344  5924.6969  3172.2039  8584.650875  2478.525671
#> 1  -447.2724  2505.0877  3129.7406 -2511.1049  4511.552409  6559.993784
#> 0  6364.1110  4147.0095  8691.5909  1809.9617  7449.154141  1403.950950
#> 1  5459.9822  4288.2024  8601.3489  2775.0537  8973.520355  5278.074742
#> 0  5874.4296  6858.7456  6665.3642  5022.8758 -2208.862637  6757.105153
#> 1  3458.2921  -245.3449  1193.2584  3058.7511  2373.660325  8665.236122
#> 0 -1043.8841 -2809.0858  -723.3936 -1944.4172   865.366750  4754.459593
#> 1 10244.7470   900.0852  3438.9816  3726.3436 -1311.608646  9025.592428
#> 0  1530.9274  3899.2670  1389.9435  4672.3942 -1428.584255  4679.855469
#> 1  2421.5438  1526.8480  3421.5004  -599.7348 -4756.891385  2962.010105
#> 0  6066.2986  1215.1125  1365.3016  4153.3910  3260.164767  8030.582230
#> 1  2328.6803  3218.1544  3585.7605  6212.7847    88.455837  6313.646847
#> 0  2654.5519   738.9111  3991.3800  6128.8150  1671.650172  4121.682890
#> 1 17141.5633  2472.0536  2610.6603  1128.3048  4416.993164  3416.492279
#> 0  2472.0536 13937.6307  7765.9865 -1866.4849  3464.099709  -165.559604
#> 1  2610.6603  7765.9865 22473.2082 -4023.8176  4133.122347  1140.261175
#> 0  1128.3048 -1866.4849 -4023.8176 18194.2220   534.585031  7646.342778
#> 1  4416.9932  3464.0997  4133.1223   534.5850 15709.630654     1.344517
#> 0  3416.4923  -165.5596  1140.2612  7646.3428     1.344517 25743.271708
#> 1  4855.2612  6631.4351  4754.6622  3690.5613  2475.919576  6049.009312
#>            1
#> 0 10071.4809
#> 1  1511.8113
#> 0  1118.1553
#> 1  3700.5099
#> 0  -373.8356
#> 1  9078.1105
#> 0 12945.4367
#> 1  3707.6758
#> 0  4981.2787
#> 1  8093.6199
#> 0  2187.6488
#> 1  4531.2366
#> 0  7283.2146
#> 1  2163.2311
#> 0  2074.7074
#> 1  5785.3566
#> 0  5159.3496
#> 1  4032.3147
#> 0 -3736.5245
#> 1  5190.0387
#> 0  7763.5784
#> 1  2344.4329
#> 0  4387.3436
#> 1  6010.3790
#> 0  2451.9255
#> 1  4855.2612
#> 0  6631.4351
#> 1  4754.6622
#> 0  3690.5613
#> 1  2475.9196
#> 0  6049.0093
#> 1 17298.7611
#> 
#> attr(,"class")
#> [1] "aat_intercategorical_alpha"

aat_splithalf(ds=ds,subjvar="subj",pullvar="is_pull",targetvar="is_target",rtvar="rt",
              algorithm="aat_doublemeandiff",iters=100,plot=FALSE)

#> 
#> Full-length reliability (Raju's beta):
#> beta (40) = .21, 95%CI [-0.17, .53], p = .192
#> 
#> Uncorrected, average split-half correlation:
#> r (40) = .13, 95%CI [-0.1, .37], p = .424