Simulate AAT datasets and predict parameters aat_simulate() generates approach-avoidance task datasets.

Simulate AAT datasets and predict parameters aat_simulate() generates approach-avoidance task datasets.

aat_simulate(
  npps = 40,
  nstims = 32,
  stimreps = 2,
  meanrt = 743,
  meanrt_jitter = 66,
  sdrt = 133,
  sdrt_jitter = 38,
  pullfx = 25,
  pullfx_jitter = 40,
  stimfx = 10,
  stimfx_jitter = 35,
  biasfx = 35,
  biasfx_jitter = 75
)

Arguments

npps
Number of participants

nstims
Number of stimuli

stimreps
Number of repetitions of each stimulus within each group (i.e. within approach target, avoid target, approach control, avoid control)

meanrt
Mean sample reaction time

meanrt_jitter
Extent by which participants' mean RTs deviate from mean sample RT.

sdrt
Standard deviation of samplewide RTs, ignoring effects of movement, stimulus, and approach bias. In essence, this represents the amount of pure noise present in the data.

sdrt_jitter
Extent by which standard deviations of individual participants' RTs are larger or smaller than the samplewide SD.

pullfx
size of the effect of approach-versus-avoidance, in milliseconds

pullfx_jitter
Individual variation in the effect of approach-versus-avoidance

stimfx
size of the effect of stimulus category, in milliseconds

stimfx_jitter
Individual variation in the effect of stimulus category

biasfx
Size of the approach bias effect, in milliseconds

biasfx_jitter
Individual variation in the approach bias effect

Value

aat_simulate() returns a data.frame with the following columns: subj (participant ID), stim (stimulus number), rep (stimulus repetition number), is_pull (0 = avoid, 1 = approach), is_target (0 = control stimulus, 1 = target stimulus), meanrt (participant's mean RT), sdrt (participant's residual standard deviation), pullfx (participant approach-avoidance effect size in ms), stimfx (participant stimulus category effect size in ms), biasfx (participant approach bias effect size in ms), and rt (trial reaction time). Additionally, the data.frame has the attribute population_reliability which represents the expected reliability of the data given the provided parameters.

Details

Defaults of aat_simulate() are based on Kahveci, Van Alebeek, Berking, & Blechert (2021).

Examples

ts<- aat_simulate(pullfx = 50, stimfx = 10, biasfx = 100)
mod<-lm(rt~is_pull*is_target,data=ts)
coef(mod) #these should be somewhat close to the provided coefficients

#>       (Intercept)           is_pull         is_target is_pull:is_target 
#>        693.909920         54.232076          8.621564         80.587690 

print(attr(ts,"population_reliabilty"))

#> NULL

print(q_reliability(ts,"subj",rt~is_pull*is_target,"is_pull:is_target"))

#> q = 0.6739759

#these two should be not too far apart,
# and should converge when the process is repeated a bunch

# Here's how to derive the parameters used in this function from a real dataset
if (FALSE) {
mod<-lmer(decisiontime ~ is_pull * is_food + (is_pull * is_food | subjectid),data=dsa)
fixef(mod) # from here, all the fx and mean RTs are derived
ranef(mod)$subjectid %>% apply(2,sd) #from here, all the fx jitters are derived
dsa %>% group_by(subjectid) %>% summarise(sd=sd(resid)) %>%
summarise(m=mean(sd),s=sd(sd)) # from here, sdrt_jitter is derived
}