Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

aat_stimulusscores(
  ds,
  subjvar,
  stimvar,
  pullvar,
  targetvar = NULL,
  rtvar,
  aggfunc = c("mean", "median"),
  iters = 5
)

Arguments

ds

the data.frame to use

subjvar

Name of the subject-identifying variable

stimvar

Name of the stimulus-identifying variable

pullvar

Name of the movement-direction identifying variable

targetvar

Optional. Name of the stimulus-category identifying variable

rtvar

Name of the reaction-time identifying variable

aggfunc

The function with which to aggregate the RTs before computing difference scores. Defaults to mean but can be changed to median.

iters

If there are missing values (which is almost inevitable) then multiple imputation will be used to complete the covariance matrix - this argument sets the number of multiple imputations to be used.

Value

Exports a list containing a data.frame with stimulus-specific bias scores, indicated in the column names, a covariance matrix of that same data, and a data.frame indicating to which stimulus category each stimulus belongs.

Examples

ds<-aat_simulate(biasfx_jitter=40,nstims=16)
ds$stim<-paste0(ds$stim,"-",ds$is_target)
aat_stimulusscores(ds,"subj","stim","is_pull","is_target","rt")
#> $data
#>          0-1-0      0-10-0      0-11-0      0-12-0       0-13-0      0-14-0
#> X1   98.864922  164.087481  104.372630  137.066771   97.4888957  157.851864
#> 2    11.952243  -95.191233  -88.265923  -84.863641  -58.9885785   30.975705
#> 3   111.297676  171.750660  141.956369  122.017940  117.6301200  190.777288
#> 4   -40.797938    4.097875   31.518534   74.479393  -58.0775155 -162.528796
#> 5   -72.201145 -128.734027  -14.896493   20.302889  -20.5576916   23.682171
#> 6   238.965220   29.347105   58.916132  235.403735  180.1681907  -45.843423
#> 7    69.395550  -42.018742   72.890180   76.912383 -149.2615752   41.549355
#> 8    21.406017   98.437782   54.405461 -192.574959   58.8015574 -116.289323
#> 9   201.200973 -232.996461  154.615030  118.396994   59.8295913 -234.730767
#> 10  134.270350  324.301568  -37.443348  142.996947  197.9566210  181.524152
#> 11  220.643145  -87.592468  117.464177  153.200853   78.8392998   58.304426
#> 12  -21.910647    6.267868   47.442660  -24.870298  -27.6619533   -4.003584
#> 13 -127.725304 -381.337953   51.786057  236.941951   83.1832205 -142.149450
#> 14   24.162967   -1.313555 -211.516594  -51.642569 -159.1120526 -101.021699
#> 15  -71.286276   13.774646   17.858585    4.879889   13.3015975   37.359104
#> 16  -61.419179  184.703811    4.867534 -105.032079  134.2202215   58.788356
#> 17 -121.021504  -95.093636  -19.842118  -70.911983   57.9390551   32.074713
#> 18   83.540819  169.379673   30.949658   74.636372   30.8190026 -189.549188
#> 19   60.674379   -1.841205  -69.470014   24.381232  -11.8477966  105.473910
#> 20   66.409970  118.513298  203.381383    7.762124   44.7359170  101.008143
#> 21  -10.040508  310.178788  196.390131  108.038382   56.9330264   -2.933673
#> 22    3.983384   31.867910  -21.898172   47.478103    0.6447895  132.553659
#> 23 -113.895447 -159.212045  -96.845611  -43.368074 -136.1989572 -265.536785
#> 24 -153.590330 -141.163896 -159.169638  -33.411877   -5.3662318 -360.348890
#> 25  211.977096 -120.187894 -226.196613   18.407782  -82.3689766 -143.796177
#> 26  149.030087 -136.535028 -129.240432  -94.580588  105.9672890 -135.050702
#> 27  239.340894   58.289167 -157.896081 -239.853008   89.8836355  -94.372980
#> 28  -99.896050 -445.866825  238.493340 -159.858094   66.2922113  102.275958
#> 29 -152.227790  -24.270343  -21.215103   32.647433 -118.9764601  -26.886269
#> 30  149.479869  -79.228693 -116.828762  243.188272  -23.3376173  145.606033
#> 31  176.921585  250.023227 -143.127939  342.580823   40.2074178  291.906266
#> 32  -16.910019   42.085793 -112.657207  -82.901985  -29.2844358   41.101194
#> 33   24.440951  -52.120519  212.519780   77.398845  136.9569837  158.752847
#> 34 -282.404397  162.025861   32.387488  -41.433952 -112.6891268 -227.695174
#> 35  -50.995663  -64.201869  -14.139607   -5.821851 -183.1030174   39.964451
#> 36 -139.021329  -28.166866  168.496831 -196.297434  132.0757593 -266.894976
#>          0-15-0       0-16-0       0-2-0       0-3-0       0-4-0       0-5-0
#> X1    0.1796050  139.4501188  161.045835  141.552845  207.870183  132.670107
#> 2   -57.9355091   56.8754188   81.010744   -2.449653   -1.428043    6.504599
#> 3   110.7985910  128.0594509  247.592729    2.115978  181.579492   62.892093
#> 4  -184.2663591 -159.4836655  210.109700 -121.367321   82.337659   61.546579
#> 5    -7.4394754   53.8865727   60.725039   -9.198100 -120.308006  -30.649301
#> 6    36.8411884   62.7279928  172.111068  306.345510  102.904674  -72.860234
#> 7   -67.7624512  124.7386326  -18.666016 -221.632709   25.210757   10.045300
#> 8    68.9777091  -22.1651590   87.702078   49.998050  -26.521211   91.180075
#> 9    -1.8652400  131.7410449  100.149278 -136.555702   50.535525  136.299861
#> 10   56.2762210   75.0578388  124.428864  171.694755   29.582332  122.779915
#> 11 -343.9917367  -13.5064651 -126.493542  200.823871 -116.646395  -75.619455
#> 12   75.8212445  -29.3411558   26.921112   38.396945   36.137840 -130.213750
#> 13   99.9749901   49.2715189  -56.755715  158.495010   86.255403  -27.296803
#> 14   31.1882784 -202.5229594  -19.814212  -68.977253   25.106996  -98.901852
#> 15  106.8727637   35.7819434  -58.211321   83.641674  -82.075298    2.591176
#> 16 -178.3331179    0.7062584  -35.885114  -69.296920   54.968758  -37.521254
#> 17   33.3284224   61.0564218   29.585824  -44.199048  -46.001167   81.400181
#> 18  -19.6812545  -23.1401042 -105.381273  -27.659696   44.602777   33.299124
#> 19   72.6569502   99.2579911  -54.320162    6.849794  -13.018137   -8.584350
#> 20  167.7000728   81.1028846 -115.508116  456.289738   50.554667  -63.141618
#> 21  202.4224260 -240.5306731   59.633037  -80.420330 -172.427337  -82.016758
#> 22   17.5106313  175.4390575   16.567376  -58.956358   23.648214  109.198514
#> 23 -249.8396811  -67.2267997 -193.575065  -60.541010 -119.955881  -88.360761
#> 24 -149.3942452 -140.5757925  -92.062537 -215.156941  -93.616266   13.110475
#> 25   59.7865467  350.1954224  251.977199  294.313580 -106.497072  325.337105
#> 26 -156.8459274    2.1887553    2.063387   -2.647429  -94.651367  -30.784600
#> 27   25.5902339  -81.5514750  212.579170 -163.591190 -324.185960   47.688073
#> 28  217.5051179 -110.8332804  110.818138 -304.124706  -88.420675  -81.504984
#> 29 -152.4991397  -40.5111810  -62.996945  115.221556  -42.706776   19.349637
#> 30  115.5136150  -43.8889614 -191.498126 -164.622012   15.963580  250.121793
#> 31   71.0069834 -234.6637496  101.157073  194.116683  -67.407852   31.203341
#> 32  -28.1918488   36.6752767  119.847564   87.012459 -109.519697  -10.931369
#> 33  131.1888490   53.5288709   -2.331889  156.262643  162.259592   61.275311
#> 34  -83.4417563  -46.1789187 -194.950192  -52.480897   12.147413 -173.504049
#> 35   80.2185182  291.0847605 -118.222021   37.168988   95.429261  139.810649
#> 36   -0.3663344  -75.0358990   -1.940808  -51.336821   47.311191  142.174756
#>          0-6-0       0-7-0         0-8-0        0-9-0      1-1-1     1-10-1
#> X1   20.426698   97.203346  215.15912175  109.4157273  142.37597   11.72161
#> 2  -140.087900    2.569076  -88.00621962   12.1703069   21.47559   91.34993
#> 3   218.328927  220.349347   62.50109438  119.9989193  249.51594  157.12130
#> 4     6.283783 -125.149175  101.31448397  -71.9888265  144.06590  -28.15689
#> 5   -20.379356  106.423086  -40.56899670   89.1527502   18.84263   65.71652
#> 6   137.945972  106.397119 -126.19676133  147.5007998  388.76234  -31.99487
#> 7    47.129457   26.913058  107.14399082   -8.5220743  132.84590  316.10680
#> 8   -84.723168   10.749607    2.94543342   48.2502726  192.17323  127.64045
#> 9   291.002056  -54.870363  119.24624877  -83.5333077  -93.40211   17.99303
#> 10  113.281670  181.702253  184.45643259  203.6852219   63.94563  111.16703
#> 11   -8.655203  179.262066   30.55359708  178.8219259  237.60347   94.97320
#> 12  -35.399180   55.183960 -117.09159435   31.4388402  -34.26533   25.93926
#> 13 -166.631857 -152.788513   45.00872249  250.8198125  186.65209   79.00329
#> 14  113.724587   36.958618  -14.78368613 -150.5389772   82.75880  124.70176
#> 15  137.345022  -56.737539   -9.98265175    4.5032105   21.46055  162.24334
#> 16   54.089464  -77.448890   -0.22802296    4.2348548  149.54060   66.89335
#> 17   14.747562    2.636603  -50.62036479  -23.9713559   86.65126  178.24370
#> 18  144.347907  -29.535919   20.78577187  120.2899300   45.16826  151.47298
#> 19   24.204974 -124.288908    4.28032871    0.9600934   38.99915   54.52061
#> 20  287.402386  109.984206   66.31380884  187.4744396 -224.61494  117.75825
#> 21   55.835029  -48.238246   37.52836153  -42.1703745   58.61202   89.70426
#> 22   70.228719   87.161693   80.55548253   31.4742276  164.21901   28.21324
#> 23  -57.465553 -213.011936  -88.30330460  294.9397319   93.88594 -147.48903
#> 24 -193.220073   69.931427  -23.11292227 -106.5669197   92.84766  256.27061
#> 25   35.543183  435.092986  460.64052768  215.9352320  344.42269  256.29250
#> 26   93.086217  -70.508485  -54.37662055  -47.6087904  -19.02200  -96.22008
#> 27  -17.831678  -92.262896 -249.74557416  320.0924776  134.47104  125.17378
#> 28  343.068858 -237.787236 -211.08870384 -171.3875160  -68.87184  218.00401
#> 29  -69.497561  -82.481691    0.02594058  -16.1987589  166.47636  170.45638
#> 30  -48.612728  -72.835037 -101.79304155 -113.1749002   54.73948  102.89123
#> 31 -153.932063   41.391422  225.22589576   54.6741611   45.45018 -162.74603
#> 32   -4.955255   76.121575   38.76968558  -61.0041828  177.64733   11.21803
#> 33   60.283733  124.677475   91.09461964 -105.9357199  229.99329 -129.43539
#> 34  -21.248896 -139.245171   -7.92278380  104.3888064  108.97277  130.49176
#> 35  -73.663014  168.499924  146.14300591  187.5972323  304.38695  204.60253
#> 36   13.340442  111.401496   51.09828511 -219.9543870   27.52640  108.87764
#>         1-11-1        1-12-1      1-13-1      1-14-1      1-15-1       1-16-1
#> X1  160.744481  9.526193e+01   40.698008  -16.885257  100.400857  152.9823319
#> 2   134.104490  1.074480e+01   68.859469 -126.581812   58.774790   80.5621325
#> 3   182.338952  1.254929e+02  187.157548  222.458582  303.300358  170.3229279
#> 4   -83.046402  1.799887e+01  -62.865905  -18.480770   57.329150  -90.4586558
#> 5   -43.716452  4.487105e+01  109.297878   68.530807  -47.037801    9.7545031
#> 6    72.274281  1.021981e+02  132.479503  249.659468  293.118854  233.0587241
#> 7   122.089218  7.075492e+01  150.456214  206.325418  172.813977   52.5016713
#> 8   137.621575 -9.303398e-03  112.421235   24.273573   26.497731   -8.1382891
#> 9   102.804075  9.938897e+01 -197.945010   92.099061   68.158335  -91.7908438
#> 10  145.223539  1.559726e+02   73.143430  145.602999  146.731535  126.0802756
#> 11  -73.342510 -5.558271e+01   34.430472   26.019667  163.818014  515.5624663
#> 12  155.434660 -1.010219e+02  -63.910290 -118.944452  -72.106743  -36.4413266
#> 13 -112.897771  1.449961e+02   78.514147 -121.286239  -17.492404 -271.9032786
#> 14   21.313700 -8.201545e+01   84.997423    3.649304  -55.232723   11.2370364
#> 15  159.894023  1.068005e+02   -9.848103   72.287369   52.288686   90.1351288
#> 16   75.392779  3.115211e+01  112.723828  118.711719 -159.979859    0.9836114
#> 17    1.262357  1.551318e+02   12.991026  -71.229100  100.778229  155.3854792
#> 18   92.782058  2.080188e+02   -5.504101  140.712692   97.055206   19.9732561
#> 19   91.682177  2.050125e+02  -45.187378   20.966846   99.738743  229.7823472
#> 20  209.675884 -9.761831e+01  150.742935  225.634016  194.684750  185.1886688
#> 21  209.104295  4.372346e+01  116.600665  203.695359   90.307679  -86.0433727
#> 22  -86.832717  3.891717e+01 -181.396335  333.628738  -46.242225  -12.7966437
#> 23   55.976289  5.673190e+00   98.345822 -172.074855   39.009837   98.8972679
#> 24  143.635968  2.204237e+02  -51.950981    6.103003   45.458475  -21.6786171
#> 25   63.525684  2.629459e+01  196.599577   -5.951970  304.141493 -132.7433417
#> 26  134.286130 -1.818356e+01 -103.514764  -46.815103   -2.233538  104.3492614
#> 27  -79.360925 -2.047254e+02  164.697831   90.846277   64.627465  135.9011861
#> 28 -181.897630  2.395859e+02  226.814883 -227.765461  448.784960 -185.9453371
#> 29   64.017456  1.614557e+02  382.865013  -98.720238   -8.052679   85.2565717
#> 30  183.754410  1.350502e+02  -58.582471  -46.798700  121.359262  315.6695511
#> 31  192.364095 -7.084914e+01  -69.789422  178.117093   46.482585  -45.6601992
#> 32  262.912925  8.006430e+01 -136.746318  163.885245  -20.284111   22.7153867
#> 33  145.156825  5.348012e+01  173.938896  195.426771   37.226437  -41.3526555
#> 34   63.619239 -7.076480e+01   45.787007 -108.982836  -19.875052 -224.1736775
#> 35  228.056116  2.601646e+02  358.848001  -61.113013  229.029339   15.2528662
#> 36   40.783246  6.967860e+01  207.831821  101.942307  169.942873 -244.1769650
#>          1-2-1       1-3-1       1-4-1      1-5-1       1-6-1      1-7-1
#> X1  102.867338   76.653307  167.443634  175.52013  240.594818   52.30050
#> 2    15.972536   53.607494   54.319512  108.91319  -96.434390   22.15267
#> 3   138.187665   69.206669  100.496329  112.68296  245.075290  171.75506
#> 4    26.935296  205.454147  103.907379   11.63590  -16.870643  -61.88472
#> 5    54.680302   62.506079   88.302003   64.49152   34.641960  184.46978
#> 6    34.037820  144.271270  280.467390   34.97596  -73.258091  151.57504
#> 7   -39.366298  296.342904  153.679441  112.04199  220.998766   54.65209
#> 8   192.146733   97.349716  152.880672  113.91165  -12.272952  186.09960
#> 9    -9.949093 -113.469886  -42.678380  -19.25209 -142.905719   85.92698
#> 10   18.609722  379.824786  282.478756  209.06601    1.119966   53.76525
#> 11  198.188858  347.646499  195.821208  260.91532  370.083588  -21.93067
#> 12   19.971931  -67.912045  115.218483  -13.16003    5.835859  -79.36314
#> 13  162.511188  -30.631127 -261.105383  115.63103   86.349814  129.93541
#> 14  -14.010016   97.788274  -47.559532   86.11556  137.458408  -47.26092
#> 15  138.860163  -16.749597  188.936575  -49.43625   22.692595  112.00379
#> 16 -147.565913   42.755668  -91.286028  271.33192  -56.699059   -8.55956
#> 17   35.498391   52.273331  206.981477  -15.63053   -7.537649   74.25341
#> 18 -332.781805  141.923072   91.216399 -192.72183 -220.953689  286.37064
#> 19  116.453097   62.134374  -41.051103   36.58313  108.228290   47.70572
#> 20  -20.184308   29.542779  107.566342   77.95290  -63.361812 -322.85202
#> 21  292.213581  -32.761138  209.437517   80.47988  -56.375111  197.68807
#> 22  141.702585   81.127900   83.569743   35.60314  143.000033 -200.27540
#> 23  -89.764522  -69.479185  -22.111408  -83.20438  -17.345214  -64.88224
#> 24 -184.090993  192.284192  -71.781263   85.10116  117.040750  125.78571
#> 25  -63.544234   93.381346 -130.177601  -47.24068  133.006011 -135.93340
#> 26  -55.560733  144.988933  -55.775306  -48.52663  119.566574 -165.18816
#> 27   31.795836  137.926740   69.096529  -56.77540   -6.450429   69.41034
#> 28  101.490253   -6.672957  352.392849  124.02411   86.508705   17.64697
#> 29  165.202701   90.851947  212.314431  -26.10946   64.412300   90.59569
#> 30  115.176619  144.098059   -4.209229  220.99638  -51.353016  182.43971
#> 31  162.091915   16.694086  -46.424457  176.70034  204.624572  268.72417
#> 32  131.121086   81.582004   -1.890195  171.24663   71.997996  -22.98398
#> 33  144.805977   98.326808  -53.761059 -136.62724  138.350110 -158.92079
#> 34   60.272109  134.275246  160.390694  150.97710  123.039402   23.69791
#> 35   79.052506 -107.780465  351.911572   58.71579  -19.306434  214.65872
#> 36   79.588539   11.719323  114.103378   88.13596 -341.691796  -66.65035
#>          1-8-1       1-9-1
#> X1   -9.997321  103.439584
#> 2   -24.003420  103.189497
#> 3   253.041655  200.450778
#> 4  -110.298960  137.007431
#> 5    -4.475086   76.248189
#> 6   220.776926  154.407312
#> 7    63.050689  127.347681
#> 8   298.225661  251.963493
#> 9   129.036549    5.514281
#> 10   71.232700  213.198929
#> 11  220.058019   38.293668
#> 12  146.236425    6.067369
#> 13   55.195736 -125.983268
#> 14   75.346267   61.601291
#> 15  116.686741   -6.557475
#> 16  122.693667   -3.863883
#> 17   42.122711  172.914266
#> 18   65.600931   70.020759
#> 19   75.579642   90.529869
#> 20   85.038704 -151.787234
#> 21   19.599729    9.596527
#> 22   18.193285  -41.399487
#> 23   39.114108  -29.462882
#> 24  -33.008912  -56.464530
#> 25  271.612243   25.615680
#> 26  177.167959  -26.911989
#> 27  -46.073653  329.003467
#> 28   45.248421  107.547667
#> 29 -128.972370  211.026672
#> 30  172.404240  -96.649552
#> 31  422.945363  -68.035246
#> 32  132.870130  -47.577110
#> 33  113.488740  -24.620099
#> 34   -1.953654   67.712856
#> 35 -194.946202  200.575120
#> 36 -199.580813   84.753254
#> 
#> $covmat
#>              0-1-0      0-10-0      0-11-0     0-12-0     0-13-0     0-14-0
#> 0-1-0  16829.02024  4151.29927 -1970.71985  6328.7456  4431.8804  6278.4924
#> 0-10-0  4151.29927 27857.29589   -19.63495  3600.0122  3247.9147  8286.0931
#> 0-11-0 -1970.71985   -19.63495 15185.09058  1424.3743  4904.4770  3038.1827
#> 0-12-0  6328.74559  3600.01221  1424.37431 17252.1386  1680.3888  7723.5853
#> 0-13-0  4431.88040  3247.91465  4904.47697  1680.3888  9631.7210  3460.2469
#> 0-14-0  6278.49236  8286.09313  3038.18270  7723.5853  3460.2469 23340.8488
#> 0-15-0  1524.97777  1641.65347  4348.11385  1791.7601  2582.0308  7772.0677
#> 0-16-0  3794.89108 -3160.24665   316.71430  1072.7234  -596.0120  3614.7292
#> 0-2-0   6949.63969  2830.58215    66.85909  -147.2595  4334.6879  3594.7497
#> 0-3-0   6359.24771  7192.86726  1120.54718  7749.5243  3707.4469  7030.5110
#> 0-4-0   -517.84290  2047.63984  5374.57849  5264.2215  1497.5369  3540.6260
#> 0-5-0   4720.96904  -227.12329 -2439.38424  1849.3185   726.3526  1610.5004
#> 0-6-0   4396.91479   224.05585  7644.79026  -742.1220  3811.4557  3190.2116
#> 0-7-0   7042.05513  5752.96351 -1114.89701  3228.8507  1488.5192  4775.1088
#> 0-8-0   3645.64268  5611.81636  -678.73338  6800.0512  -905.2270  2882.7518
#> 0-9-0   4693.58377  3755.88787 -2564.34858  3157.5869   230.0917   669.1426
#> 1-1-1   2291.17467  1355.43071 -2698.96298  3098.6800  -384.6519   274.6376
#> 1-10-1 -2674.18925 -2191.05709   148.35554 -3915.3516 -2915.1209 -2231.9992
#> 1-11-1  2103.41800  9111.88680  -729.47549  2423.6422  -553.5038  3656.4009
#> 1-12-1 -3350.50851 -4711.44554  2025.46561  2556.7074   282.3568   732.8096
#> 1-13-1 -3131.63976  -754.59080  3690.94029 -3625.2630 -1472.0314  1016.1417
#> 1-14-1  7389.11509 12629.89782  3277.19916  5083.6464  5089.3455  6694.6474
#> 1-15-1  5062.03270 -3684.55250  4922.13615  1890.3318  1869.3896  3152.2941
#> 1-16-1 10642.94741  4984.80205 -2390.07919  4689.1815  1945.3435  9787.4827
#> 1-2-1   -305.43329  1119.55481  4530.90622  3383.0046  1409.0421  8535.1909
#> 1-3-1   4344.65629  4628.99885 -2253.20688  2159.2354  1719.5024  1484.0879
#> 1-4-1  -1343.13751  5061.49002  7082.72781 -1802.3465   381.3872  5289.5185
#> 1-5-1     60.02175  3481.94377   615.92766  2681.3881  1460.6145  6158.0247
#> 1-6-1   2901.30114  -660.50756 -2160.59188  5281.5712 -1653.3128  8301.2832
#> 1-7-1    444.70616  3553.32150 -1010.55266  5836.0390   194.8325   909.7788
#> 1-8-1   9610.78725  3534.65138 -1807.14911  6551.6704  3799.4850  6670.0341
#> 1-9-1   1319.49595  3503.93943    91.82673 -4697.5807   279.0871   172.1900
#>            0-15-0     0-16-0       0-2-0      0-3-0       0-4-0       0-5-0
#> 0-1-0   1524.9778  3794.8911  6949.63969  6359.2477  -517.84290  4720.96904
#> 0-10-0  1641.6535 -3160.2467  2830.58215  7192.8673  2047.63984  -227.12329
#> 0-11-0  4348.1139   316.7143    66.85909  1120.5472  5374.57849 -2439.38424
#> 0-12-0  1791.7601  1072.7234  -147.25953  7749.5243  5264.22153  1849.31852
#> 0-13-0  2582.0308  -596.0120  4334.68785  3707.4469  1497.53695   726.35261
#> 0-14-0  7772.0677  3614.7292  3594.74968  7030.5110  3540.62596  1610.50045
#> 0-15-0 15760.5514  1919.1516  3936.40674  2453.9280  2068.41455  2461.85090
#> 0-16-0  1919.1516 16581.3199  2194.93894  7174.9858  4628.21601  6611.79294
#> 0-2-0   3936.4067  2194.9389 15182.05953  1669.9192    44.81978  3944.17109
#> 0-3-0   2453.9280  7174.9858  1669.91922 25541.2984  3894.73247   817.33713
#> 0-4-0   2068.4145  4628.2160    44.81978  3894.7325 11231.72548  1815.93769
#> 0-5-0   2461.8509  6611.7929  3944.17109   817.3371  1815.93769 10832.66473
#> 0-6-0   5289.3572  2715.7582  3311.73779   235.1607  2523.34276  -714.26563
#> 0-7-0   1943.3561  9201.5701  5684.98832 11455.6638  2949.82299  6500.91852
#> 0-8-0    809.7209  7540.1178  3651.49354  9180.4434  4424.32255  7379.67149
#> 0-9-0  -1588.0991  6294.0433   564.65854 10047.7341 -1762.05404  -255.07066
#> 1-1-1  -2700.8409  5268.5782  3649.82591  4221.8436  2110.37315  3331.77731
#> 1-10-1  2134.9164  3500.4129  -762.84793 -4322.9850 -1284.92173  2177.13485
#> 1-11-1  2799.7665  1822.9496 -1181.53234  4856.7732  2436.65331  1567.59413
#> 1-12-1  2397.2519  3904.0555 -1467.23405 -3211.1793  3502.39003  3318.22798
#> 1-13-1  2565.7855  2408.8574   138.41523  3377.3571   736.47503   375.20290
#> 1-14-1  3497.6126  2162.9218  4254.86488  5598.6803  2139.81756  1871.13490
#> 1-15-1  5525.8567  4495.5778  4383.44092  1983.5624  1198.76983  3566.24523
#> 1-16-1 -4921.1424  2733.6168 -2489.72004  6220.3718 -1108.41913   175.20179
#> 1-2-1   5170.7375 -1135.0758  2265.46261  3435.4703  -410.90266   -12.39325
#> 1-3-1  -5457.8373  -753.5710   866.95193   677.4141  -620.11895   737.20089
#> 1-4-1   2513.9355   602.0944  1167.61365  -555.0028   418.55373 -1515.23171
#> 1-5-1   -968.4293 -1508.8967  -337.78843   260.5289  1211.37739  -652.00713
#> 1-6-1  -2793.0256  1946.8058  1322.51354  2849.2460  -141.84689 -1894.61576
#> 1-7-1   2651.5175 -3186.0228   541.46601 -4877.9828  -703.35896  1209.27594
#> 1-8-1   2461.4290   312.9346  3568.23799  7550.4986  -394.95269   255.15035
#> 1-9-1   -345.7131  1433.1088  5839.88920 -3239.3830  -613.57114  1642.51014
#>             0-6-0      0-7-0      0-8-0      0-9-0      1-1-1     1-10-1
#> 0-1-0   4396.9148  7042.0551  3645.6427  4693.5838  2291.1747 -2674.1892
#> 0-10-0   224.0558  5752.9635  5611.8164  3755.8879  1355.4307 -2191.0571
#> 0-11-0  7644.7903 -1114.8970  -678.7334 -2564.3486 -2698.9630   148.3555
#> 0-12-0  -742.1220  3228.8507  6800.0512  3157.5869  3098.6800 -3915.3516
#> 0-13-0  3811.4557  1488.5192  -905.2270   230.0917  -384.6519 -2915.1209
#> 0-14-0  3190.2116  4775.1088  2882.7518   669.1426   274.6376 -2231.9992
#> 0-15-0  5289.3572  1943.3561   809.7209 -1588.0991 -2700.8409  2134.9164
#> 0-16-0  2715.7582  9201.5701  7540.1178  6294.0433  5268.5782  3500.4129
#> 0-2-0   3311.7378  5684.9883  3651.4935   564.6585  3649.8259  -762.8479
#> 0-3-0    235.1607 11455.6638  9180.4434 10047.7341  4221.8436 -4322.9850
#> 0-4-0   2523.3428  2949.8230  4424.3226 -1762.0540  2110.3731 -1284.9217
#> 0-5-0   -714.2656  6500.9185  7379.6715  -255.0707  3331.7773  2177.1348
#> 0-6-0  15439.0363   457.3680  -400.2670 -2416.3501 -4841.4119  1268.5950
#> 0-7-0    457.3680 18085.4704 10785.9782  3457.3767  6760.6516  2884.7112
#> 0-8-0   -400.2670 10785.9782 16579.7242  2395.6923  4367.8645   554.8325
#> 0-9-0  -2416.3501  3457.3767  2395.6923 17892.9035  5033.7742  -442.0549
#> 1-1-1  -4841.4119  6760.6516  4367.8645  5033.7742 14983.6723   788.5377
#> 1-10-1  1268.5950  2884.7112   554.8325  -442.0549   788.5377 12285.0684
#> 1-11-1  -852.6224  4496.0932  4000.8228  -900.1225  -735.7633  -963.2610
#> 1-12-1  1173.7670  -544.4449  1371.8090 -3278.2329  1827.1231  4018.5172
#> 1-13-1   226.1791  3106.8166   208.1049  3913.4301  5248.2134  5901.6659
#> 1-14-1  5176.5005  7878.6338  4991.8234   354.9328  2207.3443 -1655.1845
#> 1-15-1  6932.4188  4891.8561  2086.0872  1600.0288  2224.1428  5023.0042
#> 1-16-1  1218.9115  3755.4793 -3801.3476  5310.1937  1901.5908  -951.4222
#> 1-2-1  -1747.6680   173.2301   464.1027 -1424.4135  2525.0819 -1990.8468
#> 1-3-1   -785.4927  3603.9308  1233.4000  1025.1318  3305.7405  2361.7030
#> 1-4-1   4995.7938  1058.7233 -3237.5530   358.8186   562.8210  4639.3081
#> 1-5-1  -2067.3823  1264.3711  1133.5248  -975.2265   541.5419  1403.1628
#> 1-6-1  -1946.8447  3901.2330  4201.1568  2900.4396  5387.2378    50.4639
#> 1-7-1  -4514.7882 -2445.1559 -1540.2719  1335.2740  3300.1963  2892.1396
#> 1-8-1   2183.2008  4283.1684  3186.2849  2290.1467  1698.1181 -3838.0704
#> 1-9-1    355.4094  1000.9703 -2097.9351  2485.7606  4587.8667  4030.6576
#>              1-11-1     1-12-1     1-13-1     1-14-1     1-15-1     1-16-1
#> 0-1-0   2103.417998 -3350.5085 -3131.6398  7389.1151  5062.0327 10642.9474
#> 0-10-0  9111.886801 -4711.4455  -754.5908 12629.8978 -3684.5525  4984.8020
#> 0-11-0  -729.475489  2025.4656  3690.9403  3277.1992  4922.1362 -2390.0792
#> 0-12-0  2423.642167  2556.7074 -3625.2630  5083.6464  1890.3318  4689.1815
#> 0-13-0  -553.503765   282.3568 -1472.0314  5089.3455  1869.3896  1945.3435
#> 0-14-0  3656.400931   732.8096  1016.1417  6694.6474  3152.2941  9787.4827
#> 0-15-0  2799.766489  2397.2519  2565.7855  3497.6126  5525.8567 -4921.1424
#> 0-16-0  1822.949573  3904.0555  2408.8574  2162.9218  4495.5778  2733.6168
#> 0-2-0  -1181.532343 -1467.2341   138.4152  4254.8649  4383.4409 -2489.7200
#> 0-3-0   4856.773179 -3211.1793  3377.3571  5598.6803  1983.5624  6220.3718
#> 0-4-0   2436.653315  3502.3900   736.4750  2139.8176  1198.7698 -1108.4191
#> 0-5-0   1567.594128  3318.2280   375.2029  1871.1349  3566.2452   175.2018
#> 0-6-0   -852.622387  1173.7670   226.1791  5176.5005  6932.4188  1218.9115
#> 0-7-0   4496.093201  -544.4449  3106.8166  7878.6338  4891.8561  3755.4793
#> 0-8-0   4000.822758  1371.8090   208.1049  4991.8234  2086.0872 -3801.3476
#> 0-9-0   -900.122527 -3278.2329  3913.4301   354.9328  1600.0288  5310.1937
#> 1-1-1   -735.763260  1827.1231  5248.2134  2207.3443  2224.1428  1901.5908
#> 1-10-1  -963.261034  4018.5172  5901.6659 -1655.1845  5023.0042  -951.4222
#> 1-11-1 11496.587396   766.5603  -862.9841  4062.9331  -564.1910  3056.8223
#> 1-12-1   766.560270 11497.4383  1988.6861 -1698.6913  4791.6443  -848.3689
#> 1-13-1  -862.984096  1988.6861 17842.9006 -2631.5726  6794.6655 -1528.9357
#> 1-14-1  4062.933094 -1698.6913 -2631.5726 18145.3565   380.2877  3317.7409
#> 1-15-1  -564.190987  4791.6443  6794.6655   380.2877 15234.4821  2723.9996
#> 1-16-1  3056.822264  -848.3689 -1528.9357  3317.7409  2723.9996 25200.9329
#> 1-2-1      7.314876  -813.1260  1883.3705  1031.9034  1045.1941  1031.8285
#> 1-3-1  -1490.479753  -851.2386  -946.9827  3427.6009  1160.2054  7381.1339
#> 1-4-1     84.266347  3185.7907  6987.6715   187.3919  7872.6639  4808.0472
#> 1-5-1    844.898292  -340.2292  -469.0261   359.4885   194.6801  2742.8554
#> 1-6-1  -1207.303377 -1849.1242  -974.0909    -4.7462   569.4978  6060.9845
#> 1-7-1   2197.847961  6131.2986  1613.4564  -461.0776  1413.4419   570.4748
#> 1-8-1   3297.436578 -3551.6122 -5621.6521  4616.9133  1327.7216  5280.9719
#> 1-9-1  -2033.246639  1007.7985  6924.8629 -1057.1664  3682.3297  2379.3007
#>               1-2-1      1-3-1       1-4-1       1-5-1      1-6-1      1-7-1
#> 0-1-0   -305.433292  4344.6563 -1343.13751    60.02175  2901.3011   444.7062
#> 0-10-0  1119.554814  4628.9989  5061.49002  3481.94377  -660.5076  3553.3215
#> 0-11-0  4530.906219 -2253.2069  7082.72781   615.92766 -2160.5919 -1010.5527
#> 0-12-0  3383.004635  2159.2354 -1802.34650  2681.38812  5281.5712  5836.0390
#> 0-13-0  1409.042126  1719.5024   381.38723  1460.61454 -1653.3128   194.8325
#> 0-14-0  8535.190859  1484.0879  5289.51847  6158.02470  8301.2832   909.7788
#> 0-15-0  5170.737515 -5457.8373  2513.93552  -968.42931 -2793.0256  2651.5175
#> 0-16-0 -1135.075760  -753.5710   602.09444 -1508.89668  1946.8058 -3186.0228
#> 0-2-0   2265.462611   866.9519  1167.61365  -337.78843  1322.5135   541.4660
#> 0-3-0   3435.470340   677.4141  -555.00282   260.52888  2849.2460 -4877.9828
#> 0-4-0   -410.902657  -620.1189   418.55373  1211.37739  -141.8469  -703.3590
#> 0-5-0    -12.393248   737.2009 -1515.23171  -652.00713 -1894.6158  1209.2759
#> 0-6-0  -1747.668018  -785.4927  4995.79376 -2067.38230 -1946.8447 -4514.7882
#> 0-7-0    173.230116  3603.9308  1058.72327  1264.37113  3901.2330 -2445.1559
#> 0-8-0    464.102692  1233.4000 -3237.55303  1133.52479  4201.1568 -1540.2719
#> 0-9-0  -1424.413465  1025.1318   358.81860  -975.22653  2900.4396  1335.2740
#> 1-1-1   2525.081885  3305.7405   562.82103   541.54189  5387.2378  3300.1963
#> 1-10-1 -1990.846798  2361.7030  4639.30810  1403.16277    50.4639  2892.1396
#> 1-11-1     7.314876 -1490.4798    84.26635   844.89829 -1207.3034  2197.8480
#> 1-12-1  -813.125971  -851.2386  3185.79074  -340.22917 -1849.1242  6131.2986
#> 1-13-1  1883.370523  -946.9827  6987.67146  -469.02608  -974.0909  1613.4564
#> 1-14-1  1031.903386  3427.6009   187.39188   359.48852    -4.7462  -461.0776
#> 1-15-1  1045.194055  1160.2054  7872.66387   194.68011   569.4978  1413.4419
#> 1-16-1  1031.828485  7381.1339  4808.04718  2742.85539  6060.9845   570.4748
#> 1-2-1  14395.296958 -1822.9883  4044.39910  4156.65550  5122.3882  1625.5345
#> 1-3-1  -1822.988277 12399.7591  2074.92867  3213.92700  5376.3092  -886.6992
#> 1-4-1   4044.399095  2074.9287 19147.08002  1344.05052 -1564.6967  3754.4553
#> 1-5-1   4156.655496  3213.9270  1344.05052 11483.68154  4240.8316  1766.3647
#> 1-6-1   5122.388221  5376.3092 -1564.69669  4240.83165 18736.0334 -1825.3911
#> 1-7-1   1625.534455  -886.6992  3754.45526  1766.36466 -1825.3911 18758.0326
#> 1-8-1   1522.803269  1571.4138 -5088.66980  2306.09241  6669.9345  1520.1663
#> 1-9-1   1002.925342  3078.7112  8819.44075  -977.91865  -847.6050  4930.3578
#>             1-8-1       1-9-1
#> 0-1-0   9610.7873  1319.49595
#> 0-10-0  3534.6514  3503.93943
#> 0-11-0 -1807.1491    91.82673
#> 0-12-0  6551.6704 -4697.58074
#> 0-13-0  3799.4850   279.08711
#> 0-14-0  6670.0341   172.19000
#> 0-15-0  2461.4290  -345.71307
#> 0-16-0   312.9346  1433.10882
#> 0-2-0   3568.2380  5839.88920
#> 0-3-0   7550.4986 -3239.38299
#> 0-4-0   -394.9527  -613.57114
#> 0-5-0    255.1503  1642.51014
#> 0-6-0   2183.2008   355.40935
#> 0-7-0   4283.1684  1000.97031
#> 0-8-0   3186.2849 -2097.93509
#> 0-9-0   2290.1467  2485.76057
#> 1-1-1   1698.1181  4587.86675
#> 1-10-1 -3838.0704  4030.65763
#> 1-11-1  3297.4366 -2033.24664
#> 1-12-1 -3551.6122  1007.79852
#> 1-13-1 -5621.6521  6924.86292
#> 1-14-1  4616.9133 -1057.16639
#> 1-15-1  1327.7216  3682.32968
#> 1-16-1  5280.9719  2379.30068
#> 1-2-1   1522.8033  1002.92534
#> 1-3-1   1571.4138  3078.71117
#> 1-4-1  -5088.6698  8819.44075
#> 1-5-1   2306.0924  -977.91865
#> 1-6-1   6669.9345  -847.60499
#> 1-7-1   1520.1663  4930.35780
#> 1-8-1  17439.7378 -3771.20803
#> 1-9-1  -3771.2080 12348.06352
#> 
#> $stimcats
#>      stim cat
#> 1   0-1-0   0
#> 2  1-14-1   1
#> 3   0-2-0   0
#> 4   1-3-1   1
#> 5   1-4-1   1
#> 6  1-11-1   1
#> 7   0-8-0   0
#> 8   1-9-1   1
#> 9  1-16-1   1
#> 10 1-10-1   1
#> 11 0-14-0   0
#> 12  0-5-0   0
#> 13 1-15-1   1
#> 14  0-4-0   0
#> 15  1-5-1   1
#> 16  1-6-1   1
#> 17 1-13-1   1
#> 18 0-10-0   0
#> 19 0-11-0   0
#> 20 1-12-1   1
#> 21 0-16-0   0
#> 22  0-7-0   0
#> 23  1-2-1   1
#> 24  1-8-1   1
#> 25  0-6-0   0
#> 26 0-12-0   0
#> 27 0-13-0   0
#> 28  0-3-0   0
#> 29  0-9-0   0
#> 30  1-1-1   1
#> 31 0-15-0   0
#> 32  1-7-1   1
#>