Correlation tools

Helper functions to compute important statistics from correlation coefficients.

r2z(r)

z2r(z)

r2t(r, n)

r2p(r, n)

rconfint(r, n, alpha = 0.05)

compcorr(r1, r2, n1, n2)

Arguments

r, r1, r2
a correlation value

z
a Z-score

n, n1, n2
sample sizes

alpha
the significance level to use

Functions

r2z: converts correlation coefficients to z-scores

z2r: converts z-scores to correlation coefficients

r2t: Converts correlation coefficients to t-scores

r2p: Computes the two-sided p-value for a given correlation

rconfint: Computes confidence intervals for a given correlation coefficient

compcorr: computes the significance of the difference between two correlation coefficients

See also

Examples

z <- r2z(.5)
r <- z2r(z)
t<-r2t(r,30)
r2p(r,30)

#> [1] 0.004899934

print(rconfint(r,30))

#> function (r, n, alpha = 0.05) 
#> {
#>     z <- r2z(r)
#>     zint <- qnorm(1 - alpha/2) * sqrt(1/(n - 3))
#>     confints <- c(z2r(z - zint), z2r(z + zint))
#>     return(rconfint)
#> }
#> <bytecode: 0x000000002352c688>
#> <environment: namespace:AATtools>

print(compcorr(.5,.7,20,20))

#> $zscore
#> [1] 0.927105
#> 
#> $pvalue
#> [1] 0.353872
#> 
#> attr(,"class")
#> [1] "compcorr"