Covariance Matrix-Based Reliability Coefficients

These functions allow for the computation of the reliability of a dataset from the covariance matrix of its variables.

calpha(covmat)

lambda2(covmat)

lambda4(covmat)

Arguments

covmat

a covariance matrix

Functions

calpha(): Cronbach's alpha

lambda2(): Guttman's Lambda-2

lambda4(): Guttman's Lambda-4. This algorithm tries to get the highest attainable reliability by

See also

Examples

# compute reliability from covariance
h<-cov(iris[,1:4])
calpha(h)
#> [1] 0.7084772
lambda2(h)
#> [1] 0.8530156
lambda4(h)
#> [1] 0.8136558
# Lambda-2 and Lambda-4 are significantly larger because
# some of the variables in the iris dataset are negatively correlated.