Multiple correlation Computes the multiple correlation coefficient of variables in ymat with the variable x

Multiple correlation Computes the multiple correlation coefficient of variables in ymat with the variable x

multiple.cor(x, ymat, use = "everything")

Arguments

x
Either a matrix of variables whose multiple correlation with each other is to be estimated; or a vector of which the multiple correlation with variables in ymat is to be estimated

ymat
a matrix or data.frame of variables of which the multiple correlation with x is to be estimated

use
optional character indicating how to handle missing values (see cor)

Value

The multiple correlation coefficient

See also

Examples

multiple.cor(mtcars[,1],mtcars[,2:4])

#> [1] 0.876292