Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

aat_stimulusscores(
  ds,
  subjvar,
  stimvar,
  pullvar,
  targetvar = NULL,
  rtvar,
  aggfunc = c("mean", "median"),
  iters = 5
)

Arguments

ds

the data.frame to use

subjvar

Name of the subject-identifying variable

stimvar

Name of the stimulus-identifying variable

pullvar

Name of the movement-direction identifying variable

targetvar

Optional. Name of the stimulus-category identifying variable

rtvar

Name of the reaction-time identifying variable

aggfunc

The function with which to aggregate the RTs before computing difference scores. Defaults to mean but can be changed to median.

iters

If there are missing values (which is almost inevitable) then multiple imputation will be used to complete the covariance matrix - this argument sets the number of multiple imputations to be used.

Value

Exports a list containing a data.frame with stimulus-specific bias scores, indicated in the column names, a covariance matrix of that same data, and a data.frame indicating to which stimulus category each stimulus belongs.

Examples

ds<-aat_simulate(biasfx_jitter=40,nstims=16)
ds$stim<-paste0(ds$stim,"-",ds$is_target)
aat_stimulusscores(ds,"subj","stim","is_pull","is_target","rt")
#> $data
#>          0-1-0     0-10-0       0-11-0     0-12-0      0-13-0      0-14-0
#> X1 -127.913009  -13.16562  -43.1713239   69.31695   41.770877  173.031312
#> 2  -192.969015  -36.59142  -70.0843184 -268.68926 -239.657173  296.967826
#> 3   406.067949  308.40363  129.4016246  278.43188  132.857559  109.404258
#> 4    34.563864  219.27202   -0.7367699   66.79166   50.231468   31.833787
#> 5   114.291683 -223.59586   19.7005461 -108.74509  168.813331   93.268880
#> 6    24.545607 -138.89707    0.2516157  159.52317  -19.123872  -69.947253
#> 7  -250.974777  -27.66263  172.5013541   59.37705  229.437158  -57.386262
#> 8    48.431457 -114.71084   60.1482645   29.54433   21.637396   76.716198
#> 9  -144.651087  -47.68043  -47.5964496   59.66401   84.967537  -16.452738
#> 10   39.241566  141.91355  107.9848500  126.33601  406.476399 -271.758306
#> 11  -95.192532  -48.30424 -109.2631529   23.90495 -168.475103   37.735500
#> 12  177.849966  -16.53098  150.0687213  142.87209   -2.372025   20.459162
#> 13   30.713180  -33.35557   75.8059671  121.31489   90.226441   24.809884
#> 14   -9.832876   21.86214   46.4350263 -176.92695  -73.261097  117.789298
#> 15   42.057395  -19.53120  345.0515400  130.74784  162.157464  182.366000
#> 16  205.480410  159.69940  125.6400330  -73.57547  -53.757003  144.020276
#> 17  196.991854  -79.84105   82.7383503  212.24988  207.098718  229.521308
#> 18  104.161927 -114.68255  -68.5301824 -164.83340  -96.658202  -46.284650
#> 19   86.553066  -50.17388   46.6366962  208.27201   41.045863   49.628444
#> 20   15.346377   86.16464 -191.3993130   61.25143  -62.439449 -137.445203
#> 21    5.526229  -66.16983  -90.7399978   27.85940 -126.622136  -83.585604
#> 22  -51.862325  230.00086  -57.7305615   14.80213 -201.168693  -17.491780
#> 23  142.143303   71.73306  -24.1246730   92.41573  -15.505092  113.470114
#> 24  -33.911224  -22.63128  112.7895004 -131.60503 -221.001002 -135.387991
#> 25   22.416658   42.05384   31.5252358  153.85794   73.034578   -7.791748
#> 26  -37.714455 -118.14700  -19.9242344   13.12526   19.685027    6.850642
#> 27  -98.811669   60.96466   14.7392214  -21.56261  -38.332706   48.881840
#> 28   -8.870595  128.78169  109.9013265  -23.68887  -38.452414 -186.751463
#> 29  -37.995375  -42.04554   28.6055405  173.71222  -97.828067  178.497307
#> 30   61.852463   80.35198   77.6100994   94.68307  117.976097  -67.914347
#> 31 -142.217151  -10.95141  -85.7717388  -28.49326 -177.369300 -195.975499
#> 32  124.615942   53.75247   64.5651496 -185.31345  -70.868996   79.918775
#> 33 -126.738533   55.79958  -91.1961225  -43.64516   82.381646  132.555910
#> 34   51.249827  -14.14499  130.5391525  128.17033  105.439172  216.461799
#> 35   24.685596   45.87349   21.6495214  221.08292  445.121644  131.560162
#> 36  119.016681   73.54471   46.3233560   91.07925  -10.331237  170.131397
#>          0-15-0      0-16-0       0-2-0       0-3-0       0-4-0        0-5-0
#> X1 -112.1379334  106.181735   28.804533  -85.002978  -34.050629  134.1754338
#> 2  -113.4234433 -618.672112 -183.734913  -80.771552 -115.049846   96.0711266
#> 3   234.4755302   -4.758784   -8.081897  158.611373  244.946745 -146.3160192
#> 4   -83.0451616  163.388573  310.966501  -26.195486   13.937006   -6.1814412
#> 5    52.5106853   57.110655   10.981348   55.792162   20.879809   22.1349973
#> 6   118.1108213  127.677135    9.493912  134.276001  107.873440   -0.7444983
#> 7   -54.4678189   25.443366   25.787690  183.716403  -48.547902  142.3187677
#> 8   -81.7893330  110.459391   66.360146  -82.155904 -186.567113  105.0057220
#> 9   -80.8835537  -24.273473 -146.354553   90.668101  -42.513474   80.5065956
#> 10  293.3622851  498.001462 -113.666202  124.044376 -180.422707 -286.4700687
#> 11   -0.3281095 -153.534341  -48.122028 -123.290850 -145.991342  -97.7053938
#> 12   27.0046945   18.381248  -28.771779    1.757083  161.321770  -99.5295002
#> 13   13.1619136   -5.809385  -40.480291  -14.388017   34.899915  -35.4310618
#> 14   32.8489045   -5.837717 -132.701013  -30.953034 -100.674546  125.7693200
#> 15 -280.8074766   12.247649  -78.932720  -29.374416  334.720963 -109.2384240
#> 16  154.8722663  -11.195456  275.181029   62.356184  -36.803719  117.2092925
#> 17  -65.9168883  325.960781  249.389241  251.244802  176.441193  208.2435144
#> 18 -116.8538276  -80.710744  -72.581112   -2.394180  -70.338269 -120.6030750
#> 19  121.3398631   -1.821807  132.655717  287.415913   11.458032  161.6472842
#> 20   32.6862661 -150.691683  -84.845656   13.966881 -303.808787 -218.2562172
#> 21  -41.5149602   42.940413    2.834113 -156.420223   38.934221  100.6972586
#> 22   69.3088155  266.331293   96.007346   36.666641 -113.754774 -162.5687477
#> 23   35.5624256  -64.976471   59.902203   58.934464   -3.424525   49.0615893
#> 24 -221.9015712   46.895039   -5.108915  -84.480772  169.480436  -86.1793224
#> 25   37.1833914   -8.741318  -61.583373  148.566282  180.288250   74.4474557
#> 26   40.7857252   65.125494  149.690005  -13.187090  -66.871755  101.7051109
#> 27  -91.0573315    7.940403  274.013041  -46.783492  -54.708401  -20.2290072
#> 28   49.3264857   -5.841666  -11.149219    9.342808  100.536719  204.5079083
#> 29   40.6746952  -19.486841  -78.320305  193.992411  203.529228 -242.4336393
#> 30  140.2396449   90.491930  209.254935   67.649461  -22.499634   60.1792032
#> 31  -65.4564854  122.439557  417.180354  110.938454   89.131443   32.3257092
#> 32   13.7176221   64.216274  -92.103002   50.099463   33.179358   60.8308892
#> 33    7.1312105    7.581963  -93.183484   60.074440  -72.084244  -44.0013996
#> 34  143.1147924  106.332886  222.956988  121.462788  229.464517   35.1837151
#> 35   68.6232575 -133.031713 -107.350695  243.686519   36.200451  232.8878946
#> 36  248.7863237   98.295874  -69.870359 -159.439601  145.483785  101.7862477
#>          0-6-0        0-7-0       0-8-0        0-9-0       1-1-1      1-10-1
#> X1 -134.543612   74.9395404   70.996089    0.3689566   78.896628   27.658605
#> 2   -52.119355 -402.0184830 -222.335944   44.9439091   36.230563  120.910514
#> 3   113.152803  -52.4265020   36.711048   43.4927699    2.633687   33.703402
#> 4   147.875790   49.2452789  104.961337  108.8435114  188.037258  196.260447
#> 5   137.514991 -126.1838803   76.771380  165.5838281  197.260273   20.950648
#> 6   172.105609   42.7223091  146.958884  180.9066061  168.183942  305.643057
#> 7    26.123366   80.9844730   29.203774   -1.9822025  348.972516  163.151419
#> 8    68.526386  135.7317300  -64.139498  124.9669230 -159.506941 -217.035858
#> 9    29.350928  -39.2638130   14.024702   31.1203785  -32.502941   86.777959
#> 10  -92.098656  -10.3172705  101.656220    4.0376707 -163.150959  -28.900449
#> 11 -262.818309   -4.6803043  -75.002360 -151.6956633  -72.858444    3.971424
#> 12   37.496790   11.4893130  -33.549551   69.9697415   12.391325  161.635581
#> 13   52.604496   13.7521322  -34.508972  -44.3175304   74.037178  260.278758
#> 14  210.102140  -79.7954040  -40.695782 -191.5371632   59.816003   93.014463
#> 15   96.348749  -44.9626823  541.464425  130.3512726  229.464164  272.055726
#> 16 -230.952954  -61.5573178 -130.420740  172.6075223 -224.151628   76.312112
#> 17  159.122174   37.7786730  134.774183   86.7572446  106.477772  -89.589270
#> 18   30.327637   17.9258361   32.514270   82.2847837  -25.791544  103.201513
#> 19   41.049532   10.2511082 -116.394317  -56.5010901 -103.843412  -29.673134
#> 20  -33.067724  -36.6860865   20.986262  -99.3467846  215.005060  111.291080
#> 21  -45.226325   31.7325823   76.076720    3.6292322   70.246266   76.645680
#> 22  -74.464433  -85.8611082  138.272121   -4.3010735  -18.094058 -103.201138
#> 23   80.512696   53.9723966   42.954306   42.1660245  -34.774435   86.730219
#> 24  -59.638384  204.0144253   33.993010  149.7514810  266.000335  135.591867
#> 25  -34.774367  -11.6729463   55.397095   37.6743456  131.662130   54.415214
#> 26   19.090959 -119.2104364   63.556165   24.4913130    4.118973   36.448007
#> 27   94.296619 -117.1942778   -4.582434   38.8130420   38.552533  -15.412181
#> 28    8.097291 -113.3839077  -76.827668  -61.7676407 -196.594994  143.629302
#> 29  103.640862   65.7699571  107.584027   63.9109009  118.383800  202.684224
#> 30 -105.970983    0.8245351   92.013289  -50.9337207   23.719379  134.029384
#> 31   33.421847  -83.1906047 -110.174847  -68.8814662  -35.510831  -50.271345
#> 32   -2.809058  102.6662036  -17.601313 -283.4427973   22.597725  183.521598
#> 33  -86.438094  205.7561185  165.829045 -116.6415006  -15.457465   54.805085
#> 34   69.459169  205.4920281  224.526311  191.8419263  183.899973   -8.835358
#> 35  -95.277343  -43.6236452 -155.525735 -165.7321582  112.730417   78.312918
#> 36   23.523226   17.2420608  -44.451731 -127.8208928   51.683157  122.082667
#>          1-11-1      1-12-1      1-13-1      1-14-1      1-15-1      1-16-1
#> X1   91.5602450  249.392324  157.550405 333.3146925   25.784225   39.829885
#> 2    74.0790117  -74.103542  304.199280 -62.2229441   84.679558    4.658662
#> 3    20.0041384  139.362313  -14.633526  40.2359346  -25.328821   11.724411
#> 4    90.6117453  213.221819   61.703716 173.4995780   47.902194  178.631770
#> 5   156.8855214   45.167657   10.220599  56.4557503   80.821244  176.877580
#> 6    16.5945360   84.769813  295.085186 320.0905499  143.064602  105.404432
#> 7    91.5592172  -37.310733   -9.564187  77.9849947   81.889983  111.599847
#> 8  -267.5465455  -77.812589  234.494729 216.7331787  137.661870  203.505764
#> 9   -17.5661948  -38.861127  -32.538098  90.3462301   24.445189   -9.573915
#> 10  -10.4995172   -6.838781    6.825339 -78.3608348  -52.564167  186.318427
#> 11  -42.4237274  -80.558050  -72.753319 -49.6521293  -34.947975 -158.538772
#> 12  -61.4036862 -190.555763  141.784671 124.1886805  164.261270  235.863603
#> 13  150.6472567  -23.288477   99.192677 197.6647375   69.676201  151.175540
#> 14 -276.5014693  -56.380920  132.278191  70.0031268   61.907114   46.621752
#> 15  172.8740789    7.153313  368.874590 171.3414521   11.106865  237.427581
#> 16  210.6731659 -133.503536 -219.530118 -12.0382055   64.367851  -14.077712
#> 17  129.6368237   59.897934  138.691891  89.0216428   -2.372165  247.497166
#> 18    0.0379450 -114.080165  -47.798086  55.1099110   38.541482  -63.043203
#> 19    3.6008824  337.996016  147.379880  68.5869470  167.884685   41.086419
#> 20   54.2948739  229.313640  350.932735  28.4290272   90.371404  165.563383
#> 21  122.0199630   49.749506   22.750068 142.6623207   43.837702  -77.060504
#> 22   -4.1593866  210.159589   98.088629  11.4224176   64.661951   75.000643
#> 23   90.2786634   42.198578   27.182692   0.6975325   85.236591  176.581566
#> 24  211.9538259   44.946348  318.289833 -88.6834499 -292.128539  195.580944
#> 25   83.7914690   62.055542   24.402922  74.7273843  -33.441230  -40.997396
#> 26  227.7367569  -36.046468  -83.963665  32.1605400   53.193647 -134.410635
#> 27 -341.0992454  286.432312   52.691500 -83.2711222 -170.502528   37.426238
#> 28 -127.9052209   -7.224447  244.419041  29.4607688  -73.075287 -133.531371
#> 29 -117.0215386   -7.988508    7.381566 -19.8979550  130.785590    8.689202
#> 30  -27.1035158  -39.010912   51.708634 -89.6671324 -158.230790  255.375948
#> 31   89.0414145  253.977828  182.550874  89.4772252  124.491520  130.985148
#> 32  -74.3708025  189.525943   76.297986 -21.5491067   -8.339741  207.126647
#> 33  351.4561396 -104.609187 -124.780642  39.6598365  -27.576846  170.639863
#> 34  -78.0002402  279.862068  162.473694  50.8344135  377.119504  255.320348
#> 35  206.8309280  305.934730    1.031050  20.2642554  -69.242048  261.503596
#> 36   -0.6708494   66.107149  181.221430 -85.0411384  109.352594   59.320328
#>          1-2-1      1-3-1       1-4-1      1-5-1        1-6-1       1-7-1
#> X1  123.012456  199.45137  129.218236 -148.46023  279.4264515  278.220279
#> 2   -39.209713   25.18518 -363.173915 -158.08558  223.5334806   55.866429
#> 3   119.483038   55.42960    5.825235  294.52074  140.1988888   95.589163
#> 4   144.317065  241.97137  240.300431   50.79102   77.7138119  178.293053
#> 5    24.453808   80.01674  -52.780086  131.51167  -54.3748654   -8.184786
#> 6    12.627498   14.68026  151.581160  271.67399  298.7020568   11.054245
#> 7   278.755542  111.49706 -168.262625  184.95119  187.8053354  105.379648
#> 8     1.451094  -30.80931   67.451301  391.18895   50.9621594  258.255175
#> 9    52.341899   87.28640  -84.411603   62.52414    4.8116027   37.570352
#> 10  184.757902  119.34563  183.006275  -45.58411  175.7552024   11.472304
#> 11 -105.767286   11.03364  -61.435233 -126.91018   -0.4897564 -120.943142
#> 12  -12.102161  153.08132 -124.978426   96.62871  -16.1199668   31.924960
#> 13  173.368708   94.49816  137.420537  -52.93184  125.5922687  145.554029
#> 14  -14.635492  145.26316  172.779098 -146.85480  164.1074957  119.803920
#> 15  221.280276 -124.13802  -43.788120   75.76732 -155.6411332   81.736946
#> 16 -157.671471  -58.45770  116.258640  129.25849 -211.6962242    2.332078
#> 17   72.256223  -26.96461  114.294338   81.75131   80.8650046  -17.348116
#> 18   -1.052413   37.97987   67.876448  158.40016   93.5689857  109.009488
#> 19   53.149002 -164.56350  -21.833857  102.08014  141.7717293   69.182752
#> 20   86.913491   87.06917  265.966386  289.43126  135.0162419  236.751896
#> 21  168.994506   57.87658   41.834766  104.49964  140.7061130  105.542883
#> 22  244.554889   17.31496  109.253914  194.47326  -31.4507499  204.089045
#> 23   14.621628   83.89479   34.559590   92.44191   37.1667462  108.576238
#> 24  -47.474745  255.27357  169.513897  140.56272   40.1181653  181.247204
#> 25  212.392682  110.42387   33.853943  -12.66218  -55.4829410  158.328901
#> 26   23.263177  133.48806   19.568721   12.05657  104.0797904   39.794824
#> 27  143.369582  -56.27294  442.055362  125.43388   79.3070703   84.540841
#> 28  106.672699  -19.78655  -29.396503   25.90857 -173.1969266 -135.993170
#> 29  -55.890895   62.91899 -204.232502  -71.40265   28.3552492  102.286289
#> 30   44.998098  143.11113   97.051775   18.06592  106.5180407   -3.153976
#> 31  -66.787492   88.51592  268.010332  265.12945  130.0294672 -329.183321
#> 32   66.692648  -72.19009  128.920983  162.52627  -56.1679464  156.621220
#> 33 -283.795150   64.86992  155.218265   60.99840  115.1714448 -255.794315
#> 34  315.835411  179.56345  257.367668  279.20806   -7.4292145  314.858107
#> 35  281.102183  105.58663  -12.130461  181.68192  161.9141833  261.560859
#> 36   10.685163  -20.39343 -155.799795  195.46679   -4.6758994  264.437235
#>          1-8-1      1-9-1
#> X1   45.831180  178.67417
#> 2   105.041887  -92.93037
#> 3     1.041216  305.33865
#> 4   197.102446  341.88843
#> 5   117.500617   72.14000
#> 6   130.780074  -11.85997
#> 7   151.785737   98.33206
#> 8   -67.328147  206.46923
#> 9   114.343564   50.44648
#> 10  -86.814169 -131.62459
#> 11  -75.035004  123.89769
#> 12  -81.724066  142.14530
#> 13  164.493312   92.94879
#> 14   12.675656   95.02336
#> 15  107.998103  113.12802
#> 16  100.818233  194.69569
#> 17   23.561277  145.58551
#> 18  -81.445062   58.30957
#> 19  -76.645614  -27.47913
#> 20    3.847060  383.64681
#> 21   68.743241  104.15694
#> 22   17.281259  -71.90347
#> 23  216.385790   80.11931
#> 24  140.248176   48.59745
#> 25  133.006324  160.99816
#> 26 -154.187323  -45.37622
#> 27  164.773637   44.50890
#> 28  -48.146024  132.87984
#> 29  -20.903573   28.38986
#> 30  166.994068  120.62264
#> 31   91.805762  185.10863
#> 32   62.184376   80.86674
#> 33 -243.832905 -156.68436
#> 34   52.062455   24.13375
#> 35   36.628624   70.36203
#> 36   65.915777  180.60026
#> 
#> $covmat
#>              0-1-0     0-10-0     0-11-0     0-12-0     0-13-0      0-14-0
#> 0-1-0  16007.34718  3747.9513  4419.5915  5588.3142  4217.5938  2833.97081
#> 0-10-0  3747.95128 12079.0084  1112.2391  2417.3491   893.0628 -1446.96349
#> 0-11-0  4419.59150  1112.2391 10000.8952  3483.5918  6546.7784  1974.30272
#> 0-12-0  5588.31418  2417.3491  3483.5918 16150.2135 11076.5403   456.66318
#> 0-13-0  4217.59382   893.0628  6546.7784 11076.5403 23905.4400   846.88587
#> 0-14-0  2833.97081 -1446.9635  1974.3027   456.6632   846.8859 16859.01545
#> 0-15-0  6925.36923  5130.2247   206.7157  5631.2382  5697.4910 -1465.13452
#> 0-16-0  4889.33627  2785.6860  4448.3333  7642.6342  9607.1778 -7565.71419
#> 0-2-0   1832.65432  1975.1909   629.2364  2370.5358 -1552.2532 -1269.49260
#> 0-3-0   2810.46749   711.8105  2281.2255  7462.2454  9170.6166   147.51128
#> 0-4-0   6469.24336  -132.4490  8324.8575  7122.6655  2560.6820  4719.35747
#> 0-5-0  -1479.90088 -3240.9968   784.6873 -1221.1022  2774.7693  4639.88343
#> 0-6-0   3121.76418 -2402.5464  2580.8015  2097.9152  1812.0180  2091.14692
#> 0-7-0   1615.54710  -707.4699  1760.4314  3863.9502  2690.1995  -826.57765
#> 0-8-0   1254.73869  -423.6127  4850.3529  4828.2937  5076.7415  1250.69017
#> 0-9-0   2583.53388 -2769.0250  3478.1084  1930.5473   392.1879  1477.96594
#> 1-1-1  -3219.31059 -3337.8833  1838.5941  1761.2069  2665.5463  2153.08429
#> 1-10-1  -270.75156  -184.0712  2676.6874  -186.0052  -561.1906  -247.45264
#> 1-11-1  -127.44331  -842.0961  -741.1407  1215.1884  4601.6308   987.96706
#> 1-12-1   -56.38783  3614.3265 -2264.0012  5068.4707  2203.7260    43.10286
#> 1-13-1 -1708.85992 -2110.6086  2097.4087  -300.4453 -4792.8594  -936.68110
#> 1-14-1  -380.34912 -3568.9520   362.5530  3584.8592  1945.1260   734.07728
#> 1-15-1  1139.10053 -3059.5376  -501.6754  2456.3206  -782.9398  4501.79538
#> 1-16-1  3187.18706   357.9054  4293.5892  3341.9480  8336.5573  1894.07217
#> 1-2-1    130.20685  2550.7696  4194.6899  6784.7718  9246.2576 -1601.85551
#> 1-3-1  -2374.39324   635.4560 -1386.9001  -189.0714  1231.2330 -2658.11111
#> 1-4-1   1195.82507  3993.3992 -2160.5627  -443.9128   239.1594 -6491.99312
#> 1-5-1   5628.98992   910.6896   985.5443  3315.6341  1371.8743 -2962.15427
#> 1-6-1  -4310.74354 -2193.3712 -4452.3071  1670.4852  2127.0803 -1286.39512
#> 1-7-1   3007.66625  1353.0266  1768.9611  2955.0426  2454.2832  5182.53682
#> 1-8-1   -447.82306   875.4341  1829.1992  -117.0150  -436.9350   495.11976
#> 1-9-1   5486.42761  3578.9104   371.4501  3152.2805  -581.0795  -505.76761
#>             0-15-0     0-16-0       0-2-0      0-3-0      0-4-0      0-5-0
#> 0-1-0   6925.36923  4889.3363  1832.65432  2810.4675  6469.2434 -1479.9009
#> 0-10-0  5130.22475  2785.6860  1975.19091   711.8105  -132.4490 -3240.9968
#> 0-11-0   206.71569  4448.3333   629.23639  2281.2255  8324.8575   784.6873
#> 0-12-0  5631.23824  7642.6342  2370.53584  7462.2454  7122.6655 -1221.1022
#> 0-13-0  5697.49098  9607.1778 -1552.25322  9170.6166  2560.6820  2774.7693
#> 0-14-0 -1465.13452 -7565.7142 -1269.49260   147.5113  4719.3575  4639.8834
#> 0-15-0 14850.07250  5742.2472    83.27367  4562.3232  -822.5966 -1264.1075
#> 0-16-0  5742.24717 27984.5558  9303.89027  4013.6896  2950.2912 -3190.8286
#> 0-2-0     83.27367  9303.8903 21611.34890  2587.7812  2203.4204  4112.4255
#> 0-3-0   4562.32324  4013.6896  2587.78117 12799.2430  3789.9525  1309.4908
#> 0-4-0   -822.59665  2950.2912  2203.42036  3789.9525 19095.9882  1153.7911
#> 0-5-0  -1264.10752 -3190.8286  4112.42545  1309.4908  1153.7911 16950.0869
#> 0-6-0  -1284.86693  2968.6048  1506.05720  2526.4204  5408.8689   538.8061
#> 0-7-0   -218.08875  8535.2401  1344.86078  1258.3905  3229.5708 -1418.5278
#> 0-8-0  -3178.34017  9360.3580   397.98314   594.6381  7278.8509 -6243.6786
#> 0-9-0  -2523.06325  2579.7375  5233.15016   640.3278  4962.8528 -2128.5132
#> 1-1-1  -5582.86107 -1393.6423 -1593.88858   868.5070  5841.2566 -1292.7428
#> 1-10-1 -1603.40465 -5109.0243 -5474.48940 -1004.8123  5063.6678 -1810.6369
#> 1-11-1 -1942.00333 -1288.8695  -159.19403  2085.6063  2915.9811  1223.1099
#> 1-12-1  1347.95649  4011.6671  7433.40771  4534.8880  2066.3694  1902.9005
#> 1-13-1 -5378.95203 -2194.6284 -2038.14642 -3187.9236  3537.2639  -435.4730
#> 1-14-1 -3634.99001  2959.0860   910.73264   -46.1521  1946.7422  2549.2705
#> 1-15-1  3209.41432  -529.3703  1793.13346  2029.0107   938.7905   745.6137
#> 1-16-1  -498.82381  6362.4207  2455.91306  3626.6893  2851.1323 -1227.9211
#> 1-2-1    645.24925  5302.0502   515.56706  3036.4915  3483.5132  1176.2774
#> 1-3-1   -492.57557  1693.9343   243.93548 -1218.8080  -421.6706 -1617.8647
#> 1-4-1    -76.35532 12378.7621 11768.24960  -302.2708 -3507.5962 -2974.3376
#> 1-5-1   2148.78939  4988.8932  5924.22144  2972.1638  2088.4960  -559.7027
#> 1-6-1    346.85952  -907.7392 -1448.84802  1323.3939 -4935.3803   275.4511
#> 1-7-1    224.56052  -158.2329 -3871.19739 -2225.7318   271.7479   367.0400
#> 1-8-1  -2593.50361 -1458.3997  3803.14125  -507.7832  3414.4306  2497.1110
#> 1-9-1   -948.26043  -426.6981  4257.61780 -1969.3376  1709.9256   911.4357
#>             0-6-0       0-7-0      0-8-0      0-9-0       1-1-1      1-10-1
#> 0-1-0   3121.7642  1615.54710  1254.7387  2583.5339 -3219.31059  -270.75156
#> 0-10-0 -2402.5464  -707.46994  -423.6127 -2769.0250 -3337.88325  -184.07124
#> 0-11-0  2580.8015  1760.43139  4850.3529  3478.1084  1838.59406  2676.68737
#> 0-12-0  2097.9152  3863.95023  4828.2937  1930.5473  1761.20691  -186.00522
#> 0-13-0  1812.0180  2690.19950  5076.7415   392.1879  2665.54630  -561.19058
#> 0-14-0  2091.1469  -826.57765  1250.6902  1477.9659  2153.08429  -247.45264
#> 0-15-0 -1284.8669  -218.08875 -3178.3402 -2523.0632 -5582.86107 -1603.40465
#> 0-16-0  2968.6048  8535.24014  9360.3580  2579.7375 -1393.64227 -5109.02432
#> 0-2-0   1506.0572  1344.86078   397.9831  5233.1502 -1593.88858 -5474.48940
#> 0-3-0   2526.4204  1258.39053   594.6381   640.3278   868.50702 -1004.81233
#> 0-4-0   5408.8689  3229.57076  7278.8509  4962.8528  5841.25661  5063.66778
#> 0-5-0    538.8061 -1418.52777 -6243.6786 -2128.5132 -1292.74285 -1810.63692
#> 0-6-0  11299.7485   171.30468  3898.3822  3157.2451  4810.70506  1870.64245
#> 0-7-0    171.3047 12422.09804  5160.9656   837.2205  3027.01822    51.05999
#> 0-8-0   3898.3822  5160.96564 17180.6756  5510.3065  7424.96367  2979.81493
#> 0-9-0   3157.2451   837.22050  5510.3065 12768.2162  2783.86718  -186.48232
#> 1-1-1   4810.7051  3027.01822  7424.9637  2783.8672 17014.35107  6132.39536
#> 1-10-1  1870.6424    51.05999  2979.8149  -186.4823  6132.39536 11830.00043
#> 1-11-1 -4703.9372  1275.51820  3564.0533  1928.3520  4767.75229  3262.60183
#> 1-12-1  1979.2247   989.22565   360.3441 -2394.1156  4366.81461 -3189.47113
#> 1-13-1  4622.5323 -1014.00082  2676.4405  1590.5314  5869.01654  2564.97242
#> 1-14-1  3666.3846  2518.15901  3630.8106  3292.9719  2575.04861  1555.72395
#> 1-15-1  3740.4685   706.67912   511.8197  2593.3815   252.24502  -774.17613
#> 1-16-1  3135.0220  5251.07117  4938.6604  1411.3729  5771.54276   488.94076
#> 1-2-1   2363.1182  -143.49713  4917.1835  1009.3765  6397.97719   215.42016
#> 1-3-1   -240.1695  2501.12160   486.2806  1310.5069  4824.34095  1238.15770
#> 1-4-1   1388.7339  5603.34045  3809.6893   215.4764    28.81452 -3773.75816
#> 1-5-1   3307.2208  4616.32582  1209.8754  3437.9055  1726.07254 -3289.89900
#> 1-6-1   1532.6956   -59.70173 -2380.2456 -1849.6945  3450.14832  -129.75286
#> 1-7-1   1584.5756  3349.17922  1448.1036  1708.2221  6597.99064   461.21856
#> 1-8-1   2255.3096 -1318.84686   654.7027  3012.9354  6073.05476  4026.80926
#> 1-9-1   1419.8114   806.94970 -1408.6910   123.8419  2398.11705   820.55268
#>            1-11-1      1-12-1     1-13-1     1-14-1     1-15-1     1-16-1
#> 0-1-0   -127.4433   -56.38783 -1708.8599  -380.3491  1139.1005  3187.1871
#> 0-10-0  -842.0961  3614.32655 -2110.6086 -3568.9520 -3059.5376   357.9054
#> 0-11-0  -741.1407 -2264.00116  2097.4087   362.5530  -501.6754  4293.5892
#> 0-12-0  1215.1884  5068.47070  -300.4453  3584.8592  2456.3206  3341.9480
#> 0-13-0  4601.6308  2203.72605 -4792.8594  1945.1260  -782.9398  8336.5573
#> 0-14-0   987.9671    43.10286  -936.6811   734.0773  4501.7954  1894.0722
#> 0-15-0 -1942.0033  1347.95649 -5378.9520 -3634.9900  3209.4143  -498.8238
#> 0-16-0 -1288.8695  4011.66709 -2194.6284  2959.0860  -529.3703  6362.4207
#> 0-2-0   -159.1940  7433.40771 -2038.1464   910.7326  1793.1335  2455.9131
#> 0-3-0   2085.6063  4534.88803 -3187.9236   -46.1521  2029.0107  3626.6893
#> 0-4-0   2915.9811  2066.36939  3537.2639  1946.7422   938.7905  2851.1323
#> 0-5-0   1223.1099  1902.90052  -435.4730  2549.2705   745.6137 -1227.9211
#> 0-6-0  -4703.9372  1979.22470  4622.5323  3666.3846  3740.4685  3135.0220
#> 0-7-0   1275.5182   989.22565 -1014.0008  2518.1590   706.6791  5251.0712
#> 0-8-0   3564.0533   360.34411  2676.4405  3630.8106   511.8197  4938.6604
#> 0-9-0   1928.3520 -2394.11564  1590.5314  3292.9719  2593.3815  1411.3729
#> 1-1-1   4767.7523  4366.81461  5869.0165  2575.0486   252.2450  5771.5428
#> 1-10-1  3262.6018 -3189.47113  2564.9724  1555.7239  -774.1761   488.9408
#> 1-11-1 21324.5825 -1406.46683 -3956.2584  1388.1225 -1882.0771  2119.0132
#> 1-12-1 -1406.4668 19989.90116  4901.6390  1126.7954  1247.9468  3678.4035
#> 1-13-1 -3956.2584  4901.63901 19021.1647  3505.7049  1526.1159  5379.5492
#> 1-14-1  1388.1225  1126.79545  3505.7049 11038.1630  4590.9089  1084.5244
#> 1-15-1 -1882.0771  1247.94682  1526.1159  4590.9089 12879.1014   931.4686
#> 1-16-1  2119.0132  3678.40349  5379.5492  1084.5244   931.4686 14912.6377
#> 1-2-1  -2155.9457  8865.07313  3317.4351  2411.8160  1158.1312  3118.3019
#> 1-3-1   2637.0981  -400.60737  -831.8849   780.5436  -952.1173  2097.7398
#> 1-4-1  -2899.9484  9956.07334   625.2176  1722.0673 -2994.5417  5098.6006
#> 1-5-1   -874.8045  6155.32720  3342.8590  1460.6540  3799.5503  5762.2941
#> 1-6-1    579.8460  3575.25740  1956.5380  3480.7473   536.3871   972.9806
#> 1-7-1  -3897.4300  6513.22106  4585.3416  1827.2818  2697.8110  3513.6324
#> 1-8-1    130.4308  3547.71769  2903.7679   670.8084 -1398.4506  3033.6084
#> 1-9-1  -1774.7667  2734.18831  2641.1125  2922.0493   304.1133   775.4768
#>             1-2-1      1-3-1       1-4-1      1-5-1       1-6-1      1-7-1
#> 0-1-0    130.2069 -2374.3932  1195.82507  5628.9899 -4310.74354  3007.6662
#> 0-10-0  2550.7696   635.4560  3993.39917   910.6896 -2193.37116  1353.0266
#> 0-11-0  4194.6899 -1386.9001 -2160.56275   985.5443 -4452.30712  1768.9611
#> 0-12-0  6784.7718  -189.0714  -443.91277  3315.6341  1670.48522  2955.0426
#> 0-13-0  9246.2576  1231.2330   239.15937  1371.8743  2127.08031  2454.2832
#> 0-14-0 -1601.8555 -2658.1111 -6491.99312 -2962.1543 -1286.39512  5182.5368
#> 0-15-0   645.2492  -492.5756   -76.35532  2148.7894   346.85952   224.5605
#> 0-16-0  5302.0502  1693.9343 12378.76206  4988.8932  -907.73918  -158.2329
#> 0-2-0    515.5671   243.9355 11768.24960  5924.2214 -1448.84802 -3871.1974
#> 0-3-0   3036.4915 -1218.8080  -302.27078  2972.1638  1323.39392 -2225.7318
#> 0-4-0   3483.5132  -421.6706 -3507.59616  2088.4960 -4935.38034   271.7479
#> 0-5-0   1176.2774 -1617.8647 -2974.33764  -559.7027   275.45115   367.0400
#> 0-6-0   2363.1182  -240.1695  1388.73388  3307.2208  1532.69564  1584.5756
#> 0-7-0   -143.4971  2501.1216  5603.34045  4616.3258   -59.70173  3349.1792
#> 0-8-0   4917.1835   486.2806  3809.68928  1209.8754 -2380.24556  1448.1036
#> 0-9-0   1009.3765  1310.5069   215.47641  3437.9055 -1849.69451  1708.2221
#> 1-1-1   6397.9772  4824.3410    28.81452  1726.0725  3450.14832  6597.9906
#> 1-10-1   215.4202  1238.1577 -3773.75816 -3289.8990  -129.75286   461.2186
#> 1-11-1 -2155.9457  2637.0981 -2899.94837  -874.8045   579.84597 -3897.4300
#> 1-12-1  8865.0731  -400.6074  9956.07334  6155.3272  3575.25740  6513.2211
#> 1-13-1  3317.4351  -831.8849   625.21765  3342.8590  1956.53805  4585.3416
#> 1-14-1  2411.8160   780.5436  1722.06733  1460.6540  3480.74730  1827.2818
#> 1-15-1  1158.1312  -952.1173 -2994.54171  3799.5503   536.38713  2697.8110
#> 1-16-1  3118.3019  2097.7398  5098.60064  5762.2941   972.98064  3513.6324
#> 1-2-1  16584.2341  1375.7833  2113.41882  2707.6073   780.22088 10045.8785
#> 1-3-1   1375.7833  8984.9582  2540.96604 -2367.3776  3529.25321  2455.7785
#> 1-4-1   2113.4188  2540.9660 24366.61443  5822.3503  1872.81374   386.4967
#> 1-5-1   2707.6073 -2367.3776  5822.35031 18386.7345 -1339.06189  3360.6403
#> 1-6-1    780.2209  3529.2532  1872.81374 -1339.0619 13480.76306  1270.2591
#> 1-7-1  10045.8785  2455.7785   386.49673  3360.6403  1270.25907 19654.0895
#> 1-8-1   4721.9009  1568.5831  1687.42280   625.4804  -390.32671  3833.7590
#> 1-9-1   1641.4452  1249.5239  2937.02237  4619.3524 -2146.45194  3847.5010
#>             1-8-1      1-9-1
#> 0-1-0   -447.8231  5486.4276
#> 0-10-0   875.4341  3578.9104
#> 0-11-0  1829.1992   371.4501
#> 0-12-0  -117.0150  3152.2805
#> 0-13-0  -436.9350  -581.0795
#> 0-14-0   495.1198  -505.7676
#> 0-15-0 -2593.5036  -948.2604
#> 0-16-0 -1458.3997  -426.6981
#> 0-2-0   3803.1412  4257.6178
#> 0-3-0   -507.7832 -1969.3376
#> 0-4-0   3414.4306  1709.9256
#> 0-5-0   2497.1110   911.4357
#> 0-6-0   2255.3096  1419.8114
#> 0-7-0  -1318.8469   806.9497
#> 0-8-0    654.7027 -1408.6910
#> 0-9-0   3012.9354   123.8419
#> 1-1-1   6073.0548  2398.1171
#> 1-10-1  4026.8093   820.5527
#> 1-11-1   130.4308 -1774.7667
#> 1-12-1  3547.7177  2734.1883
#> 1-13-1  2903.7679  2641.1125
#> 1-14-1   670.8084  2922.0493
#> 1-15-1 -1398.4506   304.1133
#> 1-16-1  3033.6084   775.4768
#> 1-2-1   4721.9009  1641.4452
#> 1-3-1   1568.5831  1249.5239
#> 1-4-1   1687.4228  2937.0224
#> 1-5-1    625.4804  4619.3524
#> 1-6-1   -390.3267 -2146.4519
#> 1-7-1   3833.7590  3847.5010
#> 1-8-1  11117.4124  3773.2122
#> 1-9-1   3773.2122 13960.5890
#> 
#> $stimcats
#>      stim cat
#> 1   0-1-0   0
#> 2  1-14-1   1
#> 3   0-2-0   0
#> 4   1-3-1   1
#> 5   1-4-1   1
#> 6  1-11-1   1
#> 7   0-8-0   0
#> 8   1-9-1   1
#> 9  1-16-1   1
#> 10 1-10-1   1
#> 11 0-14-0   0
#> 12  0-5-0   0
#> 13 1-15-1   1
#> 14  0-4-0   0
#> 15  1-5-1   1
#> 16  1-6-1   1
#> 17 1-13-1   1
#> 18 0-10-0   0
#> 19 0-11-0   0
#> 20 1-12-1   1
#> 21 0-16-0   0
#> 22  0-7-0   0
#> 23  1-2-1   1
#> 24  1-8-1   1
#> 25  0-6-0   0
#> 26 0-12-0   0
#> 27 0-13-0   0
#> 28  0-3-0   0
#> 29  0-9-0   0
#> 30  1-1-1   1
#> 31 0-15-0   0
#> 32  1-7-1   1
#>