Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

aat_stimulusscores(
  ds,
  subjvar,
  stimvar,
  pullvar,
  targetvar = NULL,
  rtvar,
  iters = 5
)

Arguments

ds
the data.frame to use

subjvar
Name of the subject-identifying variable

stimvar
Name of the stimulus-identifying variable

pullvar
Name of the movement-direction identifying variable

targetvar
Optional. Name of the stimulus-category identifying variable

rtvar
Name of the reaction-time identifying variable

iters
If there are missing values (which is almost inevitable) then multiple imputation will be used to complete the covariance matrix - this argument sets the number of multiple imputations to be used.

Value

Exports a list containing a data.frame with stimulus-specific bias scores, indicated in the column names, a covariance matrix of that same data, and a data.frame indicating to which stimulus category each stimulus belongs.

Examples

ds<-aat_simulate(biasfx_jitter=40,nstims=16)
ds$stim<-paste0(ds$stim,"-",ds$is_target)
aat_stimulusscores(ds,"subj","stim","is_pull","is_target","rt")

#> $data
#>          0-1-0     0-10-0      0-11-0      0-12-0     0-13-0      0-14-0
#> X1  180.213299   89.66595   57.915850  120.337981  200.45889   37.792246
#> 2   102.768111  147.68913 -116.722002  154.496865  270.84571   74.867830
#> 3    17.609990   24.88523    7.994467   78.787905  109.00202   66.232241
#> 4   -33.755979   25.25334   97.711114   83.247562   83.85070  111.250037
#> 5    99.013684   14.09619  -82.333596  155.568179  -69.14821  -86.466946
#> 6   164.047925   93.90856 -179.222038 -127.270176   58.66142   53.670544
#> 7    67.172244  -40.36862   85.892240   61.026319   22.57180  102.340461
#> 8  -179.050691  -82.96282  167.462637   98.173893  -22.02852  -66.148068
#> 9   119.778534  105.54749   22.036061   69.846037  -29.09410  -90.100957
#> 10   18.086566  283.24893  178.568957  180.447513  224.78869  271.995652
#> 11   69.518634   30.25334 -139.137465  -15.788497 -242.98105  111.126378
#> 12 -192.132518  124.19643    8.911673  -91.297615 -239.19558  245.608154
#> 13 -239.272531   75.91865  -65.845840  125.718986  155.27294  342.416003
#> 14    1.316199   18.45058  -27.246503 -168.429791 -149.75915  -33.308034
#> 15  -45.964168 -133.47584  214.729654  440.425917 -226.48320 -155.617872
#> 16  314.483712 -129.98404   88.997945  -24.465574  162.37866   10.304382
#> 17  -77.844241  -56.37260  -91.332975   78.552898 -130.45435  -30.044944
#> 18    7.908308  -66.34645    6.615855   14.521084   53.32121    3.814929
#> 19  139.440903   74.98482   -5.061690   75.081971   54.63356  134.454006
#> 20  -60.756673  -68.64771   76.560276  111.832205   33.48243  -56.422103
#> 21  102.284495  139.30132   -3.946575  281.880749   59.31732  126.602896
#> 22    9.654527  -48.61890   76.172775   99.785096  308.69464  -75.723083
#> 23   -6.810995  -38.01242   62.409207  -84.836292  180.25158  109.852335
#> 24 -381.401825   92.96894   20.029951  118.523702  -38.88733   -6.681229
#> 25   87.901425  366.57421   -2.900496   42.306794  -39.66805  -86.477799
#> 26  110.241054   47.80341   40.073862   57.285593  -10.03557  -16.031388
#> 27  -52.617161  -45.98067   41.633800  -38.825760  140.23230   62.473357
#> 28   91.149301  -29.55848  -45.010797    3.448958  179.37582  108.272072
#> 29  -20.240583   29.86499  181.359408   60.806550   50.63188   76.157064
#> 30  -90.405801  -25.95231   31.454378   17.891419  197.55976 -105.658368
#> 31 -245.341694   77.22768   19.107189   98.901028  -12.42772   66.099487
#> 32  215.437763  220.15560   46.159718  110.849465   75.32300  143.740610
#> 33  -10.372487   79.00517  -24.540719  -38.741822  -36.22202   14.265485
#> 34   84.706969   45.31853   80.686053   80.839156   50.73550  -62.662128
#> 35  175.005128  193.43493   67.721551   99.171003   56.75904  -19.738702
#> 36   85.558130 -137.53071  -11.161158   14.870781  202.86135  150.310984
#>         0-15-0      0-16-0       0-2-0       0-3-0      0-4-0         0-5-0
#> X1   41.350137   45.561493   22.444788  136.170287   72.49180   86.45315393
#> 2  -101.390410   76.877623   82.135950   -9.859288   98.68674    5.67108752
#> 3   145.135812  -75.453994   66.329338   89.601743   45.98233   33.07733082
#> 4   -29.974770 -137.882612  -69.635671 -153.059008 -137.46642  -92.02453895
#> 5    60.181457 -130.671949 -462.416569  -33.507784  229.24441  104.47303488
#> 6   -29.198457    9.410344  186.421755   -8.508695  151.03634   68.99408194
#> 7    -5.086884  127.256203  151.086796   14.158463   77.38428  -23.83828226
#> 8    82.728969  -83.502720  255.042819 -199.994319 -229.60367   62.78347556
#> 9   106.615205  136.652104   42.229638  -24.291508  -28.84610 -161.71097021
#> 10  162.123022  190.718332  163.745772  123.327991  103.11843  182.80388819
#> 11   25.006412   81.009143   56.968454  -24.664335   73.62064  102.60884810
#> 12  191.524411  130.568175  361.545253  108.141114   50.94541  111.03065733
#> 13  397.248428 -105.640023  140.709303 -147.344912 -125.18513   92.99315323
#> 14   16.666280   50.883202  -51.440325  190.630497 -222.18066  -20.45267625
#> 15  -19.977561  398.728794  291.000967  394.544921 -375.11374   53.38195809
#> 16  193.307928  197.890672  249.056149  141.450200  152.81741   -5.25575386
#> 17    7.250742  -49.959350  164.546453   84.003497 -146.19345   21.91985346
#> 18   85.829543   48.071748   91.417132   20.091649  -72.70299   39.18626116
#> 19   58.963809   43.197691  201.341439  299.720046  154.09509    8.94751452
#> 20  130.582422   86.545591   56.662159   93.034211  -30.00181   46.34875466
#> 21  123.044142  -99.159169 -102.505709  -43.138208   42.82195  183.69826754
#> 22 -226.887041 -191.035323 -533.644136  144.429837 -140.22393   50.63131709
#> 23   37.987899   -4.078363  146.539252  -29.720339   90.80221    0.06551501
#> 24 -138.145715  176.128786    4.835806  107.240126  120.00369   86.52748974
#> 25  -77.808500   97.305693   39.159560   14.088893  273.29696 -106.18613444
#> 26  -57.649638   75.773632  108.807869   89.651177   60.31505  114.03189565
#> 27 -183.677801  -44.466818   23.017756   75.706072  -71.56183   80.26153182
#> 28  108.000403   61.214456  -91.292964 -118.775143 -123.26387  -45.29887901
#> 29   87.107315   62.982750   90.876529  -23.712525   39.95617   71.03389684
#> 30  -24.123627   -6.268141   67.135171  -47.920040   27.16004  -15.80225394
#> 31  -18.407265  225.782042  -35.411114   60.527349 -102.98228  163.19452825
#> 32   99.582253   -6.848936    5.918223   88.775599   55.36457   96.52938173
#> 33    1.575581   52.124188  393.958282  141.442270  145.77567  222.04056797
#> 34   62.477800  131.404634  -13.634018 -100.371702   83.93113    3.50164558
#> 35   89.676881  128.716521  -24.607337   83.637732   60.46930   50.04407296
#> 36 -137.596253  -42.243313   50.357547 -124.167260  -34.49975 -227.20354675
#>            0-6-0       0-7-0        0-8-0       0-9-0       1-1-1     1-10-1
#> X1  106.90886690   21.985375  167.0817110   78.990378   61.122549  139.64012
#> 2    19.45354511  135.068427  -94.1089084  -80.370017  -45.844730   76.78392
#> 3    14.81904614  161.078937   -0.7695019  170.262504  247.975895   78.92000
#> 4     0.01743218    6.708068   96.0849908  -21.128796   64.040182  171.00928
#> 5   100.78999728   -8.264095  -50.6836704  -19.702397  -41.700814 -230.13872
#> 6    52.42792213   -9.223977  167.9345172  158.062323  -36.098188   66.92070
#> 7   -42.59493339  -70.170189  102.5175932  249.596137   29.466218  129.04332
#> 8   -11.54274205 -155.295646   34.6767887  -88.070804  289.040701  111.53826
#> 9    34.01164988   16.014862  -76.1737361   16.372827  150.330065  101.03408
#> 10   79.16495153   57.431297   14.3505393   80.701079  103.720024  309.80743
#> 11 -137.31878190  -72.845995   60.9002312  -32.792241 -129.187679  102.75541
#> 12  -70.74257425  314.824445  288.4465200  178.716039  -33.269565  220.42297
#> 13   26.89356462   44.130197   -9.3207641  122.482878   -7.475636   38.01102
#> 14   14.92343568   89.174157  186.2218820   57.876100  -51.549035 -132.67981
#> 15  489.80619374 -326.103572   89.8975209  493.430855  -31.639985  630.72784
#> 16   40.01584500  141.454397  159.3952650  222.840503   18.324983  -57.94716
#> 17  -41.02125092  154.009911  130.8739194 -171.152408  -25.359771  160.78986
#> 18  -22.62547507   53.620217  -11.1834158   88.943559   36.971741   80.03631
#> 19   17.95447962  114.096944   57.4669157  147.227988  -88.777108  209.70930
#> 20   64.69332319   17.432703  -71.1778496   84.072770  210.954002   93.50853
#> 21  189.53333804   49.588690  143.3771608   43.502168   59.469278  -87.45427
#> 22  340.61838753  176.261611  239.7566715  257.464551  -49.021143  -69.42234
#> 23   62.60942509   12.099113  114.0157029   40.974648  117.855904  -51.53624
#> 24    1.19715205   78.271360   68.3194414 -134.255323  130.225135  243.80352
#> 25  101.19589187 -209.039050  -18.3452964   40.628295  -94.534842  142.99486
#> 26   35.70925697  -25.227909   98.5631397    2.592104   22.759873  105.91516
#> 27 -165.98806724  110.536278  163.9809876   46.249605 -137.706102  162.46592
#> 28  141.87518750  198.460905   51.1420190   -1.851429  296.964298   92.59570
#> 29  -72.90311002    2.468744  108.9049009  -31.715911   49.179390   80.46276
#> 30  132.49719053   22.198995   -2.5280438  128.747312   51.254276  -26.34021
#> 31    3.71012454   24.768755  145.6059473   78.241841  178.964185  169.09955
#> 32  -38.17129132  -14.258739  150.9862915  144.099106  101.697662  131.87299
#> 33  174.50049711  106.750032   54.1675581  111.561773   66.644482   15.36209
#> 34  203.54509772  -56.423816  181.3537024   24.032734  125.213626   56.49673
#> 35   20.29706101   36.287989  214.8990182   77.443358  -73.317578  -23.40703
#> 36  188.49511347   74.812957 -303.7482026   97.445623   12.635341 -136.92673
#>        1-11-1      1-12-1      1-13-1       1-14-1      1-15-1     1-16-1
#> X1   56.94217   53.838265  156.337743  130.6543840    5.183526   65.29966
#> 2    71.28464   77.467219  -41.952750   91.8180957  185.392245 -143.94385
#> 3   -59.23308   71.484836  189.704423  269.1681848   67.354210  170.47771
#> 4   -54.16923  110.527010  207.275898  126.9886123   70.313014   97.06211
#> 5    68.07931 -268.762018  260.532721  -77.1228207  175.593685 -172.61968
#> 6    57.67357  226.805844   66.816151  250.5806950  132.198568  128.06236
#> 7   104.20317   97.300188   31.237309   63.1584465   83.930505   94.41524
#> 8   -91.68652  111.678355   98.433313   77.4571817 -107.899540   53.92514
#> 9   -52.03421   62.480319  142.473141 -116.6277013   28.484864  -50.94122
#> 10  185.00846  -23.863991  150.011304   47.5415053   78.098714  193.40731
#> 11  282.14297  135.254782   31.852683   47.9256360   46.321826   53.63323
#> 12   33.81754 -122.337846 -382.678668   67.6481387  127.868619  143.51745
#> 13  202.82478  167.391786   -1.693042  -36.6355105    6.121983  -66.04513
#> 14   22.10334   64.305827 -223.642141   27.9958421  159.768153  -24.18804
#> 15  139.87506  383.118739 -100.815810  193.2581908  166.450418 -198.90511
#> 16   45.82002   74.699357   66.341784  279.0407221  212.783754  -23.43607
#> 17  213.47702   28.080798  167.092220  102.0567426  110.864606  147.88986
#> 18  131.25750   74.033937   78.172082   50.1110377   23.053014   13.27948
#> 19  -32.59199  166.488256  -26.014829   89.7151842  -66.763644  170.97600
#> 20  133.65175  145.851906  -81.211792  142.6066820    4.312980  -84.08440
#> 21   97.41928  246.676291  156.008893  150.4141251  138.533866  165.25970
#> 22  101.50017  130.664594 -183.273273  143.7776611  456.977633  355.21903
#> 23   39.63457    2.895265  -51.943664  -22.0846580   60.854799   24.65341
#> 24   76.27464  188.420032   92.798741    3.0091241   -2.114528   15.16402
#> 25   94.26389   56.445160   65.643101  153.7818407  205.690824  213.51031
#> 26   29.32907   32.107598    4.557055   -0.7538427  167.049601  -50.52732
#> 27  -18.43176 -163.291182   15.909181  104.8420451   60.119753   44.75868
#> 28  121.18225 -127.312134  -46.003522   88.2230157  233.260471  130.33506
#> 29  113.33039  125.420772   65.042091  -28.9795839   19.101393   17.27670
#> 30   66.91206  -35.623447   86.215478  122.0584949  172.898461  154.26494
#> 31   73.76894  -27.878804   15.208400   59.8877070  131.510856  149.03189
#> 32  169.60141   92.137312  289.175197   86.4942987  -53.584299   84.03841
#> 33  164.18899  -38.113415   30.502920   84.2738205 -116.845968  -35.93564
#> 34 -224.68242  -44.848673   79.405676  -59.4835855 -307.230384 -172.29892
#> 35 -156.03176   22.095487  -38.094410   40.6597102  107.584298   51.73773
#> 36  -92.56967   98.212032  -28.303602  143.0365366  -35.218931 -116.53493
#>          1-2-1         1-3-1       1-4-1       1-5-1       1-6-1       1-7-1
#> X1  -20.298502  4.851039e+01   64.440356  132.362105   33.697350 -102.403069
#> 2   117.922130 -1.155054e+02   -5.994714   13.017805  100.106100  363.152113
#> 3    68.740723  6.569649e+01   32.110735   90.946060  183.506983   50.395964
#> 4   133.809620  2.199216e+00  149.646033  -37.475561  114.288412  120.255477
#> 5   -82.434896  1.589141e+02 -230.406090 -115.491603   47.357534  -66.820181
#> 6   104.002546 -7.515476e-01  210.765928  104.803622   53.414372  178.429982
#> 7    57.948777  2.112810e+02  112.516954   67.159528  126.766508   67.689066
#> 8  -128.879378  2.086285e+02  174.571835  246.859828  139.748985  127.813649
#> 9  -119.627709 -7.155777e+01   29.286344  -42.782167   92.536090  108.802336
#> 10   75.958664  1.219960e+02  220.109696   58.075164  128.653369  178.372983
#> 11  192.613295  1.149990e+02  135.398177    4.698531  -34.832337   20.587072
#> 12  212.263284  3.419692e+02  -88.074796   29.584507  226.876850  303.338709
#> 13   99.619912 -4.267245e+01  289.454466 -154.828149   -3.469449  217.766128
#> 14 -380.038118  5.881332e+00  -29.712986   15.555900  -69.718754   81.815932
#> 15  325.452750  3.893555e+02 -211.499641   48.502639 -205.844804  141.869538
#> 16   -0.733923  2.023008e+02  321.894264  149.911272  -25.765824  204.431045
#> 17  102.348896 -7.090770e+00  202.640114   98.969477  -21.148410   71.534740
#> 18   91.151101  1.366813e+02   60.849466  130.715372  134.832874  227.333146
#> 19  240.072036  3.624261e+01  -25.029591   47.876554  144.650299  236.014827
#> 20  147.943290  8.184265e+01  116.167365   -4.091683  -46.420857  228.823741
#> 21  245.908957  1.282005e+02    6.421705  128.608473  223.718380   93.015931
#> 22  181.296507  1.122160e+02 -349.656220  285.932355 -133.690920 -494.977261
#> 23  106.952415  2.642865e+01  -90.163417  -26.942013 -179.161718 -134.995919
#> 24   91.393369  1.777853e+02  133.779257   33.732025   53.487288   93.380538
#> 25  158.662330 -1.155871e+00   63.551378   84.958358  -59.127920   93.567329
#> 26   63.502658  1.384081e+02   29.516030  230.046316  -16.597518   70.417125
#> 27   43.609878 -2.939326e+01  185.845524  -13.217728  185.141559  194.234504
#> 28   12.447493 -7.494220e+01  111.558500  101.088650   45.880931  -18.532673
#> 29   33.147224  8.654933e+01  139.407370   -1.324370   47.761790   63.379899
#> 30  123.653570  1.251539e+02   32.140022   94.008955   -2.098190    6.978547
#> 31   -4.499480  6.154257e+01  -71.152556  120.377447  199.814821   80.156821
#> 32  155.363918  1.548118e+02   98.942643   95.932386   -4.817466   11.479223
#> 33 -192.676791 -2.193413e+01  135.046590  175.996176   51.212175  233.969083
#> 34   20.249641  3.205494e+01   33.173142 -232.582662 -124.582647  -55.149651
#> 35   85.359670  1.536631e+02  -81.769218   19.708962   25.003619   55.930399
#> 36  127.963723 -7.716368e-03  141.781233  166.782869   12.952778  -59.813340
#>          1-8-1       1-9-1
#> X1   47.273344   73.039699
#> 2   -24.752724  188.552158
#> 3   -32.250318 -160.166212
#> 4    28.939281   31.335587
#> 5    94.132104  -55.224537
#> 6   153.730255  182.905831
#> 7    36.825654  139.900388
#> 8   113.279594  -75.886359
#> 9   182.998429  113.578558
#> 10  134.539826   91.250899
#> 11  110.383606   19.932712
#> 12  223.738677  266.947454
#> 13  201.953246  235.563936
#> 14   -2.505575   50.747013
#> 15  373.017692  330.773541
#> 16   89.320990  114.803429
#> 17  -66.313750 -221.597911
#> 18  109.988811 -109.903162
#> 19  -43.304557  120.828007
#> 20  -34.643473 -110.844488
#> 21  100.840943  187.843117
#> 22    9.273674  226.338930
#> 23    7.474542 -128.800820
#> 24   -7.509426  113.305363
#> 25 -124.423058   -2.604557
#> 26   42.459000  100.028044
#> 27 -162.660788  -63.613087
#> 28   61.875143   49.163129
#> 29  194.111221   43.802265
#> 30   64.456093   38.717944
#> 31  183.399195  169.754154
#> 32   18.430212  -43.649466
#> 33  -79.806052  257.944432
#> 34 -185.240812  148.381617
#> 35   30.667850  -45.099774
#> 36 -300.349394  167.146925
#> 
#> $covmat
#>              0-1-0     0-10-0      0-11-0     0-12-0     0-13-0       0-14-0
#> 0-1-0  19949.48307  1901.0660 -1283.61970 -1027.5152  4892.7555 -1435.977556
#> 0-10-0  1901.06603 12715.7041 -1437.00387   846.9530  -111.4431  3301.378130
#> 0-11-0 -1283.61970 -1437.0039  7657.02737  3806.6275  1224.9371 -1512.722022
#> 0-12-0 -1027.51523   846.9530  3806.62749 12092.2278   137.7964 -1672.718825
#> 0-13-0  4892.75555  -111.4431  1224.93709   137.7964 18875.6125  3510.806739
#> 0-14-0 -1435.97756  3301.3781 -1512.72202 -1672.7188  3510.8067 12278.129589
#> 0-15-0   486.81991  1725.7727   560.79006   838.6992 -1446.4856  5739.547150
#> 0-16-0  -145.15194  1437.9704  3224.83208  2571.7440 -5804.1328 -2005.278985
#> 0-2-0  -2894.13119  -363.9594  1944.79649 -2971.9999 -6939.9119  4763.059051
#> 0-3-0   1768.44756   393.1399  2179.70832  2604.1512 -4361.1403 -2782.611306
#> 0-4-0   7594.73579  8048.7850 -3551.82442 -3504.2726  2658.3291  2012.957357
#> 0-5-0  -3053.96716  2475.1149   182.99048  2033.2909 -2109.3474  2094.341808
#> 0-6-0   2001.82974 -2311.6768  2749.05322  6814.0411  2509.7668 -5092.317197
#> 0-7-0   -789.29655  -419.8427 -3414.50572 -5477.1554  4741.1768  5223.018979
#> 0-8-0     37.86623  1441.9429  1573.78538 -2022.2419 -3074.9124     6.328016
#> 0-9-0   3342.25818 -2632.2768  3366.88409  2685.0690   409.8336  -155.981413
#> 1-1-1  -3466.98543 -1693.1168  2988.71677  1371.1915  2809.2384  -281.830483
#> 1-10-1 -6071.23369  1817.8168  5019.16119  7201.5634 -6449.6056   536.710707
#> 1-11-1 -2232.33758   875.6502 -2115.95383  1431.6536 -2333.5337  2684.860730
#> 1-12-1  -312.54976 -1165.9992  1647.33298  5304.0506 -1435.8471  -441.209191
#> 1-13-1  4436.57183  2753.1100   -43.17761  4295.2402  3032.2518 -1025.671556
#> 1-14-1  3047.73614 -1754.1427  -106.89236   290.6091  2271.9524  -479.953627
#> 1-15-1  1804.44488  -409.5782 -1424.22896   392.4473  2446.8261 -2132.376739
#> 1-16-1  -934.02061  4189.8071  -133.21456 -2352.8450  3408.2369  2825.224338
#> 1-2-1    -44.98375   954.9063   638.44279  6286.9855  1185.2931  3253.015059
#> 1-3-1  -2184.66138 -2066.4537  4741.94979  4018.8968 -6323.3520 -1666.525797
#> 1-4-1    -88.37071   174.3472 -1111.33057 -4053.3148  2317.3704  5992.091307
#> 1-5-1   1552.40352 -1744.3688   914.76000  -459.3170  2214.1654 -1752.329654
#> 1-6-1  -2397.14689  2786.4089 -1124.41720  -522.8882   398.1765  5027.802396
#> 1-7-1  -2832.33128  3090.5655 -1619.66016  -412.6884 -5685.6706  4395.447074
#> 1-8-1  -3339.88965   245.3231  2226.47587  4362.1778 -5525.4883  1041.059722
#> 1-9-1  -1084.48887  1414.8153   259.49220  2932.8460  -303.5341  2733.331100
#>            0-15-0     0-16-0      0-2-0      0-3-0      0-4-0       0-5-0
#> 0-1-0    486.8199  -145.1519 -2894.1312  1768.4476  7594.7358 -3053.96716
#> 0-10-0  1725.7727  1437.9704  -363.9594   393.1399  8048.7850  2475.11489
#> 0-11-0   560.7901  3224.8321  1944.7965  2179.7083 -3551.8244   182.99048
#> 0-12-0   838.6992  2571.7440 -2971.9999  2604.1512 -3504.2726  2033.29086
#> 0-13-0 -1446.4856 -5804.1328 -6939.9119 -4361.1403  2658.3291 -2109.34738
#> 0-14-0  5739.5471 -2005.2790  4763.0591 -2782.6113  2012.9574  2094.34181
#> 0-15-0 14025.7144   817.7744  6633.0039 -2046.3681   355.5305  2314.00791
#> 0-16-0   817.7744 14088.0181  9970.9044  7244.3416   152.5498   716.33041
#> 0-2-0   6633.0039  9970.9044 32845.1368  4179.1436   259.8904  1582.37049
#> 0-3-0  -2046.3681  7244.3416  4179.1436 15257.4053  -611.0298  3236.13867
#> 0-4-0    355.5305   152.5498   259.8904  -611.0298 18802.8868  1438.46831
#> 0-5-0   2314.0079   716.3304  1582.3705  3236.1387  1438.4683  8547.49675
#> 0-6-0  -2529.5563  2064.0040 -4337.6799  3778.3216 -3911.6343  -816.82893
#> 0-7-0   1488.3532 -4064.5301 -1791.0999   410.5168  1867.0191   810.91906
#> 0-8-0    338.5051  1266.9104  -324.2571  4613.1209 -1136.3738  4678.36474
#> 0-9-0   1239.1706  4488.6515  3595.2521  7956.3057 -2930.5131   331.57150
#> 1-1-1   3851.9353   316.4889  1049.6929 -4384.0655 -2832.3020   -29.65078
#> 1-10-1  -407.1358 10535.0357 12583.5429  8131.9121 -5982.2049  2140.70980
#> 1-11-1  1647.8663   855.6389  1576.4787  2505.0901  -322.3449  4085.07526
#> 1-12-1  -150.4652  2408.6248  4879.4627  3444.1853 -4689.4766  -686.31553
#> 1-13-1  1434.5880 -3888.3481 -3896.9486 -4854.4890  4982.5178   572.95678
#> 1-14-1 -1612.6670  -206.7155  3019.7599  3125.0102  -583.7492  -365.30360
#> 1-15-1 -3915.9724 -1111.4645 -9588.9255  3261.4397 -1935.1652  -472.44144
#> 1-16-1 -1550.0211 -4198.3394 -3522.5488   948.8326  1312.7057  1436.21620
#> 1-2-1  -1005.6008  1044.0561  1240.2506  2317.8913   963.0435   188.37364
#> 1-3-1   1580.2578  4804.7619  3291.5599  5098.3301 -1610.4352  3247.58788
#> 1-4-1   5128.8579  -135.1141 12754.9307 -5954.2124  2217.0950  -622.23411
#> 1-5-1  -4199.1564 -1135.7409  1403.1608  2684.3065 -1746.5913  1087.46562
#> 1-6-1   2077.5914 -1899.5351  2433.2109 -2334.4400  2118.0681  2308.05541
#> 1-7-1   6228.3561  6365.2409 18675.0984  1939.3629  1683.0882  2168.84556
#> 1-8-1   6722.8452  5127.8025  3842.8742  2463.2972 -4957.5745  4327.38012
#> 1-9-1   -884.5153  5205.4503  3467.3336  3245.7470  -769.4618  1662.96697
#>             0-6-0      0-7-0        0-8-0      0-9-0       1-1-1     1-10-1
#> 0-1-0   2001.8297  -789.2965    37.866234  3342.2582 -3466.98543 -6071.2337
#> 0-10-0 -2311.6768  -419.8427  1441.942918 -2632.2768 -1693.11682  1817.8168
#> 0-11-0  2749.0532 -3414.5057  1573.785382  3366.8841  2988.71677  5019.1612
#> 0-12-0  6814.0411 -5477.1554 -2022.241940  2685.0690  1371.19150  7201.5634
#> 0-13-0  2509.7668  4741.1768 -3074.912362   409.8336  2809.23841 -6449.6056
#> 0-14-0 -5092.3172  5223.0190     6.328016  -155.9814  -281.83048   536.7107
#> 0-15-0 -2529.5563  1488.3532   338.505057  1239.1706  3851.93531  -407.1358
#> 0-16-0  2064.0040 -4064.5301  1266.910396  4488.6515   316.48887 10535.0357
#> 0-2-0  -4337.6799 -1791.0999  -324.257113  3595.2521  1049.69288 12583.5429
#> 0-3-0   3778.3216   410.5168  4613.120902  7956.3057 -4384.06552  8131.9121
#> 0-4-0  -3911.6343  1867.0191 -1136.373819 -2930.5131 -2832.30202 -5982.2049
#> 0-5-0   -816.8289   810.9191  4678.364743   331.5715   -29.65078  2140.7098
#> 0-6-0  15350.7027 -4042.0092 -1277.226067  7825.9431   993.45330  1559.2537
#> 0-7-0  -4042.0092 13474.6241  2065.006864 -2360.7642   335.12549 -5735.8365
#> 0-8-0  -1277.2261  2065.0069 12719.589338  2511.0582 -2242.06871  2283.5779
#> 0-9-0   7825.9431 -2360.7642  2511.058168 14933.9447 -2230.67044  4856.2946
#> 1-1-1    993.4533   335.1255 -2242.068712 -2230.6704 12161.76548   771.5873
#> 1-10-1  1559.2537 -5735.8365  2283.577887  4856.2946   771.58729 21916.0234
#> 1-11-1  -577.3771  -234.2847  -286.857002   708.8807 -1792.31827  3199.9606
#> 1-12-1  4336.0051 -5702.9283  -575.320641  4721.8618  -683.46357  6298.5573
#> 1-13-1 -2367.2970 -4216.9555 -3679.713626 -5324.3849  3738.73158 -1118.4659
#> 1-14-1  2507.1494   663.7933   686.004405  4952.8202  -388.36217  1853.3468
#> 1-15-1  3464.1054  2805.8176  2522.636032  3650.6451 -3988.26365 -2099.5642
#> 1-16-1 -2113.6862  4377.0157  5301.873315   668.4276   151.43773   686.4922
#> 1-2-1   3201.5613 -1662.8092    36.505955  5012.4264 -3616.69088  8066.8001
#> 1-3-1   2332.2475 -3345.3638  4051.020104  6344.6577  -475.92621  5558.4487
#> 1-4-1  -8399.9423   797.9828 -3152.290626 -5301.6855  2556.14747  1265.4171
#> 1-5-1   2261.3446  1206.4078   594.647764  2131.7243  1365.52139  -160.9667
#> 1-6-1  -6814.3015  4943.5952  -413.269270 -2736.0891  1903.57776   594.1447
#> 1-7-1  -7228.1254   891.2125 -2713.122084 -1057.7824 -1139.96399  7860.9610
#> 1-8-1   1122.5211 -3126.6401  3791.328283  4907.3887  1048.55035  7274.6214
#> 1-9-1   7824.8995  -181.5440  1580.421967  7594.1755 -2990.90600  3968.1112
#>            1-11-1      1-12-1      1-13-1     1-14-1     1-15-1      1-16-1
#> 0-1-0  -2232.3376  -312.54976  4436.57183  3047.7361  1804.4449  -934.02061
#> 0-10-0   875.6502 -1165.99916  2753.11004 -1754.1427  -409.5782  4189.80710
#> 0-11-0 -2115.9538  1647.33298   -43.17761  -106.8924 -1424.2290  -133.21456
#> 0-12-0  1431.6536  5304.05058  4295.24017   290.6091   392.4473 -2352.84495
#> 0-13-0 -2333.5337 -1435.84706  3032.25182  2271.9524  2446.8261  3408.23693
#> 0-14-0  2684.8607  -441.20919 -1025.67156  -479.9536 -2132.3767  2825.22434
#> 0-15-0  1647.8663  -150.46517  1434.58798 -1612.6670 -3915.9724 -1550.02105
#> 0-16-0   855.6389  2408.62477 -3888.34815  -206.7155 -1111.4645 -4198.33944
#> 0-2-0   1576.4787  4879.46271 -3896.94861  3019.7599 -9588.9255 -3522.54883
#> 0-3-0   2505.0901  3444.18528 -4854.48899  3125.0102  3261.4397   948.83256
#> 0-4-0   -322.3449 -4689.47656  4982.51784  -583.7492 -1935.1652  1312.70571
#> 0-5-0   4085.0753  -686.31553   572.95678  -365.3036  -472.4414  1436.21620
#> 0-6-0   -577.3771  4336.00508 -2367.29697  2507.1494  3464.1054 -2113.68619
#> 0-7-0   -234.2847 -5702.92828 -4216.95551   663.7933  2805.8176  4377.01574
#> 0-8-0   -286.8570  -575.32064 -3679.71363   686.0044  2522.6360  5301.87332
#> 0-9-0    708.8807  4721.86176 -5324.38495  4952.8202  3650.6451   668.42764
#> 1-1-1  -1792.3183  -683.46357  3738.73158  -388.3622 -3988.2637   151.43773
#> 1-10-1  3199.9606  6298.55735 -1118.46587  1853.3468 -2099.5642   686.49224
#> 1-11-1 11267.8194  2143.97460   502.20495   609.8217  4125.8006  2224.37756
#> 1-12-1  2143.9746 14744.02960  -473.33964  3866.6700  -274.0152   -87.95579
#> 1-13-1   502.2049  -473.33964 17579.60105  -245.8933 -4948.5007   -40.90085
#> 1-14-1   609.8217  3866.66999  -245.89329  8197.7624  3699.0485  3870.77359
#> 1-15-1  4125.8006  -274.01517 -4948.50072  3699.0485 16388.8198  6283.86188
#> 1-16-1  2224.3776   -87.95579   -40.90085  3870.7736  6283.8619 15415.04662
#> 1-2-1   2873.7201  6601.94498 -1225.86163  3978.2577  2867.7952  3847.54994
#> 1-3-1    864.1022  2610.33832 -2411.14151  1363.6366  1492.7136   -31.03200
#> 1-4-1   2370.1410  1405.04735  6552.02912  1938.0903 -7218.7752  -716.94290
#> 1-5-1   1709.8851  2443.76644  -998.80545  4958.0360  4981.6339  6097.77923
#> 1-6-1  -1355.6813 -2897.76042  2602.13438   527.5735 -1505.5998  4100.25228
#> 1-7-1   1465.2462  1617.48719 -1248.54688  1181.0481 -5217.5171 -5100.63645
#> 1-8-1   4923.6273  4137.48337 -1489.59998 -1014.4111  4491.4200   177.17510
#> 1-9-1    673.2827  4683.51524 -7165.82126   451.7184  1835.5243 -2048.92321
#>              1-2-1      1-3-1       1-4-1      1-5-1      1-6-1       1-7-1
#> 0-1-0    -44.98375 -2184.6614   -88.37071  1552.4035 -2397.1469 -2832.33128
#> 0-10-0   954.90634 -2066.4537   174.34723 -1744.3688  2786.4089  3090.56552
#> 0-11-0   638.44279  4741.9498 -1111.33057   914.7600 -1124.4172 -1619.66016
#> 0-12-0  6286.98553  4018.8968 -4053.31482  -459.3170  -522.8882  -412.68839
#> 0-13-0  1185.29314 -6323.3520  2317.37040  2214.1654   398.1765 -5685.67057
#> 0-14-0  3253.01506 -1666.5258  5992.09131 -1752.3297  5027.8024  4395.44707
#> 0-15-0 -1005.60078  1580.2578  5128.85791 -4199.1564  2077.5914  6228.35614
#> 0-16-0  1044.05606  4804.7619  -135.11409 -1135.7409 -1899.5351  6365.24086
#> 0-2-0   1240.25058  3291.5599 12754.93066  1403.1608  2433.2109 18675.09837
#> 0-3-0   2317.89131  5098.3301 -5954.21244  2684.3065 -2334.4400  1939.36288
#> 0-4-0    963.04353 -1610.4352  2217.09496 -1746.5913  2118.0681  1683.08821
#> 0-5-0    188.37364  3247.5879  -622.23411  1087.4656  2308.0554  2168.84556
#> 0-6-0   3201.56127  2332.2475 -8399.94227  2261.3446 -6814.3015 -7228.12536
#> 0-7-0  -1662.80918 -3345.3638   797.98276  1206.4078  4943.5952   891.21253
#> 0-8-0     36.50595  4051.0201 -3152.29063   594.6478  -413.2693 -2713.12208
#> 0-9-0   5012.42643  6344.6577 -5301.68546  2131.7243 -2736.0891 -1057.78236
#> 1-1-1  -3616.69088  -475.9262  2556.14747  1365.5214  1903.5778 -1139.96399
#> 1-10-1  8066.80009  5558.4487  1265.41713  -160.9667   594.1447  7860.96104
#> 1-11-1  2873.72006   864.1022  2370.14096  1709.8851 -1355.6813  1465.24622
#> 1-12-1  6601.94498  2610.3383  1405.04735  2443.7664 -2897.7604  1617.48719
#> 1-13-1 -1225.86163 -2411.1415  6552.02912  -998.8054  2602.1344 -1248.54688
#> 1-14-1  3978.25775  1363.6366  1938.09025  4958.0360   527.5735  1181.04807
#> 1-15-1  2867.79518  1492.7136 -7218.77522  4981.6339 -1505.5998 -5217.51709
#> 1-16-1  3847.54994   -31.0320  -716.94290  6097.7792  4100.2523 -5100.63645
#> 1-2-1  17406.60408  4582.5255 -2715.27617   308.8204  -765.0154    90.28281
#> 1-3-1   4582.52546 12132.3588 -4483.12190  2311.8618  -409.2239  -108.89455
#> 1-4-1  -2715.27617 -4483.1219 20021.84530  -507.4577  3434.9421  9974.27867
#> 1-5-1    308.82037  2311.8618  -507.45772 11139.0736  1370.7814 -2980.84763
#> 1-6-1   -765.01544  -409.2239  3434.94212  1370.7814 11779.4700  7558.29959
#> 1-7-1     90.28281  -108.8945  9974.27867 -2980.8476  7558.2996 22638.88305
#> 1-8-1   1972.79975  6824.6698 -2755.60395  -276.6296  1176.0216  3687.10674
#> 1-9-1   2539.60396  2246.8100 -3143.49349   718.4451  -208.3853  2013.45302
#>             1-8-1      1-9-1
#> 0-1-0  -3339.8897 -1084.4889
#> 0-10-0   245.3231  1414.8153
#> 0-11-0  2226.4759   259.4922
#> 0-12-0  4362.1778  2932.8460
#> 0-13-0 -5525.4883  -303.5341
#> 0-14-0  1041.0597  2733.3311
#> 0-15-0  6722.8452  -884.5153
#> 0-16-0  5127.8025  5205.4503
#> 0-2-0   3842.8742  3467.3336
#> 0-3-0   2463.2972  3245.7470
#> 0-4-0  -4957.5745  -769.4618
#> 0-5-0   4327.3801  1662.9670
#> 0-6-0   1122.5211  7824.8995
#> 0-7-0  -3126.6401  -181.5440
#> 0-8-0   3791.3283  1580.4220
#> 0-9-0   4907.3887  7594.1755
#> 1-1-1   1048.5503 -2990.9060
#> 1-10-1  7274.6214  3968.1112
#> 1-11-1  4923.6273   673.2827
#> 1-12-1  4137.4834  4683.5152
#> 1-13-1 -1489.6000 -7165.8213
#> 1-14-1 -1014.4111   451.7184
#> 1-15-1  4491.4200  1835.5243
#> 1-16-1   177.1751 -2048.9232
#> 1-2-1   1972.7997  2539.6040
#> 1-3-1   6824.6698  2246.8100
#> 1-4-1  -2755.6040 -3143.4935
#> 1-5-1   -276.6296   718.4451
#> 1-6-1   1176.0216  -208.3853
#> 1-7-1   3687.1067  2013.4530
#> 1-8-1  16332.0756  5006.9315
#> 1-9-1   5006.9315 17427.6664
#> 
#> $stimcats
#>      stim cat
#> 1   0-1-0   0
#> 2  1-14-1   1
#> 3   0-2-0   0
#> 4   1-3-1   1
#> 5   1-4-1   1
#> 6  1-11-1   1
#> 7   0-8-0   0
#> 8   1-9-1   1
#> 9  1-16-1   1
#> 10 1-10-1   1
#> 11 0-14-0   0
#> 12  0-5-0   0
#> 13 1-15-1   1
#> 14  0-4-0   0
#> 15  1-5-1   1
#> 16  1-6-1   1
#> 17 1-13-1   1
#> 18 0-10-0   0
#> 19 0-11-0   0
#> 20 1-12-1   1
#> 21 0-16-0   0
#> 22  0-7-0   0
#> 23  1-2-1   1
#> 24  1-8-1   1
#> 25  0-6-0   0
#> 26 0-12-0   0
#> 27 0-13-0   0
#> 28  0-3-0   0
#> 29  0-9-0   0
#> 30  1-1-1   1
#> 31 0-15-0   0
#> 32  1-7-1   1
#>