Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

aat_stimulusscores(
  ds,
  subjvar,
  stimvar,
  pullvar,
  targetvar = NULL,
  rtvar,
  aggfunc = c("mean", "median"),
  iters = 5
)

Arguments

ds

the data.frame to use

subjvar

Name of the subject-identifying variable

stimvar

Name of the stimulus-identifying variable

pullvar

Name of the movement-direction identifying variable

targetvar

Optional. Name of the stimulus-category identifying variable

rtvar

Name of the reaction-time identifying variable

aggfunc

The function with which to aggregate the RTs before computing difference scores. Defaults to mean but can be changed to median.

iters

If there are missing values (which is almost inevitable) then multiple imputation will be used to complete the covariance matrix - this argument sets the number of multiple imputations to be used.

Value

Exports a list containing a data.frame with stimulus-specific bias scores, indicated in the column names, a covariance matrix of that same data, and a data.frame indicating to which stimulus category each stimulus belongs.

Examples

ds<-aat_simulate(biasfx_jitter=40,nstims=16)
ds$stim<-paste0(ds$stim,"-",ds$is_target)
aat_stimulusscores(ds,"subj","stim","is_pull","is_target","rt")
#> $data
#>          0-1-0     0-10-0      0-11-0       0-12-0       0-13-0     0-14-0
#> X1   86.768451  345.24045  182.606499  -42.9316924  165.2205495  -80.31089
#> 2   110.712490  -53.43935   47.408215 -105.3874002  153.2528266  -47.87089
#> 3   137.724374   23.68986   68.496364   30.9333573   18.7708824   11.92935
#> 4  -113.212088   19.46136  -79.562092   38.6723322  -56.6465544  -72.07627
#> 5   -14.196429   29.76915  -28.341122  -28.9402089  188.7606904   65.83551
#> 6     1.839975   22.78355 -115.936981  -97.8172916  145.5822413   94.50099
#> 7   -11.655623   54.22626  -46.585700  -62.1936588   70.2069479  -32.93416
#> 8   -11.498552 -139.22768 -364.601996 -197.7881114  -59.5219755   26.58384
#> 9   -78.898665 -269.39615 -260.872697  -36.8101690  -60.1466794 -142.75250
#> 10  246.646568  141.50532  166.051534 -234.1246183   86.2101407   28.60218
#> 11 -301.566239   10.86995 -331.588397  -99.5305165  -88.9872376   60.07757
#> 12 -227.521190  136.35450   -6.251541  328.6076453   62.9857064 -191.43285
#> 13  108.354619  110.18665 -118.042156  178.7280803  -77.4484691 -120.76841
#> 14   72.113540   20.91714   -9.272101   88.0505096  348.7130933  -38.47284
#> 15  207.325092   19.57101  116.054783  -56.6548679   77.1825945   72.06193
#> 16   26.434307  150.22468  -10.443727 -125.1397262 -337.1907525  175.32629
#> 17  -16.470086  326.01705  195.322989  226.9418865  -16.1660390   18.03974
#> 18   99.061958  111.82921   15.308617  -27.4070156  134.6681031 -123.76259
#> 19   -6.390760 -139.61200  -65.258554 -132.5113278 -142.1593466 -128.21595
#> 20  -18.647341 -159.01578  170.242470   16.8542022  107.6306954 -100.86426
#> 21  -29.146869  -49.67767  -30.121677   62.8955028 -102.2251225  -69.94263
#> 22   38.108394 -100.56396   37.019274  160.7063397  -30.6564662  -67.67708
#> 23   89.000296   20.23658  173.366907  273.8256901  277.9665238  126.59509
#> 24  110.250044 -149.83979  279.198013   21.6811788    0.4254085  -33.46760
#> 25 -129.013975  -38.71136  -51.681141 -107.6124724  -78.9738792  -76.52967
#> 26  118.088501 -145.33243   21.681447    0.5292330   54.4781234 -134.84906
#> 27  105.207809  126.22490  -41.486844   51.2510492  -16.5569281   60.76194
#> 28 -128.579470   46.85962 -111.647352   39.3938036  -34.5793467  -81.41101
#> 29  -61.290291 -216.03444 -158.083520   54.7923361  155.6200959  187.08721
#> 30   61.902593   27.79360   15.839869   -0.4587636   -6.0090899   58.44648
#> 31   11.663213   65.83280   32.596098   30.3664091  233.7466924  340.20969
#> 32   -5.195713   66.82687  -49.642431  -67.4461061  -12.0872085  -70.13469
#> 33  174.615731 -134.84086   -1.421212  114.9767050  144.8893143   18.95493
#> 34  -40.621150 -195.27804  -87.018202  -18.4868343  -31.8007737  201.23519
#> 35  121.839519   53.90287    6.187581  149.0675729   96.3542473  119.35092
#> 36  -65.388848  216.08055    9.110799   76.3882672   27.0985468   31.88934
#>         0-15-0      0-16-0        0-2-0        0-3-0       0-4-0        0-5-0
#> X1   -6.756474  166.694175   56.0769697    3.5281864  -49.917732  -51.3451042
#> 2   368.591726  108.826610  113.3284946  226.2467880   37.921253  -19.3385552
#> 3     5.880478  -28.749629   68.4087640   56.7015005 -125.490095   80.9765997
#> 4    23.658119  137.553694 -156.3214452  108.5454835  136.943909  -13.6099150
#> 5    13.824515    4.322536  -63.2578766  -37.4837477 -106.911658  -21.3778303
#> 6    67.801006 -124.587875  -90.0682850  167.5928451   -8.514446  -10.1390937
#> 7  -207.953884  133.391174   54.3283180   11.6757075  133.871052  136.9215021
#> 8   262.104697   26.926681 -259.3397004  133.9694500  146.580593  379.1437348
#> 9    14.810780  233.894410 -187.1413672   21.0436597   12.689750    0.2778031
#> 10  203.674581 -111.500864   44.2420263  251.9280340   86.170903  182.4120488
#> 11   -1.571881  -23.178595  -60.9084695  -98.6467792  -76.661137 -147.5307244
#> 12   99.705360  118.382408   88.2216956  -92.2457304  173.534614 -207.5722452
#> 13  -46.077345   69.263453  116.4174459   62.4058049 -119.509659  -45.9939517
#> 14  132.377833   29.896760  111.7859728    5.3144834  145.796781  117.5155235
#> 15  155.653141   94.333249 -114.1640573   72.8741660   29.662549 -145.3825271
#> 16  -11.904494  -30.922043  -92.6203213  143.7316141  180.018168    9.9451480
#> 17  -15.868042 -127.534117    0.2324006   -4.6174189  216.554786  221.2078420
#> 18  -92.126064  187.589974   66.6887120   19.8931717   98.801576   28.2392393
#> 19  -30.669798 -178.116598 -143.1556629  -51.0199149  -65.841311  -99.6309584
#> 20  -47.842122  -39.957022 -254.5862740  131.3956081   50.362666   38.5239219
#> 21    5.829714   46.082719  -24.8045275  -62.4870855 -101.292010 -143.3658042
#> 22  -16.132924 -144.340488  244.8059319  -22.0433074   95.130232   18.5385513
#> 23  133.827666  238.604258  152.0893677   -0.6236103   77.768237 -120.7895940
#> 24   63.007164  124.171958   24.9426430  -46.4587973   88.567791  106.5267203
#> 25 -186.292057 -123.273555   38.4386338   81.9519075   -3.314604   67.4511961
#> 26  -48.929097   13.900988   54.5399974   -2.1291088  144.125702   23.2028912
#> 27   66.035280   88.245506   -2.9473020 -152.6877150 -178.161655  -25.3227732
#> 28  -60.901695  -35.752831   68.7935073    6.7729674 -284.654968  -56.7596612
#> 29   43.003074   58.419311  209.8782699  175.3575282  -78.435618  205.4985020
#> 30 -109.258992  179.989208   33.3727590   36.3283504   -1.963471  -16.7609567
#> 31  171.120022  -31.817934    9.1557083  -22.4615767  101.029791  118.4245728
#> 32 -135.921523  -28.441279   61.2459723   29.6034142  -36.313116   27.0438331
#> 33   86.765152    4.870134  120.6275993  197.0218497   71.206170  220.8844624
#> 34   10.426037  170.272377    7.3839292    7.2503911   58.333094   20.5736619
#> 35  133.430346  186.610966   73.4236563   33.5254106   95.156634    5.2285633
#> 36  164.697379   98.860146   15.5743630  107.6175132  151.759114   48.4469990
#>          0-6-0       0-7-0       0-8-0        0-9-0      1-1-1      1-10-1
#> X1   18.020109  149.448496  124.342459  -61.3070185  225.38525  201.322517
#> 2  -198.534928 -151.438813   11.790858  288.1846533   83.12940  174.912131
#> 3   -31.870763  237.164991   14.635241  -74.5102225   11.89792  150.792709
#> 4    32.382541  -68.863759  -66.554275   92.0019676  106.88574   26.372599
#> 5   -45.835765  -10.820686 -214.621454  -20.4498943  108.66547  -17.848344
#> 6   184.479061  124.901127 -114.151417   14.2836125   30.85459   50.418555
#> 7   111.840752   51.262070    5.996982  -31.3191492  -71.90786  140.963914
#> 8  -156.448742  246.370881   11.776022  -92.4131158  429.97324   36.824696
#> 9   -98.490469   -1.807191  157.470739 -105.2708722   41.63960   96.263110
#> 10   41.284870  159.618071   11.467611  -98.6870200  166.71692 -425.969480
#> 11    2.069438   14.791606  -71.366059  -74.1177550   32.42985  122.676766
#> 12 -192.611145   85.326111   89.197140 -112.2049242  -35.33382  240.847808
#> 13   75.137396  -26.208345   69.890706  161.7225191   57.24634 -142.452332
#> 14  107.238808    3.953456  138.557664  -72.3251763   40.72722   49.538463
#> 15  -81.248473  -63.919575  137.074314    0.2237277  162.86799   76.263534
#> 16  -33.151109 -197.940315  249.352179  151.4126815 -110.72393  118.772344
#> 17  155.085160  -67.579440   46.599400  186.3179111   57.28291  167.814900
#> 18 -128.399560   88.006843   59.196084  221.0585998   35.77336  186.650995
#> 19  -35.128466  -44.724532 -201.131264 -123.4083394  116.61022   19.986255
#> 20 -250.848318   29.774514  -68.158866 -196.8255050   10.71269   56.649866
#> 21  -81.143225  136.029553  -80.295507  -15.5488958   17.61370   36.911240
#> 22  -20.425776  101.551193   43.716714   -1.6300603   86.78873  -51.991528
#> 23  154.582386  300.715676   66.897095   65.8648162  118.05103   -0.805081
#> 24   27.161135  212.211050   44.987003  -32.8996791  -16.09580 -102.245411
#> 25   28.594576  160.388435 -102.647904  -25.1975194  -32.22696   97.148472
#> 26  273.906878   48.948488 -155.456407 -150.3575374   96.76585   80.113965
#> 27  131.202887 -130.487673   -6.295938  -47.3858373   76.55750   64.965024
#> 28 -253.868318 -124.490889 -236.501657  -32.3415394  -73.69422 -169.408250
#> 29   19.949187    2.288183   66.201480  176.2722954  -69.02088    7.628206
#> 30   -4.419843   69.031864   35.867296  -30.3636603  189.94175  126.994747
#> 31  -73.460599  223.442728  121.295647  -31.4408632 -120.52966   62.595690
#> 32  -33.130958   73.649835   42.978552  -61.8860044  152.99198  -27.994186
#> 33  143.526799   63.510789  132.090726  108.2735490   96.19828  117.173602
#> 34  -49.706914  -28.962503   60.207057   76.1784255  103.98604   61.479910
#> 35  139.918624   26.334887   44.183423   51.0260027   18.30576   61.201381
#> 36  126.857312   58.150630  150.868550   76.8251559  -12.91394  -11.892731
#>         1-11-1      1-12-1      1-13-1      1-14-1       1-15-1      1-16-1
#> X1  250.023176   71.078467  196.795919   35.594619  132.9804110  218.880033
#> 2    97.791653  235.676291  -72.178549  316.053026   75.1099757  207.273089
#> 3   165.473154  -23.070595  193.406450   73.709890  142.8356578   -3.184356
#> 4     3.620559  112.975969   72.181466  -12.411881 -163.5031645   47.455088
#> 5    -8.555550   57.185583   77.087555  -94.331971   96.4346988  169.055189
#> 6    65.890766  -54.717899   19.234053   68.143558  -86.7003999   72.320218
#> 7   107.747193   66.576761  120.488083   69.075259   97.3689877  139.366607
#> 8   179.858265  217.669054  141.960376  -22.346567  -30.9222766   23.323151
#> 9    46.074339   35.559070   93.039608  -74.599962 -121.8559289  -32.042712
#> 10 -142.674878  188.598644  -58.346017   63.923129  -85.6370255  379.216716
#> 11   89.155612  -92.666138 -112.069807   34.819056  -28.9350085  -69.364676
#> 12   -4.463558 -103.605349   90.418311 -150.842918  206.9218594  127.388113
#> 13    1.640607   51.278033   31.732923   66.922638   54.7019917   70.028161
#> 14  -34.200152  -17.901765   17.910582   37.866425   82.9867841   29.572478
#> 15  106.092829  116.058655  -61.595596  -53.055859  -71.2704835  -43.210120
#> 16   25.204629 -135.287178 -426.348358  -42.046827  -38.2994344  -91.879105
#> 17  142.505579  -19.645146   42.190630  -17.402893  173.7200294 -163.129932
#> 18   72.893097   77.992986   97.182707  100.724947  200.8054011  218.140094
#> 19  -11.095869  147.787536   -2.195265  -33.383624   41.4527275   28.325062
#> 20 -101.552117  -41.933585   27.173239 -359.056992  258.7645865  -79.785009
#> 21   93.086647   68.549508  114.005452   59.890397   78.7840179   30.366304
#> 22  121.585148  -67.745935 -281.884506  174.948171  -81.9022308   20.098092
#> 23  -76.727675   -9.266695  172.432413  331.712453  119.0131051  142.965132
#> 24  170.157810  227.948830  109.608916   41.701431  134.3151969 -140.596384
#> 25  -40.884941   40.758919   26.082039 -221.646053   -0.7387592  106.897316
#> 26  -33.205682   16.620965  -21.123887  116.472373  118.5974699   33.387962
#> 27  134.130266  119.610653  152.051305  -42.864731 -167.5653298   79.123762
#> 28  -63.174785   12.263521  -70.504675  -86.660895  -32.2783123 -141.835804
#> 29   83.036310  170.159430   28.769241  117.317866  118.3395326   47.303131
#> 30   16.321668  100.476918   59.725342   85.939885   55.8489139   37.963029
#> 31   47.759021   73.510747  160.793924  154.868008  109.7754204  163.905160
#> 32  114.121629   66.071897  298.410317   46.541564  186.8346734  123.137253
#> 33   77.049764  167.922533   82.913329  163.531587   49.7232607   58.044486
#> 34   65.312856  -22.034114   40.162371   -5.722199   66.1168769   62.435350
#> 35   49.077231  152.337268  154.912390    9.547192  189.2075197  169.061829
#> 36  204.821562   88.589997   43.927441  150.199061   13.6641203   93.330194
#>          1-2-1        1-3-1       1-4-1        1-5-1        1-6-1       1-7-1
#> X1  241.531338  -49.5427721   18.375305   87.7173635 -228.2481160  219.967314
#> 2   -38.330186  148.6279109  172.652145   45.9676764  266.6762168   96.911353
#> 3   134.796968   -8.2556178  -33.819288   14.1661326   44.7936082  128.709800
#> 4   118.516544  -25.7904195  159.510831  -69.6097047    1.0914106  -65.985214
#> 5   114.579149   72.7725659 -139.337934  -41.0700569  154.4964802   69.996252
#> 6    46.462308   73.7876738   31.130707  -25.6441600   23.6516049   -9.918567
#> 7   168.904096 -137.8409257  -36.294741 -124.2751855 -226.6433775   20.219942
#> 8   -32.074636   57.0825654  143.229950 -154.2179597   14.1695807   45.958447
#> 9   103.178885   87.9819315    4.686801 -163.4063387  136.8657281  -54.804936
#> 10   -4.019391   24.5680376  109.427680  108.2815461  141.4549891  258.220020
#> 11  -72.751570 -179.3311058  -50.275668  232.8951905   42.5201309   30.406545
#> 12 -142.989938  183.7464872  393.666961 -141.3636849  343.5258034   -9.655268
#> 13  140.945332  177.0560932 -161.659782   -3.9162785  241.7220316   77.104323
#> 14 -110.564365  172.0901681   95.822514   51.8809005   16.0772484   35.067894
#> 15   45.735351    1.7960163  108.936469  -12.7254224  155.1739472  109.468663
#> 16 -115.889555  140.3170334  -22.011182   61.2955042 -118.9998059  150.512640
#> 17 -315.060382    2.8018637   22.788121  -69.0282180  -11.0877140  274.394635
#> 18  186.063913  132.5742881  254.607146  184.0372470  136.9103166   46.744401
#> 19   19.055200   70.0019182   48.575684   20.2876405    0.4762888   91.245325
#> 20 -105.913310   82.2054881    7.453273  -16.1095869   -2.5331771   24.025806
#> 21  136.712070  225.9629803  143.816452   84.4386861  184.6562876  145.561623
#> 22   43.726023   -7.3636387   32.688593   -5.0609954   58.1152766 -205.890462
#> 23  124.079427  102.8199599  233.251441  168.6152061   39.5593841  314.807622
#> 24   16.407700 -151.3127214  243.088013  195.7739836  159.7303875    5.955375
#> 25   98.740940 -202.0559916  -78.465845  -83.4077544  -57.7633750  -43.160606
#> 26   80.909447    0.3898189  -41.509233  251.3967313  236.7391907   33.323640
#> 27 -167.235190  114.3075847  131.218763  -78.5825457  100.8770876  -49.773376
#> 28   63.329092 -243.4609430  -35.304476   68.6935684   15.4576827 -211.910513
#> 29   44.007643 -107.4501865  -33.239104 -100.9025056  128.3311659  200.743779
#> 30  -41.166954   29.9257726  216.339710  136.6203222   78.9616107   13.300917
#> 31   68.981453  172.8946837  112.881425   -0.8213150  182.7524631   27.959769
#> 32   85.068511   47.8985502   90.753699  113.0923034   40.1622708   75.426241
#> 33  205.181173  230.9916962  164.757510  254.8142757  134.5405398   34.689325
#> 34   27.837980  -26.9689656   50.511207   38.8051241  106.2507268   41.289813
#> 35  138.699519  -21.5610999  172.364115   -0.4958906  117.6853561   70.690315
#> 36   48.463776  168.0059018   81.848307   63.3900405   25.8210516  154.669268
#>          1-8-1        1-9-1
#> X1   -7.120023  -98.1168530
#> 2   -71.958939  143.2369553
#> 3   -40.909205   27.4431113
#> 4   206.014026   -6.4386973
#> 5   -55.588420   28.6735172
#> 6   -51.569921  174.4362647
#> 7   -54.928746   47.3052959
#> 8    21.301389  210.9438063
#> 9   -64.367582   10.2393362
#> 10  -88.595759  358.1276272
#> 11  -46.913911 -114.5172751
#> 12   76.546048  343.7378844
#> 13 -300.772820  214.3090568
#> 14  250.667596   43.5977402
#> 15  -91.845721  146.1442367
#> 16  -78.763085  -53.6127827
#> 17   31.714564 -108.0919117
#> 18  204.249925   57.6818081
#> 19   15.904688  -23.1869697
#> 20  116.419584  -74.5690951
#> 21   52.292514  -20.4798594
#> 22   66.060497   73.6658987
#> 23  112.735822  188.9351873
#> 24  -25.521949  231.3704871
#> 25  163.609512  -60.0744815
#> 26   66.504804  101.2166684
#> 27    6.841399  183.8696178
#> 28  -71.311190  -98.4422665
#> 29  135.354379   44.0341404
#> 30  100.036238   -0.7306805
#> 31  132.291842  142.4801052
#> 32   87.663576  -35.4369912
#> 33   52.375628   88.3973564
#> 34   88.428276   41.2083328
#> 35   75.733824  142.6457555
#> 36  177.713043   27.1369659
#> 
#> $covmat
#>             0-1-0     0-10-0     0-11-0      0-12-0     0-13-0      0-14-0
#> 0-1-0  13145.1763  1205.0493  8678.9438 -1138.27531  4629.0614  1566.48720
#> 0-10-0  1205.0493 19770.3417  6702.2780  3170.91566  1064.1428   652.59330
#> 0-11-0  8678.9438  6702.2780 18848.9547  4705.40166  5919.7972  -111.56583
#> 0-12-0 -1138.2753  3170.9157  4705.4017 15301.76273  4116.0003 -1065.67550
#> 0-13-0  4629.0614  1064.1428  5919.7972  4116.00035 17030.7416  2626.30106
#> 0-14-0  1566.4872   652.5933  -111.5658 -1065.67550  2626.3011 13166.25468
#> 0-15-0  3870.7842  -345.5047   984.2342   -43.16985  5546.5800  4378.67673
#> 0-16-0  1120.8218  -299.7338   268.7808  3583.50466  3385.8436   168.65432
#> 0-2-0   3074.6462  2787.2823  4474.0625  6857.46773  5425.8593   596.15880
#> 0-3-0   3419.3552 -1517.0683   323.1970 -4208.14790  1209.7199  1529.65315
#> 0-4-0   1126.4495  1434.4031  3956.9036  1842.24033  1670.1399   926.14429
#> 0-5-0   3424.4618 -2252.5699  -699.5722 -3360.04051  2054.6149  3337.79875
#> 0-6-0   4255.9425  3461.3050  2530.0209  3035.35094  2685.3350  2779.18208
#> 0-7-0   1735.7694 -1088.1748  2925.7377   829.64711  5178.0913   920.23795
#> 0-8-0   3419.4125  3395.6696  2739.9673  2988.43912  1233.6062  3583.13792
#> 0-9-0   1926.1265  3177.1798   889.8310  2662.55712   303.9401  3310.27887
#> 1-1-1   3568.5982  -652.9840 -1464.7038 -3603.16313   517.2026 -1696.49028
#> 1-10-1 -3301.6368  1365.7286 -1193.4316  2605.62926   694.0408  -647.24190
#> 1-11-1   390.2213  2461.8862  -429.3046  -202.29538 -1223.9225  1016.46688
#> 1-12-1  4794.1007 -2592.4755   873.0632 -3484.34903  2054.6491   303.50049
#> 1-13-1  1109.1491  1183.0714  1096.7951  2986.09970  7083.2895  -895.00368
#> 1-14-1  5669.6012   948.2090  2046.9301  2267.85154  5652.1834  4101.35313
#> 1-15-1   686.5602  1463.4960  6124.7551  4752.88962  5232.1540 -1092.65011
#> 1-16-1  4026.5198  4048.6306  1772.0432 -2203.94063  6743.3494   816.73571
#> 1-2-1   2949.5644 -2413.5222  -770.8696 -1290.46479  2900.1547 -1270.68414
#> 1-3-1   3999.8089  1777.0103  1256.8016  3724.36306  2929.8525    67.53622
#> 1-4-1   1860.8763   857.5500  4519.8492  4104.04542  3589.0159  -475.24975
#> 1-5-1   3983.8247   405.4671  4776.0481  -488.97543  1623.6498   168.23401
#> 1-6-1   1780.2762 -4789.3009  -209.8462  4433.29809  2575.3623  -804.97462
#> 1-7-1   4300.0197  5664.5650  5830.7741   152.84547  3025.8772  3773.80221
#> 1-8-1  -2565.7026 -1308.6057  1420.8075  2870.99667  4653.7821   912.65450
#> 1-9-1   5676.8472  -777.9360  2252.4341  2044.49793  4580.7105  1129.56907
#>             0-15-0     0-16-0         0-2-0      0-3-0      0-4-0       0-5-0
#> 0-1-0   3870.78417  1120.8218  3.074646e+03  3419.3552  1126.4495  3424.46180
#> 0-10-0  -345.50470  -299.7338  2.787282e+03 -1517.0683  1434.4031 -2252.56986
#> 0-11-0   984.23425   268.7808  4.474062e+03   323.1970  3956.9036  -699.57215
#> 0-12-0   -43.16985  3583.5047  6.857468e+03 -4208.1479  1842.2403 -3360.04051
#> 0-13-0  5546.58002  3385.8436  5.425859e+03  1209.7199  1670.1399  2054.61492
#> 0-14-0  4378.67673   168.6543  5.961588e+02  1529.6531   926.1443  3337.79875
#> 0-15-0 14336.52871  1734.9454 -5.028571e+02  3698.3634  3275.8301  2044.19250
#> 0-16-0  1734.94536 12701.3056  1.080041e+03 -1428.5012  1394.6221 -2970.86365
#> 0-2-0   -502.85712  1080.0408  1.348380e+04  -491.2001  -781.8080   -51.19006
#> 0-3-0   3698.36341 -1428.5012 -4.912001e+02  8723.1490  2334.9409  5269.42004
#> 0-4-0   3275.83010  1394.6221 -7.818080e+02  2334.9409 13056.0777  4795.26005
#> 0-5-0   2044.19250 -2970.8637 -5.119006e+01  5269.4200  4795.2601 14564.16004
#> 0-6-0  -1261.98347  -207.3983  4.304543e+03  -282.5380  3254.4187  2691.11236
#> 0-7-0    943.37857   349.6267  1.781670e+03  -298.9251  1901.8452  3313.97708
#> 0-8-0   3736.20813  5030.3818  2.530986e+03  1957.7777  6133.9515  2520.16141
#> 0-9-0   2425.45613  2689.4124  4.791964e+03  3660.3201  1790.9525  1427.04982
#> 1-1-1   3211.47755  1103.3633 -3.160394e+03  1230.1523   467.6110  2149.50401
#> 1-10-1 -1839.16875  3850.9408 -1.173662e+03 -2588.4457  3127.5973 -2305.67901
#> 1-11-1  1110.51900  1760.5135  4.436433e+02 -1205.3171   216.2181  1978.47771
#> 1-12-1  4450.09880  2252.3020 -2.569677e+01  2761.6992  -668.1724  4364.81722
#> 1-13-1  -449.82604  5679.9849 -3.592502e-02 -2797.5615 -2216.9585  1850.08424
#> 1-14-1  5749.09267  3237.5536  8.798256e+03  1869.6366  1859.0735  1294.38528
#> 1-15-1 -2059.58504  1663.7969  3.370067e+03 -1280.8560  2131.9463   600.71898
#> 1-16-1  3260.97691  2470.5634  4.098900e+03  2359.9608  1001.5456   116.92727
#> 1-2-1  -2814.11890  3965.7308  2.833720e+03  1293.6864 -3301.0889 -1173.86051
#> 1-3-1   5819.18840  1443.6447 -6.558685e+02  1140.8005  3434.1544  -996.48602
#> 1-4-1   5874.13141  5609.2643  1.217137e+03  -807.9363  5356.0173 -1458.60681
#> 1-5-1   -499.30750   453.0229  3.905319e+03  -144.9121  -323.4536 -1957.73416
#> 1-6-1   5992.45594  2032.7344  2.740541e+03  -631.1414  -543.8866 -3011.59486
#> 1-7-1   3623.07176  1259.5602  9.676021e+02  2602.3882  2614.4501  1652.15971
#> 1-8-1   -508.45256  1414.1280  1.246078e+03  -731.2417  4881.4511  2111.48682
#> 1-9-1   8131.10173  2426.6233  2.661992e+03  1550.3977  3264.2524  1182.48342
#>              0-6-0      0-7-0      0-8-0      0-9-0       1-1-1      1-10-1
#> 0-1-0   4255.94253  1735.7694  3419.4125  1926.1265  3568.59823 -3301.63676
#> 0-10-0  3461.30498 -1088.1748  3395.6696  3177.1798  -652.98397  1365.72858
#> 0-11-0  2530.02088  2925.7377  2739.9673   889.8310 -1464.70385 -1193.43163
#> 0-12-0  3035.35094   829.6471  2988.4391  2662.5571 -3603.16313  2605.62926
#> 0-13-0  2685.33504  5178.0913  1233.6062   303.9401   517.20259   694.04082
#> 0-14-0  2779.18208   920.2379  3583.1379  3310.2789 -1696.49028  -647.24190
#> 0-15-0 -1261.98347   943.3786  3736.2081  2425.4561  3211.47755 -1839.16875
#> 0-16-0  -207.39827   349.6267  5030.3818  2689.4124  1103.36329  3850.94081
#> 0-2-0   4304.54321  1781.6699  2530.9856  4791.9635 -3160.39441 -1173.66222
#> 0-3-0   -282.53800  -298.9251  1957.7777  3660.3201  1230.15232 -2588.44569
#> 0-4-0   3254.41869  1901.8452  6133.9515  1790.9525   467.61098  3127.59731
#> 0-5-0   2691.11236  3313.9771  2520.1614  1427.0498  2149.50401 -2305.67901
#> 0-6-0  15650.48037  1728.2906  1141.0848  1434.0282    60.47234  -816.12452
#> 0-7-0   1728.29059 13897.4647   645.5101 -4442.2756  2528.61831 -1389.00832
#> 0-8-0   1141.08477   645.5101 12462.9300  3956.6015  -631.46487  3079.78999
#> 0-9-0   1434.02819 -4442.2756  3956.6015 12454.2945 -2021.37813  1938.66397
#> 1-1-1     60.47234  2528.6183  -631.4649 -2021.3781 10929.65177 -1102.50245
#> 1-10-1  -816.12452 -1389.0083  3079.7900  1938.6640 -1102.50245 14817.98211
#> 1-11-1   954.72624  1115.5705  3422.7511  2065.8730  1596.71254  4231.21128
#> 1-12-1   131.15567   996.2698  -336.8487  1800.2480  3653.83174 -3112.94952
#> 1-13-1  1754.13475  7596.8100  -591.0336 -2908.0636  3156.22969  2212.96948
#> 1-14-1  6049.75392  3528.5689  3815.2736  6937.6185  1208.75382  -757.16389
#> 1-15-1 -1370.27915  3466.9012   689.9793   336.2687 -1945.82275  4081.57770
#> 1-16-1  1597.20365  3873.4464   810.0111  1184.2163  2425.04684 -1852.38156
#> 1-2-1   1400.88771  5022.0267  -847.1680   652.1405   546.46155 -1148.69176
#> 1-3-1    -10.71896  -130.8021  5202.4644  2333.9903  1179.72424  2143.17494
#> 1-4-1  -2038.89173  3514.6275  4012.0081   927.4586  2035.09286  2561.71610
#> 1-5-1   3377.20415  2042.4065  -187.1484   793.7816   640.69117 -1744.73593
#> 1-6-1  -2349.47449  -149.8103  -224.4510  1527.1374  -405.98865 -1921.03712
#> 1-7-1   3926.61770  2607.0113  3957.2137  3279.6744  1995.99421    21.59627
#> 1-8-1   1017.27188  2901.0248   962.5083  -279.5806  -744.33243  3512.83569
#> 1-9-1   1752.27861  4456.2107  2575.6220  -163.2689  2049.18098 -4675.62103
#>            1-11-1      1-12-1        1-13-1      1-14-1      1-15-1     1-16-1
#> 0-1-0    390.2213  4794.10068  1.109149e+03  5669.60122   686.56017  4026.5198
#> 0-10-0  2461.8862 -2592.47548  1.183071e+03   948.20898  1463.49602  4048.6306
#> 0-11-0  -429.3046   873.06319  1.096795e+03  2046.93013  6124.75508  1772.0432
#> 0-12-0  -202.2954 -3484.34903  2.986100e+03  2267.85154  4752.88962 -2203.9406
#> 0-13-0 -1223.9225  2054.64909  7.083290e+03  5652.18337  5232.15401  6743.3494
#> 0-14-0  1016.4669   303.50049 -8.950037e+02  4101.35313 -1092.65011   816.7357
#> 0-15-0  1110.5190  4450.09880 -4.498260e+02  5749.09267 -2059.58504  3260.9769
#> 0-16-0  1760.5135  2252.30205  5.679985e+03  3237.55356  1663.79691  2470.5634
#> 0-2-0    443.6433   -25.69677 -3.592502e-02  8798.25597  3370.06744  4098.8999
#> 0-3-0  -1205.3171  2761.69920 -2.797561e+03  1869.63663 -1280.85604  2359.9608
#> 0-4-0    216.2181  -668.17238 -2.216959e+03  1859.07351  2131.94629  1001.5456
#> 0-5-0   1978.4777  4364.81722  1.850084e+03  1294.38528   600.71898   116.9273
#> 0-6-0    954.7262   131.15567  1.754135e+03  6049.75392 -1370.27915  1597.2037
#> 0-7-0   1115.5705   996.26980  7.596810e+03  3528.56889  3466.90119  3873.4464
#> 0-8-0   3422.7511  -336.84871 -5.910336e+02  3815.27357   689.97926   810.0111
#> 0-9-0   2065.8730  1800.24796 -2.908064e+03  6937.61855   336.26868  1184.2163
#> 1-1-1   1596.7125  3653.83174  3.156230e+03  1208.75382 -1945.82275  2425.0468
#> 1-10-1  4231.2113 -3112.94952  2.212969e+03  -757.16389  4081.57770 -1852.3816
#> 1-11-1  7852.1192  1688.21851  2.801305e+03  2992.33532   137.38180 -1285.6566
#> 1-12-1  1688.2185  9063.95768  4.399558e+03  3243.68450  -666.19826  3485.6561
#> 1-13-1  2801.3046  4399.55799  1.739634e+04   670.94153  5697.71659  4555.7216
#> 1-14-1  2992.3353  3243.68450  6.709415e+02 16888.08780   -77.16018  5010.0466
#> 1-15-1   137.3818  -666.19826  5.697717e+03   -77.16018 11847.23131  1108.8196
#> 1-16-1 -1285.6566  3485.65615  4.555722e+03  5010.04660  1108.81959 12986.6033
#> 1-2-1    890.5674  2823.83196  5.206389e+03  4391.29256   663.22179  5398.8210
#> 1-3-1   -364.7688   -75.81008  1.262619e+03  3671.19263  1303.99678  3640.7450
#> 1-4-1    706.4049  3060.06915  4.220009e+03  3528.51332  2091.49155  2905.3866
#> 1-5-1   -672.6163   330.79263 -1.285652e+03  6449.86917  2512.43562   353.4748
#> 1-6-1  -2169.4067  2264.52756  8.413676e+02  2837.94982  1405.46494  1844.1772
#> 1-7-1    900.9133  1568.16991  2.655269e+03  4428.69894  4301.99048  3564.2106
#> 1-8-1   -546.2714  -255.22468  3.367393e+03   186.32291  2964.59509  1308.5051
#> 1-9-1  -1344.5337  3717.41213  1.995681e+03  4163.20069 -1018.55230  5993.1986
#>             1-2-1       1-3-1      1-4-1      1-5-1      1-6-1       1-7-1
#> 0-1-0   2949.5644  3999.80892  1860.8763  3983.8247  1780.2762  4300.01971
#> 0-10-0 -2413.5222  1777.01025   857.5500   405.4671 -4789.3009  5664.56501
#> 0-11-0  -770.8696  1256.80158  4519.8492  4776.0481  -209.8462  5830.77405
#> 0-12-0 -1290.4648  3724.36306  4104.0454  -488.9754  4433.2981   152.84547
#> 0-13-0  2900.1547  2929.85246  3589.0159  1623.6498  2575.3623  3025.87716
#> 0-14-0 -1270.6841    67.53622  -475.2498   168.2340  -804.9746  3773.80221
#> 0-15-0 -2814.1189  5819.18840  5874.1314  -499.3075  5992.4559  3623.07176
#> 0-16-0  3965.7308  1443.64469  5609.2643   453.0229  2032.7344  1259.56019
#> 0-2-0   2833.7198  -655.86853  1217.1366  3905.3188  2740.5413   967.60213
#> 0-3-0   1293.6864  1140.80053  -807.9363  -144.9121  -631.1414  2602.38819
#> 0-4-0  -3301.0889  3434.15444  5356.0173  -323.4536  -543.8866  2614.45005
#> 0-5-0  -1173.8605  -996.48602 -1458.6068 -1957.7342 -3011.5949  1652.15971
#> 0-6-0   1400.8877   -10.71896 -2038.8917  3377.2041 -2349.4745  3926.61770
#> 0-7-0   5022.0267  -130.80214  3514.6275  2042.4065  -149.8103  2607.01127
#> 0-8-0   -847.1680  5202.46436  4012.0081  -187.1484  -224.4510  3957.21369
#> 0-9-0    652.1405  2333.99035   927.4586   793.7816  1527.1374  3279.67436
#> 1-1-1    546.4615  1179.72424  2035.0929   640.6912  -405.9887  1995.99421
#> 1-10-1 -1148.6918  2143.17494  2561.7161 -1744.7359 -1921.0371    21.59627
#> 1-11-1   890.5674  -364.76882   706.4049  -672.6163 -2169.4067   900.91331
#> 1-12-1  2823.8320   -75.81008  3060.0691   330.7926  2264.5276  1568.16991
#> 1-13-1  5206.3890  1262.61884  4220.0086 -1285.6521   841.3676  2655.26886
#> 1-14-1  4391.2926  3671.19263  3528.5133  6449.8692  2837.9498  4428.69894
#> 1-15-1   663.2218  1303.99678  2091.4916  2512.4356  1405.4649  4301.99048
#> 1-16-1  5398.8210  3640.74498  2905.3866   353.4748  1844.1772  3564.21056
#> 1-2-1  13711.5066 -1035.70454 -1789.2052  2886.7016 -1057.5005  -523.70550
#> 1-3-1  -1035.7045 14749.68894  4981.9723   422.7634  5861.4972  3344.41769
#> 1-4-1  -1789.2052  4981.97235 14025.0699  2092.6081  4912.7610   621.71102
#> 1-5-1   2886.7016   422.76343  2092.6081 12827.6819  1551.2185  2225.71449
#> 1-6-1  -1057.5005  5861.49721  4912.7610  1551.2185 14612.8586 -1048.44654
#> 1-7-1   -523.7055  3344.41769   621.7110  2225.7145 -1048.4465 12828.54804
#> 1-8-1   -468.6298  1062.06936  5315.6044   837.2477 -1301.9919  -594.41177
#> 1-9-1   -335.6278  4667.00744  6813.5616 -1114.0687  8425.0797  1007.40065
#>             1-8-1      1-9-1
#> 0-1-0  -2565.7026  5676.8472
#> 0-10-0 -1308.6057  -777.9360
#> 0-11-0  1420.8075  2252.4341
#> 0-12-0  2870.9967  2044.4979
#> 0-13-0  4653.7821  4580.7105
#> 0-14-0   912.6545  1129.5691
#> 0-15-0  -508.4526  8131.1017
#> 0-16-0  1414.1280  2426.6233
#> 0-2-0   1246.0784  2661.9925
#> 0-3-0   -731.2417  1550.3977
#> 0-4-0   4881.4511  3264.2524
#> 0-5-0   2111.4868  1182.4834
#> 0-6-0   1017.2719  1752.2786
#> 0-7-0   2901.0248  4456.2107
#> 0-8-0    962.5083  2575.6220
#> 0-9-0   -279.5806  -163.2689
#> 1-1-1   -744.3324  2049.1810
#> 1-10-1  3512.8357 -4675.6210
#> 1-11-1  -546.2714 -1344.5337
#> 1-12-1  -255.2247  3717.4121
#> 1-13-1  3367.3930  1995.6811
#> 1-14-1   186.3229  4163.2007
#> 1-15-1  2964.5951 -1018.5523
#> 1-16-1  1308.5051  5993.1986
#> 1-2-1   -468.6298  -335.6278
#> 1-3-1   1062.0694  4667.0074
#> 1-4-1   5315.6044  6813.5616
#> 1-5-1    837.2477 -1114.0687
#> 1-6-1  -1301.9919  8425.0797
#> 1-7-1   -594.4118  1007.4006
#> 1-8-1  12054.4732 -2673.6941
#> 1-9-1  -2673.6941 14607.9457
#> 
#> $stimcats
#>      stim cat
#> 1   0-1-0   0
#> 2  1-14-1   1
#> 3   0-2-0   0
#> 4   1-3-1   1
#> 5   1-4-1   1
#> 6  1-11-1   1
#> 7   0-8-0   0
#> 8   1-9-1   1
#> 9  1-16-1   1
#> 10 1-10-1   1
#> 11 0-14-0   0
#> 12  0-5-0   0
#> 13 1-15-1   1
#> 14  0-4-0   0
#> 15  1-5-1   1
#> 16  1-6-1   1
#> 17 1-13-1   1
#> 18 0-10-0   0
#> 19 0-11-0   0
#> 20 1-12-1   1
#> 21 0-16-0   0
#> 22  0-7-0   0
#> 23  1-2-1   1
#> 24  1-8-1   1
#> 25  0-6-0   0
#> 26 0-12-0   0
#> 27 0-13-0   0
#> 28  0-3-0   0
#> 29  0-9-0   0
#> 30  1-1-1   1
#> 31 0-15-0   0
#> 32  1-7-1   1
#>