Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

aat_stimulusscores(
  ds,
  subjvar,
  stimvar,
  pullvar,
  targetvar = NULL,
  rtvar,
  aggfunc = c("mean", "median"),
  iters = 5
)

Arguments

ds

the data.frame to use

subjvar

Name of the subject-identifying variable

stimvar

Name of the stimulus-identifying variable

pullvar

Name of the movement-direction identifying variable

targetvar

Optional. Name of the stimulus-category identifying variable

rtvar

Name of the reaction-time identifying variable

aggfunc

The function with which to aggregate the RTs before computing difference scores. Defaults to mean but can be changed to median.

iters

If there are missing values (which is almost inevitable) then multiple imputation will be used to complete the covariance matrix - this argument sets the number of multiple imputations to be used.

Value

Exports a list containing a data.frame with stimulus-specific bias scores, indicated in the column names, a covariance matrix of that same data, and a data.frame indicating to which stimulus category each stimulus belongs.

Examples

ds<-aat_simulate(biasfx_jitter=40,nstims=16)
ds$stim<-paste0(ds$stim,"-",ds$is_target)
aat_stimulusscores(ds,"subj","stim","is_pull","is_target","rt")
#> $data
#>          0-1-0       0-10-0      0-11-0     0-12-0     0-13-0      0-14-0
#> X1  136.213343  152.8141419   86.522452  248.18889   22.41851  213.529152
#> 2    99.750243   24.3809069  -48.137885  -31.01862   33.80562  138.012419
#> 3  -105.162978    0.5977615  -15.253518   48.51468  -98.65344  -44.050312
#> 4   224.889344   19.9312090  203.894271  -57.22730    8.23355 -305.998266
#> 5  -223.074987  -44.2394177 -243.321742   19.14706 -127.63039  -74.754189
#> 6    48.645544  -51.4455088  208.836167 -152.30989  210.85791   63.779065
#> 7   -19.135087   76.4951366  -29.468544  -17.25394  -77.59162    3.965357
#> 8  -181.691003  -88.4430764 -169.118494   88.53223   21.15896   38.044512
#> 9     1.809645  164.2204574  -15.425010 -110.74753   10.36967 -152.624868
#> 10  -41.650100  -12.0850689  458.888493   55.16346  166.14443  -93.224499
#> 11  162.324725 -124.3944222  -64.042270  173.84671  183.74129 -100.458971
#> 12  -84.100540   39.3288402   -7.855051  -49.02720   82.59188   -7.594818
#> 13   91.919381   17.8015437  -86.909051   26.23868   60.75000   13.057518
#> 14  110.568743   51.4499711  -15.373560  130.04245  152.42035   19.283195
#> 15 -158.891959 -105.9633016 -138.321730   80.91242  -73.40940  -68.506468
#> 16   11.342879   77.5874124 -118.846154 -228.53380  -29.90133 -129.459852
#> 17   66.836338 -160.3405249   28.977306   28.42396 -108.00674   68.925025
#> 18  -44.743068  -81.0354805   54.697221   48.38501   49.52521   45.933913
#> 19  -10.099810  263.1688197   89.136155  -47.50797  243.60851 -149.525828
#> 20   12.493277   44.5968995  -35.786130  196.80554 -227.80724  114.354909
#> 21  -20.350366  -37.0506743 -115.119865  -66.80070  -82.30849 -150.444676
#> 22 -330.939212 -192.4761634 -316.439269 -109.68630  266.80845   36.616897
#> 23  -94.395276  123.6382874  -57.814625   37.15371   26.18664   -4.660482
#> 24  -21.740254 -242.2048989   82.877511   16.40874 -164.33443   -6.941632
#> 25  110.722145  -40.5326112  319.708439  109.12106  222.45628  217.033009
#> 26   69.712470  202.1923604 -357.731685  162.35425  131.68916  153.695003
#> 27   68.445600   14.6958567  -63.720523  137.89656  -31.97327   85.884185
#> 28    8.043021   57.7640408  275.420694   82.02044  215.19893  128.860130
#> 29 -163.076697   45.4455911   25.261995   13.29966  171.93924  150.508783
#> 30   -6.480407   93.9161568  -13.564782  156.18286  -72.84740  -58.974496
#> 31  -29.251079  181.7777193  154.140276   19.83429  142.86961  -39.577290
#> 32 -132.610194  197.4091119  167.165153  212.53862  309.90306   68.977019
#> 33  295.977106  188.1676805   67.333963  -55.57598  169.29937 -151.686537
#> 34   47.363563  160.0023647  115.462587  -34.99235  125.15192   75.772061
#> 35   -8.902658  -58.5666102 -188.489626 -123.77003 -142.34112  -62.156984
#> 36   30.293424   74.4554678 -107.219386 -134.78776  -47.95721   -7.516706
#>         0-15-0     0-16-0       0-2-0      0-3-0       0-4-0       0-5-0
#> X1  206.793990   67.97997  -55.512849  -47.18038  -83.702835 -161.616413
#> 2   -79.688559  -72.92170 -112.438971  -99.11952   28.697064  -53.261113
#> 3    79.470484  -32.36595   61.234655 -110.46420   14.565963  -71.258899
#> 4    33.151295  159.13216  297.149955 -105.68180  112.158232   92.296274
#> 5  -172.290883 -143.19939 -168.024200 -250.75548  -31.639762  200.173645
#> 6   201.630617   -2.93553   36.335593   21.59141  168.429773  -63.817430
#> 7  -149.845056   10.19038  -76.228300  -36.36156   86.467718  -83.709790
#> 8    -2.954648  167.39259   18.627324  -96.90678  116.538006  169.105749
#> 9   -74.121580 -184.72373 -160.258590  -22.85767 -148.958183  -89.015608
#> 10  -76.405214  -54.59951  150.626678  147.59612  197.231108   36.860988
#> 11  -54.688388 -224.71971   84.939760  261.68648 -141.562058 -130.931499
#> 12  -52.685916  195.88902  -36.585812   23.39378  106.927088 -149.693100
#> 13 -289.728724   39.87193  -43.140468  288.43558  -73.516235 -118.697851
#> 14   56.718316  176.51510  139.072138  -25.26151   31.396047  141.778726
#> 15 -188.733203 -184.91747  -32.657185 -125.25409   -9.979085   12.032734
#> 16  -43.653804   66.09698  -55.355006  -25.47269   65.288211  -37.683830
#> 17   26.068633  -11.92033  139.934105  -84.19312  180.937832  143.444382
#> 18    6.343480  -71.42138  -46.090676  -62.00947   30.189755  -46.348198
#> 19 -106.860975   23.26703  -88.264847  -19.36192 -160.876897   23.780597
#> 20 -127.839096  121.61771  264.655003   85.06120  -53.242889  221.232556
#> 21  -34.999882  284.57567  -19.079178  185.32615   51.947797  -29.251102
#> 22 -104.057770 -364.80636  160.871593  198.97225  328.571682   66.943299
#> 23  -68.772605   15.20963 -196.213659   83.42773   40.952855  153.955522
#> 24  130.082717  137.78039   -3.645610   91.72041  105.580785 -238.935401
#> 25   31.010762 -124.34348 -105.822861  104.17371   -6.744193  166.893849
#> 26  240.134721 -312.30629   66.312147   31.26704 -138.535012  -36.676869
#> 27   26.726255  -51.10561   -9.446284  -48.42405  150.401542  -55.775989
#> 28  -28.293384  -41.21581   74.240360   57.67217  -19.771265  -35.555534
#> 29   63.353853  -89.70773  103.130901  197.09029  110.245232   20.926056
#> 30   79.806048  -66.49703   79.728417 -224.40727 -191.154413  -13.075397
#> 31  -18.012868   64.24843  -72.938733   19.28817  -24.848211  -37.438313
#> 32  -84.469019   79.71114  -73.297820 -136.26083  279.819940  317.481880
#> 33  -28.572382  204.20898  107.462999   14.90953   -7.230053   58.193116
#> 34  246.463074 -163.62656   26.405997 -111.69225 -192.889312  335.536723
#> 35  -88.182352   43.28087  -37.741145 -116.85502   18.582217   -7.833409
#> 36  -58.596910  -23.83022   10.608650  -88.86158  -53.369078   15.314551
#>          0-6-0      0-7-0       0-8-0        0-9-0        1-1-1       1-10-1
#> X1  -41.354907 -226.36585  -80.577982 -147.2538907  110.5124850  -43.6366458
#> 2   -72.042420   64.25398  -68.512054  -37.0946520 -130.5931406    7.5345476
#> 3    90.365032  -46.73922   65.969919 -101.4678203  128.2192371    9.3821049
#> 4   109.261776  232.84943  174.720523 -144.2262685  192.0697642  232.9255006
#> 5  -198.572324   50.19433 -214.874556   14.6389324  281.5147261   69.0245807
#> 6   -75.750019  236.72871  137.679264  169.3039464  -25.4322284  -33.2529328
#> 7   -50.046566  121.16747  -59.459917   39.8227399  139.9386701   39.4333434
#> 8   273.152153  108.10509  118.966847   81.2919136 -145.9517339  136.2712806
#> 9    70.276660   48.24603 -160.934763 -146.1153731 -100.5811476  -71.2446423
#> 10 -160.880275  298.60422   45.993886   95.4589240   61.6188064  -61.0552809
#> 11 -227.934014  -14.36564  105.816666  274.0032980   -2.4475366  152.6027750
#> 12  -96.191807  270.90112  258.059368  -78.3880953 -130.6393847  246.1421413
#> 13   29.528073  -51.14514 -175.207652   19.9440693    0.8888629  225.6568486
#> 14   23.150051 -146.37448  -15.318728   -3.2602932  118.5801774   82.4301278
#> 15   -9.719917   82.16372  -29.499328 -123.6781788  136.6550911  -57.1828780
#> 16    5.288144  -81.22738 -116.165322  -66.2467822   28.1628059   -9.1262113
#> 17  179.952692 -317.90517 -100.239667  142.3218779   -9.8907713  377.4470452
#> 18  -93.617677  137.54543  -35.255164   85.2344898  188.3893798  -38.6264561
#> 19  331.050685   60.96616   21.739796  -63.4172160  -26.6777321  121.2298450
#> 20 -111.603613   62.03045   44.089108 -110.1950130  164.1732230  -16.4895561
#> 21   51.124616   56.52856  -11.733145   -1.0316695   29.6446132   36.2333280
#> 22  -81.081335  294.74699 -192.965207    1.3446820    2.9566775   77.7560682
#> 23 -135.079330  -23.66825  -55.298755   -7.3436448  -53.9460334    0.8345058
#> 24   -9.659055 -111.68944  -56.629350 -164.2789747   96.6311472  149.2913633
#> 25 -148.999985  -41.31161  234.911410    2.4964709   13.4901933 -236.7302839
#> 26 -335.258704   67.99430  279.647683  156.8351745 -100.3099503  139.5260856
#> 27   26.477316  -37.62228  -68.892809    0.1262139  100.5900366   93.7522699
#> 28    8.276062   71.67823 -111.325345  -13.2794250  144.6230911   76.6650245
#> 29 -139.913104  -18.11021  205.529662 -104.2840643  205.8096463   51.6161670
#> 30  -76.663057   92.05049   95.761334    5.0830760  103.9347486   37.1195708
#> 31  238.967236   48.28809  -30.225966   29.8088228   94.5676708  -17.5183673
#> 32 -175.316100   42.76077  126.041373  159.4156270   40.0530843   50.1002765
#> 33  373.864431  151.88402   29.463344  -10.6078654  201.6523862  -21.5667821
#> 34  -67.438434  196.08517   77.546136   -8.0470445   53.7059363  152.2661586
#> 35 -176.702614  -53.06834   -8.133961 -113.5234307  123.5708953   59.2252537
#> 36  -20.775312   22.35139   -7.128520    7.9468473  233.3365864   45.3875511
#>        1-11-1      1-12-1      1-13-1       1-14-1     1-15-1     1-16-1
#> X1 -111.03199  -36.498255 -160.328131   51.2780968 -233.46709 -134.29823
#> 2   -14.14242 -101.048855  -50.628407  -24.9847844  -38.20513  124.93092
#> 3   121.71637   49.973645    9.153130  128.8921596  304.74661   75.85057
#> 4  -289.60188 -197.449092 -142.876069  146.9349121  170.80205   40.93319
#> 5   209.15083   59.038115  -87.341881   22.3917327 -140.25295  -33.74276
#> 6   238.44712   78.461208  -14.973145  -51.5183763 -137.08961   90.13576
#> 7   174.46111  -92.456204  -55.942832   33.2713113   21.17317   33.75024
#> 8   -95.10066 -160.044407 -142.382142  116.7454446 -102.01462 -163.00733
#> 9  -100.27427  -86.953627   51.578278   64.8038684   36.37579 -101.04250
#> 10   99.50898  163.938229  169.026493   80.6283033  226.59441 -313.79866
#> 11   45.23408   46.293491  153.598067   75.3459464  100.21741   98.95467
#> 12  -74.13175   44.223872    8.867692  -28.3610014   14.65577  -85.18343
#> 13   35.13387  -71.091844 -124.572337   74.7762108  -99.85609  339.19938
#> 14   73.10203  193.894047  144.176433  131.6414774   85.05351    2.12319
#> 15   11.69515  173.172835  137.482955  143.8673956   92.70278   76.50050
#> 16   38.43556  112.887243   40.185513   -2.3429754  116.64599  -13.59487
#> 17   36.56455  -22.048175   -8.115406  157.4570120  247.68569  194.07984
#> 18   61.69405  115.888459  -48.288989    0.1610268  310.33288   64.04794
#> 19  173.24587  -65.678610  293.352791  275.3024396  130.92874  -86.34803
#> 20 -201.00780   50.565475   98.473189 -217.2125804  102.76698  -34.81466
#> 21   10.42624  161.399322  -77.862725   28.4613191   22.46862  176.73845
#> 22  194.29337  455.342040   82.119532   -3.7664110  -64.99238  -51.22308
#> 23  216.57073  -91.642303 -179.061340 -115.1664334   72.92347   92.98382
#> 24   98.86328  -46.718152   89.270035  189.0526021  -16.26144  313.17994
#> 25  -35.35055  222.808918   60.160274  115.1575604  255.54536 -287.61139
#> 26 -151.39150  -85.097620  401.721999  371.7726079  177.75448  -63.49870
#> 27  235.33090  111.427162  197.352292  134.0896385  198.36738  112.38770
#> 28  252.70058   94.937506  287.942753  185.1199410  232.16212   33.64513
#> 29   58.55989   15.013424   87.332246  199.1964912 -182.77508  130.27713
#> 30 -109.02560   89.596661   17.306370 -222.6377795   86.81164 -109.58725
#> 31   95.55809  -51.597874  -55.290264  213.9710702   95.20534 -150.89664
#> 32  199.14474   -1.190935  -76.495072  116.4506696   52.19881   89.12214
#> 33  100.03436  100.605752  -68.130038   39.3983489  -19.63507   51.10352
#> 34  137.85790  -45.749172  327.393261   45.3236376  -21.99179  255.24986
#> 35  203.03738  132.944489   94.173674  262.1393731   94.45558  122.60969
#> 36  -25.52489  111.608318   88.338299  107.8307534   61.25520   59.16948
#>          1-2-1       1-3-1       1-4-1       1-5-1      1-6-1       1-7-1
#> X1 -227.100329 -244.596990   74.903386   23.400174 -149.31694  -49.985479
#> 2  -111.906057  -99.886339   21.806838  -39.291742  -43.81445  -55.481679
#> 3    87.095990  151.751103  185.883229  302.331779  162.58406  251.226896
#> 4    17.993294 -232.695609  194.714340   92.511315  128.02740  -19.353683
#> 5    17.157621   31.971262 -109.370244   17.000180   20.83565   80.802311
#> 6    66.173181   35.183754  191.216120   58.296159   40.36461   64.005063
#> 7    65.078669  117.896691   76.170821  126.898463   19.57927  -43.432298
#> 8    39.614010   77.064788  135.544448   42.255885  -79.97800  253.864581
#> 9    98.493953   80.472386 -109.428212  -30.011599  108.97615    6.303013
#> 10  -67.045548 -223.474605   10.716258  122.997564   47.50803 -135.797305
#> 11   75.814362   61.384359  -33.453403  -17.630405   45.76682   91.179769
#> 12  -32.859642   66.905489 -202.323115   72.240489  112.46922   74.869012
#> 13 -143.046289 -205.523073  287.718036  206.634801   76.92073  112.113354
#> 14   73.956808  124.890895  119.448993   86.454675  139.77713   39.986884
#> 15   41.323259  -21.107547  121.418365  147.332981   86.16641   66.686445
#> 16   76.071348  112.779849  -88.706530   84.176062  188.92575   66.909799
#> 17  363.572769 -264.864829  -45.913521  156.669621  120.41695 -148.088811
#> 18    1.881313   25.434118  150.312342  -78.289327   61.40007   57.784995
#> 19  -58.110837 -175.157256   44.598031  -38.305116 -125.33955 -326.332990
#> 20 -128.537812   76.492550  -41.553542  -51.285992   64.41389 -100.521114
#> 21   67.829798   58.554094   30.878657  146.464047   83.59060   73.319696
#> 22  120.585526  -51.837306  238.146896 -174.232089  124.30568   54.445869
#> 23  -10.896957   84.422000    7.633180  130.643049  160.68984  213.018475
#> 24  -12.176226   -4.918944  157.424636  -62.270393   35.84351  -10.173222
#> 25  216.311108  132.379721 -467.158704   42.925591   84.92853  -81.038454
#> 26 -230.737218   35.616408    8.923361  232.869758  280.77508    9.758998
#> 27  109.371653  220.135660  -69.353438  103.169602  126.60031  109.774123
#> 28 -108.208417   64.859672  308.245522  231.382271   80.39062  -83.296952
#> 29  -21.416168  316.396125   18.063933 -118.459290   59.72129  221.842835
#> 30   -2.663314 -122.071445   79.541221 -114.211468   85.97719  150.804602
#> 31  243.217652   77.185676  -47.191560  145.874282   14.70658   35.425120
#> 32  175.623819  201.739636   52.251332  291.562981  -15.33928   27.568400
#> 33   18.942012 -123.017170  357.607643   -3.490593  159.54386 -133.108586
#> 34   52.623133   55.591322   30.558918  -45.249028  192.58429  139.211540
#> 35   48.493832   83.420529   29.329222    7.056419  172.07580   88.929234
#> 36  167.939180  213.701825  117.331001  -63.364897   65.99389   71.964646
#>          1-8-1      1-9-1
#> X1  135.207044 -265.77546
#> 2    34.595487 -143.81392
#> 3    44.709896  277.00679
#> 4  -150.596680 -104.36850
#> 5  -152.606578  -42.29845
#> 6     7.037167   64.07300
#> 7    21.675942  -22.96150
#> 8   -60.674705  294.90853
#> 9    96.017902 -109.97008
#> 10  114.231151  -18.64018
#> 11 -159.336248   39.81824
#> 12   58.813983 -151.34168
#> 13   72.292652   19.25357
#> 14  152.545279  156.84731
#> 15 -150.428131  120.62904
#> 16   -7.816235  -26.00231
#> 17  174.221591   93.16131
#> 18   22.436791   62.08914
#> 19   10.915387  257.81153
#> 20  132.852884  -89.08349
#> 21   78.824757  112.10095
#> 22  413.853863 -165.80367
#> 23   25.385235   11.14815
#> 24  106.713101   86.06654
#> 25 -133.154392  287.27720
#> 26   21.815551 -511.17540
#> 27  187.770235  115.79933
#> 28  121.644747  151.38552
#> 29  531.478656  387.65431
#> 30  111.512327  -14.80840
#> 31 -154.017722  163.39431
#> 32   69.693825 -139.15634
#> 33   -6.264422  126.17725
#> 34  137.591508   30.63148
#> 35  -63.833015  160.55537
#> 36   88.382261   85.38976
#> 
#> $covmat
#>             0-1-0     0-10-0      0-11-0      0-12-0     0-13-0      0-14-0
#> 0-1-0  15443.3127  4020.3567  6937.68942  1009.61295   705.1849 -1206.86799
#> 0-10-0  4020.3567 14062.1101  2950.69378  1042.52727  4918.4600  -322.70052
#> 0-11-0  6937.6894  2950.6938 28674.95842  2161.45050  8639.8251   642.72134
#> 0-12-0  1009.6130  1042.5273  2161.45050 12494.66650   861.8480  5967.46987
#> 0-13-0   705.1849  4918.4600  8639.82511   861.84804 18834.7608  2794.82240
#> 0-14-0 -1206.8680  -322.7005   642.72134  5967.46987  2794.8224 13189.04377
#> 0-15-0  4063.5524  1511.9495  3974.05905  2076.23693  2144.7525  4583.29956
#> 0-16-0  4939.6464  1896.8581  6089.77921 -1160.93305 -4700.2492 -4321.96554
#> 0-2-0   3102.8335 -2974.0315  2563.94707  1729.79713   342.7013 -1116.38869
#> 0-3-0   1099.8694 -2937.0941  1930.43117   -77.02285  6012.8708  1461.75841
#> 0-4-0  -5788.4536 -6653.6437  2372.37373 -2179.95781  2511.4609   987.43433
#> 0-5-0  -2797.1953  2611.9467  1616.42921  2229.40591  2847.1574  1485.73303
#> 0-6-0   3933.7882  2941.5737  4015.31659 -4381.68456  -243.4899 -6861.24190
#> 0-7-0  -3712.5413  1017.4058  3051.33719 -4381.39200  6231.7378 -4215.54911
#> 0-8-0   2754.2719  2606.0502  5113.94102  3176.47140  5188.0836  1925.45214
#> 0-9-0    667.9133  -991.3052   374.50896  2565.15689  5909.5318  1873.81918
#> 1-1-1    549.3872  -529.6293  2204.08015   754.60147 -4050.5469 -1699.48716
#> 1-10-1   762.5526 -2294.3641 -4652.33376  -469.20608 -1722.8936 -1869.07515
#> 1-11-1 -5278.3136 -1197.1420  1806.69667 -2843.35543  4955.3209  1173.60759
#> 1-12-1 -4248.6676 -5591.2879 -1731.28724 -1424.13840  3004.3457  1130.60354
#> 1-13-1   679.3795  2026.2630   -39.69338  1171.12209  4574.9410  2694.21601
#> 1-14-1   987.7963   914.1632    12.96591    62.88799  3890.4172   104.28878
#> 1-15-1  2797.2582  -589.2004  5905.76220  2006.80421  -824.4702 -2090.42652
#> 1-16-1  2101.6550 -4205.5230 -5571.30876 -2971.74364 -5171.0863  -336.26854
#> 1-2-1  -1798.8193 -3608.8970  1999.95330 -3419.08097   330.2519 -2343.49441
#> 1-3-1  -5755.3975  1075.7691 -4588.17314  -345.53216   448.1511  3989.83894
#> 1-4-1    926.0919  -874.3306 -1502.25446 -1720.59440  1441.9144 -3333.25096
#> 1-5-1    941.7588  2556.8554  2781.16201  3117.36199   753.2029    36.29778
#> 1-6-1   1331.9807  -239.7591 -3865.77334 -2311.46589 -1303.1706 -1400.70691
#> 1-7-1  -5892.3745 -2836.2502 -8186.74856   -41.99052 -3059.4341   747.72093
#> 1-8-1  -4864.0681  -504.4013 -1357.08140   -14.59059  2756.4136  6446.63984
#> 1-9-1  -2178.6913 -4456.3503  7833.38063 -2665.94656   375.4047 -1701.40961
#>             0-15-0     0-16-0       0-2-0       0-3-0       0-4-0       0-5-0
#> 0-1-0   4063.55244  4939.6464  3102.83346  1099.86941 -5788.45364 -2797.19528
#> 0-10-0  1511.94948  1896.8581 -2974.03146 -2937.09408 -6653.64367  2611.94674
#> 0-11-0  3974.05905  6089.7792  2563.94707  1930.43117  2372.37373  1616.42921
#> 0-12-0  2076.23693 -1160.9331  1729.79713   -77.02285 -2179.95781  2229.40591
#> 0-13-0  2144.75254 -4700.2492   342.70125  6012.87078  2511.46094  2847.15739
#> 0-14-0  4583.29956 -4321.9655 -1116.38869  1461.75841   987.43433  1485.73303
#> 0-15-0 14198.56705  -874.6594  3006.20994 -2224.77192 -1666.95041  -587.06588
#> 0-16-0  -874.65941 21245.1761  1296.44488  -771.55072  3135.64769  -424.67987
#> 0-2-0   3006.20994  1296.4449 12788.48220  2645.39553  2868.71681  2239.43910
#> 0-3-0  -2224.77192  -771.5507  2645.39553 16248.85929  1917.80309 -4889.82618
#> 0-4-0  -1666.95041  3135.6477  2868.71681  1917.80309 15676.37356  1942.13920
#> 0-5-0   -587.06588  -424.6799  2239.43910 -4889.82618  1942.13920 17248.06193
#> 0-6-0  -1688.92900 10844.3815  1248.29582 -2422.99927  -360.45952  -509.09218
#> 0-7-0  -1594.46085 -2686.1258  2956.52125   812.34010  2774.82605  1879.03457
#> 0-8-0   6872.23966  1634.9741  4198.11506   618.92918  -456.48983  2221.83684
#> 0-9-0    734.75701 -3738.5938   580.48628  2050.45486  1811.89539  2423.92581
#> 1-1-1   -898.71877  1467.8301  3240.57015 -3108.97311  -577.23419  1591.57114
#> 1-10-1   281.95065  2255.0711  3982.67630    56.43244  1404.45694  -217.27364
#> 1-11-1 -2244.86023 -2703.5887 -5056.25411   690.14220  5027.98893  1093.43315
#> 1-12-1 -1931.12126 -4561.2276  2599.94860  3310.35962  4588.80634   855.42881
#> 1-13-1  4372.39277 -9699.9720  3655.32065  1862.85122 -5290.27994   352.62357
#> 1-14-1  2302.65806 -2083.9621  -210.59541   531.74546   -77.18507 -1984.56236
#> 1-15-1   -91.81219 -2027.6088  2871.76023 -1643.63993  -366.65963   992.96620
#> 1-16-1  -190.93635  2377.7653    94.48158  1863.54387   667.44779 -2002.56681
#> 1-2-1   -959.39666  -105.7109  -682.49272 -2645.55949  5250.20508  4709.35344
#> 1-3-1    877.07754 -1321.0331 -3604.91815   103.75348  1531.75875  2332.64752
#> 1-4-1   -178.21814  3249.7216  6279.88680   702.81421  2018.82076 -3095.15950
#> 1-5-1   -895.53929  3320.3582 -1186.47088  -483.43456  2246.79615   381.16111
#> 1-6-1   1835.28291 -2703.9985  2534.05863   670.85490  -323.10576  1214.88183
#> 1-7-1   1835.34139 -1488.6827 -1619.43329  -859.13849   775.98804   108.78925
#> 1-8-1   3180.17625 -2871.6679  4926.78180  5613.48027  5480.83561    50.94533
#> 1-9-1  -2129.13292  6263.6055  -428.88130  1415.68252   995.39429  2223.58340
#>              0-6-0       0-7-0      0-8-0       0-9-0       1-1-1      1-10-1
#> 0-1-0   3933.78819 -3712.54127  2754.2719   667.91333   549.38725   762.55261
#> 0-10-0  2941.57368  1017.40583  2606.0502  -991.30522  -529.62925 -2294.36412
#> 0-11-0  4015.31659  3051.33719  5113.9410   374.50896  2204.08015 -4652.33376
#> 0-12-0 -4381.68456 -4381.39200  3176.4714  2565.15689   754.60147  -469.20608
#> 0-13-0  -243.48994  6231.73778  5188.0836  5909.53177 -4050.54694 -1722.89358
#> 0-14-0 -6861.24190 -4215.54911  1925.4521  1873.81918 -1699.48716 -1869.07515
#> 0-15-0 -1688.92900 -1594.46085  6872.2397   734.75701  -898.71877   281.95065
#> 0-16-0 10844.38153 -2686.12582  1634.9741 -3738.59379  1467.83007  2255.07109
#> 0-2-0   1248.29582  2956.52125  4198.1151   580.48628  3240.57015  3982.67630
#> 0-3-0  -2422.99927   812.34010   618.9292  2050.45486 -3108.97311    56.43244
#> 0-4-0   -360.45952  2774.82605  -456.4898  1811.89539  -577.23419  1404.45694
#> 0-5-0   -509.09218  1879.03457  2221.8368  2423.92581  1591.57114  -217.27364
#> 0-6-0  24138.30053 -1682.21296 -4093.1534 -3593.62188   -43.78732  2661.91339
#> 0-7-0  -1682.21296 18429.11274  4747.1235  1503.67582 -1079.13267 -1747.71708
#> 0-8-0  -4093.15343  4747.12347 15945.2709  2470.33727 -2484.78442   173.18965
#> 0-9-0  -3593.62188  1503.67582  2470.3373 10518.51280 -2826.90587  1800.57856
#> 1-1-1    -43.78732 -1079.13267 -2484.7844 -2826.90587 11987.04900 -2074.06505
#> 1-10-1  2661.91339 -1747.71708   173.1897  1800.57856 -2074.06505 12360.03459
#> 1-11-1  -443.11474    20.57738 -6280.6874  3302.88676  1970.44362 -1039.52782
#> 1-12-1 -4247.18654  2430.22913 -2476.3528   721.59638  3379.93125 -4941.90615
#> 1-13-1 -4764.10662  1766.10030  3526.4041  1098.36481  -582.34391  1138.97659
#> 1-14-1  2811.10295 -4105.10516  2302.2142   192.79018    80.35214  3202.48853
#> 1-15-1   407.19112  -973.30733  2387.3436  1971.18146   909.75882  -692.90879
#> 1-16-1   137.26669 -5325.46301 -4381.3071  -462.84620  1975.64993  7697.17848
#> 1-2-1   4951.42575 -2772.11922 -1127.5328  3084.21341   741.87155   344.74890
#> 1-3-1  -6343.81045  -805.02824  3807.1304   395.97527  1969.63272 -4329.80852
#> 1-4-1   7873.92033  3765.06987 -4783.1797  -240.45079  4423.73458  3817.22035
#> 1-5-1    640.17311 -2869.86017  1108.0097  2717.13101 -1561.88136  1537.16685
#> 1-6-1  -3201.43963   760.52834  1468.8682    50.20652   363.07047  1234.72133
#> 1-7-1  -4695.75273    59.31458  1580.9265   271.57867   -99.60362   494.85546
#> 1-8-1  -2187.31947 -1087.28197 -1828.1026 -2220.75730   533.52907  2291.23951
#> 1-9-1  12290.35538 -4551.24113   602.3095 -1907.83627  4878.01467 -2438.54760
#>             1-11-1      1-12-1      1-13-1      1-14-1      1-15-1      1-16-1
#> 0-1-0  -5278.31357 -4248.66764   679.37946   987.79627  2797.25824  2101.65501
#> 0-10-0 -1197.14202 -5591.28790  2026.26297   914.16317  -589.20042 -4205.52305
#> 0-11-0  1806.69667 -1731.28724   -39.69338    12.96591  5905.76220 -5571.30876
#> 0-12-0 -2843.35543 -1424.13840  1171.12209    62.88799  2006.80421 -2971.74364
#> 0-13-0  4955.32088  3004.34569  4574.94101  3890.41720  -824.47024 -5171.08632
#> 0-14-0  1173.60759  1130.60354  2694.21601   104.28878 -2090.42652  -336.26854
#> 0-15-0 -2244.86023 -1931.12126  4372.39277  2302.65806   -91.81219  -190.93635
#> 0-16-0 -2703.58868 -4561.22755 -9699.97199 -2083.96213 -2027.60884  2377.76534
#> 0-2-0  -5056.25411  2599.94860  3655.32065  -210.59541  2871.76023    94.48158
#> 0-3-0    690.14220  3310.35962  1862.85122   531.74546 -1643.63993  1863.54387
#> 0-4-0   5027.98893  4588.80634 -5290.27994   -77.18507  -366.65963   667.44779
#> 0-5-0   1093.43315   855.42881   352.62357 -1984.56236   992.96620 -2002.56681
#> 0-6-0   -443.11474 -4247.18654 -4764.10662  2811.10295   407.19112   137.26669
#> 0-7-0     20.57738  2430.22913  1766.10030 -4105.10516  -973.30733 -5325.46301
#> 0-8-0  -6280.68742 -2476.35282  3526.40410  2302.21423  2387.34363 -4381.30712
#> 0-9-0   3302.88676   721.59638  1098.36481   192.79018  1971.18146  -462.84620
#> 1-1-1   1970.44362  3379.93125  -582.34391    80.35214   909.75882  1975.64993
#> 1-10-1 -1039.52782 -4941.90615  1138.97659  3202.48853  -692.90879  7697.17848
#> 1-11-1 18105.76763  5267.53383  2716.07195  2089.46692  -295.84900  5298.94072
#> 1-12-1  5267.53383 16010.67688  4229.14694 -2439.85966  2614.99025 -3231.70490
#> 1-13-1  2716.07195  4229.14694 20242.08215  7349.26233  6681.74421  -763.51958
#> 1-14-1  2089.46692 -2439.85966  7349.26233 15117.37193  3431.78880  1011.99598
#> 1-15-1  -295.84900  2614.99025  6681.74421  3431.78880 18306.25264 -3266.40150
#> 1-16-1  5298.94072 -3231.70490  -763.51958  1011.99598 -3266.40150 21069.22893
#> 1-2-1   4047.35427  3765.05918 -2456.21479  1175.87142  4468.32516   582.77216
#> 1-3-1   5026.06487  2972.66352  3549.06100  1258.82093   486.26919  1328.95430
#> 1-4-1   3740.88901  -452.85839 -1954.59690   761.49199 -3650.89111  9077.56624
#> 1-5-1   2208.84352 -3141.08598 -1003.12183  4689.07198  5585.16453  1163.03120
#> 1-6-1    -63.40157  2699.50442  4066.92936   622.12194  4610.63668  2809.09454
#> 1-7-1    774.22074   -47.00816 -4117.91855 -2550.91166 -3781.90627  4378.29120
#> 1-8-1   2561.87123  4828.86222  4189.37426 -1504.52630 -4577.88763  3856.00649
#> 1-9-1   7928.57138  3077.55684  -301.86221  4075.75556  3037.15820  2573.60585
#>             1-2-1      1-3-1      1-4-1       1-5-1       1-6-1       1-7-1
#> 0-1-0  -1798.8193 -5755.3975   926.0919   941.75884  1331.98074 -5892.37448
#> 0-10-0 -3608.8970  1075.7691  -874.3306  2556.85539  -239.75914 -2836.25020
#> 0-11-0  1999.9533 -4588.1731 -1502.2545  2781.16201 -3865.77334 -8186.74856
#> 0-12-0 -3419.0810  -345.5322 -1720.5944  3117.36199 -2311.46589   -41.99052
#> 0-13-0   330.2519   448.1511  1441.9144   753.20289 -1303.17056 -3059.43408
#> 0-14-0 -2343.4944  3989.8389 -3333.2510    36.29778 -1400.70691   747.72093
#> 0-15-0  -959.3967   877.0775  -178.2181  -895.53929  1835.28291  1835.34139
#> 0-16-0  -105.7109 -1321.0331  3249.7216  3320.35824 -2703.99852 -1488.68269
#> 0-2-0   -682.4927 -3604.9182  6279.8868 -1186.47088  2534.05863 -1619.43329
#> 0-3-0  -2645.5595   103.7535   702.8142  -483.43456   670.85490  -859.13849
#> 0-4-0   5250.2051  1531.7588  2018.8208  2246.79615  -323.10576   775.98804
#> 0-5-0   4709.3534  2332.6475 -3095.1595   381.16111  1214.88183   108.78925
#> 0-6-0   4951.4257 -6343.8104  7873.9203   640.17311 -3201.43963 -4695.75273
#> 0-7-0  -2772.1192  -805.0282  3765.0699 -2869.86017   760.52834    59.31458
#> 0-8-0  -1127.5328  3807.1304 -4783.1797  1108.00967  1468.86819  1580.92652
#> 0-9-0   3084.2134   395.9753  -240.4508  2717.13101    50.20652   271.57867
#> 1-1-1    741.8716  1969.6327  4423.7346 -1561.88136   363.07047   -99.60362
#> 1-10-1   344.7489 -4329.8085  3817.2204  1537.16685  1234.72133   494.85546
#> 1-11-1  4047.3543  5026.0649  3740.8890  2208.84352   -63.40157   774.22074
#> 1-12-1  3765.0592  2972.6635  -452.8584 -3141.08598  2699.50442   -47.00816
#> 1-13-1 -2456.2148  3549.0610 -1954.5969 -1003.12183  4066.92936 -4117.91855
#> 1-14-1  1175.8714  1258.8209   761.4920  4689.07198   622.12194 -2550.91166
#> 1-15-1  4468.3252   486.2692 -3650.8911  5585.16453  4610.63668 -3781.90627
#> 1-16-1   582.7722  1328.9543  9077.5662  1163.03120  2809.09454  4378.29120
#> 1-2-1  14949.9827  4727.2414 -5085.5614   754.35589  1432.20211  1243.73058
#> 1-3-1   4727.2414 20406.4213 -6012.7449   620.00957  2750.48624  8862.94586
#> 1-4-1  -5085.5614 -6012.7449 23301.3974  1026.50958  -235.37249  1139.84801
#> 1-5-1    754.3559   620.0096  1026.5096 13610.91179  1910.25963   527.98086
#> 1-6-1   1432.2021  2750.4862  -235.3725  1910.25963  7691.76871  2952.12536
#> 1-7-1   1243.7306  8862.9459  1139.8480   527.98086  2952.12536 14532.85674
#> 1-8-1  -1666.6508  2866.6698  4096.1743 -5004.88949   898.82928  1637.16193
#> 1-9-1   8850.9967  7439.7959   821.2518 -1307.47032 -2244.20611  3333.46797
#>              1-8-1      1-9-1
#> 0-1-0  -4864.06815 -2178.6913
#> 0-10-0  -504.40127 -4456.3503
#> 0-11-0 -1357.08140  7833.3806
#> 0-12-0   -14.59059 -2665.9466
#> 0-13-0  2756.41359   375.4047
#> 0-14-0  6446.63984 -1701.4096
#> 0-15-0  3180.17625 -2129.1329
#> 0-16-0 -2871.66788  6263.6055
#> 0-2-0   4926.78180  -428.8813
#> 0-3-0   5613.48027  1415.6825
#> 0-4-0   5480.83561   995.3943
#> 0-5-0     50.94533  2223.5834
#> 0-6-0  -2187.31947 12290.3554
#> 0-7-0  -1087.28197 -4551.2411
#> 0-8-0  -1828.10257   602.3095
#> 0-9-0  -2220.75730 -1907.8363
#> 1-1-1    533.52907  4878.0147
#> 1-10-1  2291.23951 -2438.5476
#> 1-11-1  2561.87123  7928.5714
#> 1-12-1  4828.86222  3077.5568
#> 1-13-1  4189.37426  -301.8622
#> 1-14-1 -1504.52630  4075.7556
#> 1-15-1 -4577.88763  3037.1582
#> 1-16-1  3856.00649  2573.6058
#> 1-2-1  -1666.65080  8850.9967
#> 1-3-1   2866.66980  7439.7959
#> 1-4-1   4096.17429   821.2518
#> 1-5-1  -5004.88949 -1307.4703
#> 1-6-1    898.82928 -2244.2061
#> 1-7-1   1637.16193  3333.4680
#> 1-8-1  20881.43068   134.7768
#> 1-9-1    134.77683 30395.3127
#> 
#> $stimcats
#>      stim cat
#> 1   0-1-0   0
#> 2  1-14-1   1
#> 3   0-2-0   0
#> 4   1-3-1   1
#> 5   1-4-1   1
#> 6  1-11-1   1
#> 7   0-8-0   0
#> 8   1-9-1   1
#> 9  1-16-1   1
#> 10 1-10-1   1
#> 11 0-14-0   0
#> 12  0-5-0   0
#> 13 1-15-1   1
#> 14  0-4-0   0
#> 15  1-5-1   1
#> 16  1-6-1   1
#> 17 1-13-1   1
#> 18 0-10-0   0
#> 19 0-11-0   0
#> 20 1-12-1   1
#> 21 0-16-0   0
#> 22  0-7-0   0
#> 23  1-2-1   1
#> 24  1-8-1   1
#> 25  0-6-0   0
#> 26 0-12-0   0
#> 27 0-13-0   0
#> 28  0-3-0   0
#> 29  0-9-0   0
#> 30  1-1-1   1
#> 31 0-15-0   0
#> 32  1-7-1   1
#>