Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

aat_stimulusscores(
  ds,
  subjvar,
  stimvar,
  pullvar,
  targetvar = NULL,
  rtvar,
  iters = 5
)

Arguments

ds
the data.frame to use

subjvar
Name of the subject-identifying variable

stimvar
Name of the stimulus-identifying variable

pullvar
Name of the movement-direction identifying variable

targetvar
Optional. Name of the stimulus-category identifying variable

rtvar
Name of the reaction-time identifying variable

iters
If there are missing values (which is almost inevitable) then multiple imputation will be used to complete the covariance matrix - this argument sets the number of multiple imputations to be used.

Value

Exports a list containing a data.frame with stimulus-specific bias scores, indicated in the column names, a covariance matrix of that same data, and a data.frame indicating to which stimulus category each stimulus belongs.

Examples

ds<-aat_simulate(biasfx_jitter=40,nstims=16)
ds$stim<-paste0(ds$stim,"-",ds$is_target)
aat_stimulusscores(ds,"subj","stim","is_pull","is_target","rt")

#> $data
#>            1-0         1-1        10-0       10-1        11-0        11-1
#> X1  222.783004 -196.207250   19.324794  300.65037  427.673181 -190.820207
#> 2   263.375934 -110.797255 -275.463494  157.89933  249.546295  -35.289747
#> 3    16.861680  289.738059  -79.040719   56.54952 -130.531714   75.761037
#> 4   -62.265167  147.028084   83.924315  -46.42705 -126.394243   -4.688545
#> 5   -24.306955 -112.122388  -28.233226 -144.70622   75.799981   81.805102
#> 6    67.016983  327.694066  103.245790  134.12248  -23.369313  -18.637994
#> 7   -79.686000    4.563078  166.615023 -195.22844   59.284821  -17.393912
#> 8    30.926012   31.571029 -179.244780  -43.22290  -64.212328   13.043797
#> 9   138.938194   62.579053 -256.227766   29.67064  -90.686298  105.080603
#> 10    1.591647  111.982883  -54.694866  -49.33154 -208.592169 -129.438990
#> 11   39.678470  157.489633   95.823135  137.42169    5.059891   77.998416
#> 12 -138.643232 -139.464795  134.688335  198.98507   42.560742  284.059549
#> 13 -123.216263  115.855028  -13.426021 -104.13172  -69.068489  -99.069559
#> 14 -175.997486  -38.804561  260.412393  119.53663  109.872117 -288.597213
#> 15 -117.664786    1.048111   35.760634  350.40775 -253.705607   77.790516
#> 16  -21.851041   15.557048    3.000805  -69.81134  -51.643017   39.859794
#> 17  -42.974812   85.660532  -56.421346 -108.46729   57.909855   54.213845
#> 18   -4.721596  -36.532070   33.838779  172.30829  -39.702307 -244.017839
#> 19 -203.459875   21.014245  -66.260611 -188.29114 -177.338236  255.592887
#> 20   23.676960  110.795404 -129.976514  -21.07666  -43.670642   -6.370075
#> 21  106.522277  -31.225607  -89.431261  -54.63155  193.661780  177.715756
#> 22  -75.160920  242.044040  101.137859   61.42006   31.570352  -95.094573
#> 23  411.567292 -422.700177  118.793237  -40.56607  -95.056749 -273.894351
#> 24  -34.546517  231.548335 -114.433606   47.92381   84.560861  371.942258
#> 25   46.176271   64.772620 -146.284761  236.64149  214.257069   10.362453
#> 26  -80.313208 -207.285608   36.498727   76.04823  -21.709569 -106.002488
#> 27  110.881140  -24.850767 -154.712865 -278.00643   40.929708  274.671613
#> 28  108.897706  134.928709  477.737662   49.65273   61.013639 -192.422487
#> 29   83.950010  -48.840829  -73.620883  -72.40411  215.956927  103.001102
#> 30   35.876961  143.241988  -49.557148  145.66165  604.191318  376.804809
#> 31  174.863100  405.395187 -221.111353  201.84160 -241.893055  332.273176
#> 32 -127.696330 -209.567166 -317.741568  -90.87736 -233.135614    1.946613
#> 33   -8.793893  -92.906070 -226.650843  -21.82817   15.533288  -59.194307
#> 34 -173.393178  -75.098948   19.819057   55.72788  105.666187  426.823647
#> 35 -195.742073  626.557851 -191.077919  325.99919 -424.319744   69.716367
#> 36   -4.365512   84.702089   15.190491   75.45700  214.109305   51.714120
#> 37   93.203200   12.659898 -202.187239   81.56282 -162.102224   17.445484
#> 38  204.419993  -46.911482  278.471908  -73.70020  169.235432  -52.969310
#> 39  -88.519161    4.029608  -22.335117  163.55196 -145.445539 -140.785723
#> 40  276.570470   51.761262  190.874612  217.91855  -36.698294  -21.987506
#>            12-0       12-1         13-0        13-1        14-0        14-1
#> X1  -36.2465667  149.86871 -178.7932904 -121.453928 -130.938636  172.337871
#> 2    84.2921381  -13.01829  -14.4127053  229.539059   13.911041  116.480144
#> 3   142.8626608  262.12651  -66.0719008   -7.691965   50.974045  140.178728
#> 4   -63.2199904   69.96409  -94.3693759   18.879997  -24.114047  -37.918388
#> 5   -33.6968492  -46.98501  157.0730838   11.777061  -56.139519  198.302932
#> 6    22.2983115  297.40950 -123.7434955  215.591484  -22.239276  155.597411
#> 7     0.6412893   28.45559   33.3555058    3.123998   -4.438222  -48.179421
#> 8   -91.1225304  -14.37701 -355.2754366   15.776576  -10.686718   21.506070
#> 9  -103.7513456   49.61391  -19.7770209   90.327370   -9.929436  119.315337
#> 10  -43.1433310 -125.50177  107.0174775  122.540026  128.386542  311.669417
#> 11 -194.4152597   71.78816  -12.7800786   64.519017  -21.484398  223.543177
#> 12  112.6996373  159.23203  -12.7496983    2.589519 -312.743770   81.425911
#> 13   26.4604931  316.45537  -17.2507428 -163.398745  143.526835 -220.556907
#> 14   34.5435480  -88.83686  326.1742313   69.928191   26.072394  -82.333357
#> 15  -68.6108998 -154.79882  540.8333676  259.808193  -30.770930   97.795003
#> 16    3.0196871  -90.40492 -235.9304050  130.652216   37.246778  -40.206522
#> 17  -76.1806548 -129.84706   -2.3536183  292.164276  -65.109523  166.641299
#> 18   76.0023257   50.09445  -56.0993195   75.915383  -24.671505  103.025354
#> 19   23.5241438  312.40518  -78.4111783  101.746354  162.622981  -82.436787
#> 20   15.9152588  167.21722 -472.7861644  324.960539   -5.803797 -281.536942
#> 21   51.3201576 -200.89389    0.7791266  313.473077    3.368978  101.128622
#> 22 -130.5620355   39.57522   49.7849071   50.443627  181.126018  297.419301
#> 23 -591.1293552  247.07118  276.4485017 -247.966297 -592.679894  119.568176
#> 24   35.1741745  117.29325  146.2992160  -33.866113  122.264963  -17.682071
#> 25  -97.2876498  278.09079  -46.0347190  270.298865  191.914364  -11.856264
#> 26 -307.6824664  -12.08850  -11.0619747 -129.608112   65.806103   14.083077
#> 27  105.8945857  157.34505   56.2728757   73.192154  253.403604  235.389246
#> 28 -116.0297670  232.09638  212.8472316   78.013734  355.452646   18.255803
#> 29 -159.4261734   50.21903    5.9359154   39.084194    6.392275   -1.400198
#> 30   97.5201661  123.64683 -129.9524556  -24.961689   93.275848  165.180477
#> 31    3.6914200 -229.13191   99.2015149   85.197190  327.406333  183.929939
#> 32  -67.8459356   15.78653   41.4678117   -4.473400  126.881283   17.788592
#> 33  -20.8011782  -64.38919 -175.4959569 -172.956130  -56.835223 -183.450424
#> 34  100.8192559  138.06464   47.4406327 -146.750041  -78.428816 -117.410149
#> 35  299.0357413  144.82497 -173.9496041   25.640338  -85.647258  157.657123
#> 36   91.8956209   34.40484   95.3852368  184.130027  159.897391  140.452691
#> 37    5.5023878 -163.26558 -323.4351758  -64.727873 -136.440377   97.716066
#> 38  122.9545756   -2.32100   97.6876265  -35.356752   22.246891 -185.511628
#> 39 -258.2301371 -148.12344 -101.3127784   -5.624821  -72.773785 -144.892026
#> 40   81.9503038   18.99826  -24.3669552  142.287417 -111.413467  281.536420
#>           15-0       15-1        16-0        16-1          2-0         2-1
#> X1 -122.212118 -170.22376 -361.760460   89.349067   63.7256491  -99.571347
#> 2   413.623998  302.53414  819.554860  149.832653  -14.1922404  224.285776
#> 3   117.090673  114.28465   52.378934  139.272224   78.1427284  -31.077277
#> 4   -28.191411   94.14885  126.187282  -82.774598   57.8467203 -134.888528
#> 5   -51.637950   16.08082   86.404391   18.535384  307.8799891  -52.588918
#> 6   -92.670043 -205.28347 -229.855772   40.858593 -126.9256748  124.193787
#> 7    99.582007  109.70871   12.723615  -46.353984   62.5018089 -101.043364
#> 8   -53.454834   25.03398   87.772721 -468.007082   52.9002022  -74.307848
#> 9   -93.380495   76.95230  -59.787074  -41.719663  228.7673281   85.179343
#> 10  112.037707  -60.65792  226.138472   19.118220  141.2495760 -109.072103
#> 11   17.637728  -24.11352   15.859650  142.137609   83.4203506 -127.956848
#> 12  -24.946227 -183.93664  135.003901   75.356998 -138.7554355   65.483092
#> 13   83.332723  -94.38960 -181.997918  -53.382021  129.2398901   79.574294
#> 14  102.727691   33.82533 -194.977402 -209.797841   -3.6830240  243.351385
#> 15   -9.602665  177.38989 -354.598195  -46.746831  250.2490684  314.822966
#> 16  -83.618595   42.16712  -84.279049   57.558292 -165.8472564  -93.824839
#> 17  -16.689497  -44.12917 -282.541983  -79.498113  117.1776627  -70.558204
#> 18 -110.792413   43.33480 -145.991920 -109.751057 -154.2290029   32.519761
#> 19  -99.670802 -179.47081    3.677818 -162.330112  -34.8570507  197.897145
#> 20  -34.416856  204.72780   15.234511  276.057452   30.5301454  352.037401
#> 21  180.050202  -30.19004   26.983778  226.976236  187.4210458   17.587152
#> 22   17.719781  -63.74173  -58.260452   63.227364   52.4885066    1.651787
#> 23 -116.544448  119.72298  119.868908  230.092257  271.8055970  355.638901
#> 24  120.542392   45.31688   20.747666    5.093877   24.7838427   69.938070
#> 25  205.652523   61.22006   -7.412231  105.395110   46.2202121   22.068442
#> 26 -244.203310   28.29959  162.657810   60.371017   43.4477187  -97.398217
#> 27   24.564033  -53.24709  -54.713724  168.805056  -81.0063406   77.651199
#> 28  136.934832  297.32541  -37.602183   61.201716  407.3150094  -15.107617
#> 29   65.065797   69.56390   73.424117    8.185990   46.4086892  111.968394
#> 30  251.593074  104.30928  205.734097    8.442342  -83.0094980  445.786640
#> 31   88.113859  415.93055  -75.355814 -144.682178  278.7113302  253.103386
#> 32  -31.274879  -44.86852 -149.319656  -54.808141 -105.8704867   69.459714
#> 33  -27.915880 -138.28414  -96.587963 -103.710103  -85.5160012  -50.385456
#> 34   24.186223  -54.71827   69.007152   62.971385  -53.2131809   49.872412
#> 35   15.156328  303.08988  659.290456 -145.893550  304.3796698 -144.322006
#> 36  103.647013   56.20436  100.192683  124.099405   75.0564803  221.073321
#> 37 -182.684988 -318.42334   -1.884933  -56.421401   -0.2414084 -139.545189
#> 38  123.994483   40.68156  170.318578  -41.449857  160.2039105  227.983093
#> 39 -209.579377   89.96486  -48.027811 -215.046456   86.4724516  205.928517
#> 40  -66.000439  150.15383 -114.552191  105.614962   -6.7815281 -128.395927
#>            3-0         3-1         4-0         4-1        5-0         5-1
#> X1   32.048558  -20.435185    1.919814 -231.286587  107.73345  302.586754
#> 2   103.995258  519.717997   27.822090  323.165940  370.71025   -4.900436
#> 3   -14.295698   65.808966    4.889316  167.994437 -162.43161  133.205062
#> 4   -29.251874   -8.592681   41.761909   67.803046   52.03091  137.553184
#> 5  -117.500046  246.405202   71.569902   74.761733  151.88360   21.602321
#> 6    64.915484   13.023950  122.742676   66.013460   72.07215  328.653417
#> 7    84.656421  127.250514    2.313098  -30.960423  -45.26625   56.277372
#> 8  -212.381251 -177.517472   -1.116991  -74.078981   99.18331  102.980343
#> 9   218.223013  -32.296263   -9.400233  -62.762596  -63.36493   49.060610
#> 10   28.708437 -319.013355  154.948035  116.528268 -122.16892 -223.723167
#> 11  130.860520   95.699391 -103.356647  142.038209   58.82504  164.629973
#> 12 -297.295370  -29.557545 -122.190385   -0.824127  -90.05195   67.355395
#> 13  153.076083   -5.773116   19.121145  -67.992537  169.89959  248.834578
#> 14   54.086846 -259.002397  -27.727233 -187.874525  -39.97319  -79.890740
#> 15   36.696553  160.873715  372.463132   23.005486   21.33715  -61.091853
#> 16  -79.985787  -35.780295  -75.981488 -193.676924 -152.30776 -159.236668
#> 17  -71.878787 -206.527106 -290.363896   35.527742 -287.58165  -34.787773
#> 18  -64.335374  134.335924  -50.558942   13.461807 -193.55734   68.038069
#> 19    8.124697 -138.457810  124.723460   40.018230 -143.68573 -202.999408
#> 20  113.425909  216.969451  172.728882 -147.790826  282.99275 -286.960516
#> 21  -79.833018  128.340929  -24.688994   99.915962   97.27660   74.243472
#> 22   72.659774   82.315512    4.180755    6.605117   -9.47402  205.437605
#> 23   88.731862  854.801763 -723.182685  -37.348256  356.66761 -326.927555
#> 24  226.335927  312.972382  -83.562209 -135.137625  130.93325  -15.201147
#> 25 -157.229276   40.740701  -32.956769  117.950939  -93.81830  -11.142318
#> 26  -52.036897  104.087749   -5.887059   -8.828471   21.38429  -80.433263
#> 27  184.917841   50.647363 -320.023424  383.247247 -123.08987 -238.688853
#> 28  233.669240  112.024838    3.979256   28.567320  124.21248  129.423832
#> 29  210.955624  -79.854084  150.805048  -14.432987  219.25462   98.033106
#> 30  389.810473   44.225805  -19.199151  390.642476  -92.59582   69.576802
#> 31  147.578985  362.303895 -192.285742  285.093511 -200.27284  207.714465
#> 32   31.573927 -156.053120   30.411361 -226.740361 -175.97666 -107.962590
#> 33  -45.877587 -279.789009   -9.308810  189.328699   42.12924 -300.554809
#> 34  134.249977  -29.401804  -32.602188  151.761404  187.67739  148.290494
#> 35  117.559038   -4.551052 -412.917483  361.602230 -119.49030   26.119706
#> 36   35.347571  153.353443   70.902915   -6.678037  131.37167   86.571483
#> 37 -317.078529  136.655830 -374.275903 -151.739599 -344.47088  309.488556
#> 38  -65.184587  396.587894    6.440572  444.451159  -12.07743   14.593668
#> 39   -3.444930 -169.328113  -89.170376   54.878388  -30.65589 -203.743225
#> 40   91.698899  162.946470  -13.619850  -86.574927  -65.98508  -71.257539
#>            6-0        6-1          7-0         7-1         8-0         8-1
#> X1   92.286311  330.90589 -295.0329706  -47.923269   -6.165806 -149.120489
#> 2   -91.813123   22.81281  296.8910856   26.755141   50.926456  129.482229
#> 3   -42.648790   31.20927    5.6082528   89.719375  276.477250  -23.328596
#> 4   141.337467  166.31821  -24.0449125  121.842019  -23.517720  -69.527808
#> 5   -86.511340   36.61452  -96.9190656  -15.425982   -1.505668    6.399858
#> 6   216.137399  124.77596  141.9098225  -16.856946  -46.548212  202.347685
#> 7  -116.715464   24.70459  116.2496040   49.543336 -125.398206   63.861101
#> 8   -63.888254  -28.17394 -228.0356740   58.037446   68.804499  292.392093
#> 9    83.846532  131.44786   62.1063667  105.001119  -37.001299   85.015941
#> 10    8.216006  -27.71380   -0.4495214 -156.024136 -232.230526 -205.717100
#> 11  -38.636482  163.82035  -16.4082192  -48.201567    1.854557   43.216776
#> 12 -258.560618  121.29501  199.7757777  -50.099739 -214.199222   60.708486
#> 13  158.320255  -32.80852  -69.3375344  392.069854 -183.753858  191.657810
#> 14  -82.787464 -228.42316  -66.5983939   86.722169   68.997305   18.516344
#> 15   44.947281  -70.20712  157.2064329   96.559690  184.947078  130.691423
#> 16   81.444738  -13.47280  -29.3234902  -77.643513   28.951876   73.884935
#> 17  -38.934862  -82.58937 -119.0683821 -110.371645  -22.616235  217.946877
#> 18   51.827212   80.64821   -4.9829128   39.575683 -202.984371   65.695617
#> 19  168.856416  -16.97828  -70.2421737  -78.472760  -85.867991  402.865988
#> 20  162.070951  146.12329  -25.9591020  -74.438306 -219.633679 -204.396103
#> 21  184.051584  214.65453  349.0992800  -37.640879  -24.861434   20.967872
#> 22   41.051247   87.26501  116.5417035  157.645090  216.682772  -26.401762
#> 23 -306.468418  425.50614 -165.4640133  199.308181  -73.232675   54.224562
#> 24   23.475284   35.63191    1.1724676   20.699727   54.725528  110.123855
#> 25   73.344596  161.22491 -294.6567607  134.295899   48.457649   32.737958
#> 26 -163.261511   38.34310  176.6896764  -61.528896 -300.530427   42.318491
#> 27 -236.698458 -152.68610  383.9414731 -305.485406  135.158753   72.599991
#> 28  108.569985  -16.93277  123.0998215  148.926475  128.224013 -114.181477
#> 29   67.910179  206.88385  155.9456584  245.494298  -72.714093  251.959568
#> 30  -24.681789  364.10972  142.6183030  268.958758  -88.616052  175.975670
#> 31  569.005994  200.27852  284.1922046  -57.435058  266.599247   26.327669
#> 32  122.298772  -52.17573 -162.0183834  -58.491308  -53.478940  156.436307
#> 33   36.253422  -21.27541  155.6240337 -127.870244   61.165175  -56.280022
#> 34  276.048926   99.08397  144.2631846  142.210225  -10.532340   66.209617
#> 35  181.795773 -341.70216   36.7673122 -197.234318  353.879806  605.150707
#> 36   95.626442  120.33498   47.6179385  122.489595  131.222359  124.439555
#> 37 -352.211684 -213.72903 -157.1313830   87.506215 -193.952170  218.069649
#> 38 -202.219393  106.43756  331.8753816  119.443954   70.931095  -91.245025
#> 39  -25.167212  -13.57316  157.3146048   -3.703657  -86.211008  -35.252758
#> 40 -124.642981 -101.11442   85.4876299  -59.597749  345.750797   89.595809
#>            9-0         9-1
#> X1  -17.840673 -321.905215
#> 2   -55.520586  141.442912
#> 3  -128.841713  199.798192
#> 4    40.414408    7.443898
#> 5   112.927093 -123.072295
#> 6    15.334486  -71.633940
#> 7   128.036478  121.723604
#> 8    21.272886   20.521904
#> 9    91.647938  -39.705677
#> 10 -190.861311  122.905022
#> 11  138.355935  147.040189
#> 12 -156.614204   49.776608
#> 13   -6.722106   83.749342
#> 14  278.175505  -73.542279
#> 15  -10.052974  218.217425
#> 16  -45.251106   -8.091693
#> 17  -50.326558  265.680401
#> 18  159.036910  -43.441682
#> 19   52.523500   32.553074
#> 20   55.160766 -148.001104
#> 21   83.261264   57.652172
#> 22   -0.815422  176.131446
#> 23  160.686913  -77.786107
#> 24  193.969205  -31.453018
#> 25 -143.014786  111.069107
#> 26  -20.703920   21.669366
#> 27  460.894338   32.408710
#> 28  -43.010425  143.568024
#> 29   45.981170    2.906390
#> 30  146.500211  283.419191
#> 31 -160.055238  100.978277
#> 32  206.835446   23.342961
#> 33  110.521916  -88.079772
#> 34   -2.053721  -44.627919
#> 35  247.219705   58.003474
#> 36   23.411006   98.682599
#> 37  100.133821 -136.280015
#> 38   98.257497   37.672997
#> 39   62.529881 -113.003474
#> 40  288.421144  -80.638495
#> 
#> $covmat
#>             1-0        1-1         10-0         10-1          11-0        11-1
#> 1-0  19163.9391  -6047.030    622.23605  1177.307165   7622.825899  -4796.6702
#> 1-1  -6047.0301  33620.096  -2763.44376  7032.085613 -10402.951718   8843.8988
#> 10-0   622.2360  -2763.444  27147.16075  1352.081173   5973.567166 -10457.6909
#> 10-1  1177.3072   7032.086   1352.08117 21554.071264    306.211911  -2099.2994
#> 11-0  7622.8259 -10402.952   5973.56717   306.211911  36171.989449   3647.4181
#> 11-1 -4796.6702   8843.899 -10457.69088 -2099.299407   3647.418055  31002.2206
#> 12-0 -5659.9541  12884.791  -3369.22703  2520.324194   2044.080761  10682.7284
#> 12-1  -229.3365   2466.301   4488.52887 -1113.226556   3446.057532    827.1243
#> 13-0  -183.3064  -6075.210  11880.15058  1688.669205   -642.371207  -1750.6319
#> 13-1   199.0426   7647.671  -2047.52773  2200.089218    289.803458   2364.2369
#> 14-0 -4887.0707  13347.286  -2229.27817 -3614.472987   1517.562233   5920.5014
#> 14-1  6315.2523   5233.888  -1033.04373  5225.021955    206.810063   2070.8325
#> 15-0  2435.8114   5386.430   -662.82257   286.738979  11014.612779   5328.1682
#> 15-1  4682.0942   8514.301    720.52704  5809.980334  -3306.778755   -346.9312
#> 16-0  2732.9543   5547.159  -7896.16857  1482.809311    802.802497   5372.9676
#> 16-1  7965.0412  -4730.763   2817.67321  -214.764484   8666.625963   1651.8282
#> 2-0   4117.0601   5226.604   3655.30857  1532.745408  -6006.036716  -3625.1946
#> 2-1   3129.8053  -3891.297   -287.77739  1955.032012   6757.507200   3897.7245
#> 3-0   2684.5837   7216.531   1403.89513    -5.426017   5630.922192   4838.6136
#> 3-1  19028.4707  -6226.029   4536.93008  3141.177716   4246.677492   -367.6176
#> 4-0  -7691.0885   1787.625   2452.97678   898.082087   5166.581833    945.0541
#> 4-1   3116.9189   9984.577  -2637.59709   920.188986   3683.873986  10423.3480
#> 5-0   6480.4771  -8655.702   4015.18445 -1013.494006   9990.086937  -4106.2351
#> 5-1    437.0549  11198.615   3835.43746  7583.091309   6549.477815   3454.7792
#> 6-0  -5306.4164  15785.566  -5891.63089  4070.702193  -4919.087376   7854.9838
#> 6-1   9833.0001  -7702.559   2456.64332  1012.894647  15209.904307   2590.2994
#> 7-0   2209.4377   3567.900   2938.92538  -887.020187   2934.147409   9755.9127
#> 7-1   1303.6680  -2192.554   6130.18322  1132.173821   8248.613237  -2890.3871
#> 8-0   3339.5334  13710.817    600.20529  6887.744935  -3736.505879   3941.6367
#> 8-1  -6071.3678   8032.936  -8195.40106  1154.191813  -7000.862649   8477.5907
#> 9-0    811.5313  -2334.234     74.69304 -4040.629176      7.188923   -103.1231
#> 9-1  -3165.0316   7953.409    842.59238  -467.362221    615.990535   5773.9386
#>            12-0       12-1        13-0       13-1        14-0        14-1
#> 1-0  -5659.9541  -229.3365  -183.30645   199.0426 -4887.07075  6315.25232
#> 1-1  12884.7909  2466.3006 -6075.21013  7647.6714 13347.28619  5233.88763
#> 10-0 -3369.2270  4488.5289 11880.15058 -2047.5277 -2229.27817 -1033.04373
#> 10-1  2520.3242 -1113.2266  1688.66921  2200.0892 -3614.47299  5225.02195
#> 11-0  2044.0808  3446.0575  -642.37121   289.8035  1517.56223   206.81006
#> 11-1 10682.7284   827.1243 -1750.63190  2364.2369  5920.50144  2070.83251
#> 12-0 21522.4575   982.3783 -5586.63774  5110.5466  7231.79261   184.57312
#> 12-1   982.3783 21912.5874 -3299.48835 -4023.9854  -132.43313 -3099.75773
#> 13-0 -5586.6377 -3299.4883 32945.05143  -234.0975  2015.00011  5642.18994
#> 13-1  5110.5466 -4023.9854  -234.09750 18910.1974  6931.62339  3889.97729
#> 14-0  7231.7926  -132.4331  2015.00011  6931.6234 26318.28390   -91.46037
#> 14-1   184.5731 -3099.7577  5642.18994  3889.9773   -91.46037 21192.17923
#> 15-0  7425.4407  1199.2547  5809.20967  5453.9843  8370.74765  1067.13646
#> 15-1   225.8113 -2418.9115  6248.33094  5000.8400  5179.39791   519.83858
#> 16-0  6617.5419  1646.2739 -5922.08909 -1085.0947 -2748.46303  2395.46142
#> 16-1 -1381.4168  5916.0839  2206.81105  3695.0244 -2506.13377  4419.98331
#> 2-0  -5116.2751 -2403.0884 10286.03610   426.4805  1111.80299  3060.92652
#> 2-1  -1875.2600  2382.9543  8111.73772  2765.4169    69.70221 -5560.66979
#> 3-0    181.9862  5795.9979  6530.79409  -890.1846  7552.46207  1478.06587
#> 3-1  -6869.6045  3828.1476 10431.38137 -2078.8768 -9124.10434  3903.27929
#> 4-0   5816.6579 -1142.6105  2137.71831  9201.8358 12492.51321 -5375.77316
#> 4-1   8303.2638   556.8930  2686.58996   920.7028  5670.10579  4279.96790
#> 5-0  -7003.8936  6579.4713  3763.85314 -2390.2566 -5877.59894 -6036.53782
#> 5-1   6124.6476  2650.2364 -3042.16441 -1506.4115  3995.01858  5801.74978
#> 6-0   5571.9288   785.3576 -1932.98006  6165.8060 13143.90138 -3021.14545
#> 6-1  -8837.4716  5076.6485   -65.93226 -2333.0337 -5116.87396  1006.65338
#> 7-0   6514.5467 -5846.1018  7058.07232  3063.6130  5923.51648   903.07075
#> 7-1  -4296.4734  5139.0140  4904.85114 -4832.6429  -663.14087 -5048.51394
#> 8-0   7224.4602  -948.6867  6724.36877  3600.9188  5910.39200  7512.92709
#> 8-1   5378.7572  3312.8098 -3362.43024   251.8681 -1703.57295  2018.83624
#> 9-0   2075.5821   584.9132  1410.51136 -3058.2187 -2523.29170   354.33770
#> 9-1   2173.5072  -419.1209  7008.18867  5319.2480  7928.11065  4679.98411
#>            15-0       15-1       16-0        16-1         2-0         2-1
#> 1-0   2435.8114  4682.0942  2732.9543  7965.04118  4117.06013  3129.80526
#> 1-1   5386.4303  8514.3013  5547.1588 -4730.76292  5226.60358 -3891.29711
#> 10-0  -662.8226   720.5270 -7896.1686  2817.67321  3655.30857  -287.77739
#> 10-1   286.7390  5809.9803  1482.8093  -214.76448  1532.74541  1955.03201
#> 11-0 11014.6128 -3306.7788   802.8025  8666.62596 -6006.03672  6757.50720
#> 11-1  5328.1682  -346.9312  5372.9676  1651.82818 -3625.19457  3897.72452
#> 12-0  7425.4407   225.8113  6617.5419 -1381.41684 -5116.27507 -1875.26003
#> 12-1  1199.2547 -2418.9115  1646.2739  5916.08389 -2403.08837  2382.95432
#> 13-0  5809.2097  6248.3309 -5922.0891  2206.81105 10286.03610  8111.73772
#> 13-1  5453.9843  5000.8400 -1085.0947  3695.02439   426.48053  2765.41686
#> 14-0  8370.7477  5179.3979 -2748.4630 -2506.13377  1111.80299    69.70221
#> 14-1  1067.1365   519.8386  2395.4614  4419.98331  3060.92652 -5560.66979
#> 15-0 16760.7060  7630.0277 12191.9972  4214.42186  2001.87865  5455.39619
#> 15-1  7630.0277 22447.9776 12404.2946   683.05579 10377.35077  6766.73915
#> 16-0 12191.9972 12404.2946 48362.8792  3323.80901  3552.92646   738.58007
#> 16-1  4214.4219   683.0558  3323.8090 19836.75379   -72.60658  2410.62708
#> 2-0   2001.8786 10377.3508  3552.9265   -72.60658 19081.66264   333.21529
#> 2-1   5455.3962  6766.7392   738.5801  2410.62708   333.21529 25316.69560
#> 3-0   6591.1348  9576.7339  2120.5591  4116.26895  3610.88857  8818.71958
#> 3-1   5400.5706 13940.7445 14750.7616 14959.04301  9958.21120 12969.49627
#> 4-0   4285.9159  -328.3050 -7594.7195 -1588.81762 -3238.37565  2667.48873
#> 4-1  10481.0865  7820.7343 19742.0992  1372.21650  3804.71700  5172.85458
#> 5-0   4860.5079  5462.8578  9488.5403  8177.40769  4705.63315  7920.08959
#> 5-1   2430.6974 -3791.5263 -3908.5440 -2023.59911  1829.22255 -6936.01145
#> 6-0   3276.4780  7499.1027 -5724.3191 -2989.45895  3446.34858  2775.03595
#> 6-1   2207.3472   905.5923 -1747.5901  8934.52659   562.56745  8840.89624
#> 7-0   6525.2364  6549.2436  9790.2574  4760.09643   176.68689  6935.66288
#> 7-1   4377.6592  1396.1627 -1526.9075  -532.43348  2878.18581  7043.65900
#> 8-0   5716.8863 10884.1660  1828.3795 -1072.19068  6578.46225 -1238.65478
#> 8-1  -1085.6566  -800.9554  8166.5153 -8490.76013 -1381.84893  -583.82883
#> 9-0  -2514.1015  -615.3500 -1024.6514  -752.06097 -2549.22787  1037.48340
#> 9-1   9080.7330  5622.6708  5746.0260   321.21719  2746.58492  2253.49174
#>               3-0         3-1         4-0        4-1        5-0        5-1
#> 1-0   2684.583666  19028.4707  -7691.0885  3116.9189  6480.4771   437.0549
#> 1-1   7216.531244  -6226.0295   1787.6247  9984.5771 -8655.7019 11198.6153
#> 10-0  1403.895130   4536.9301   2452.9768 -2637.5971  4015.1845  3835.4375
#> 10-1    -5.426017   3141.1777    898.0821   920.1890 -1013.4940  7583.0913
#> 11-0  5630.922192   4246.6775   5166.5818  3683.8740  9990.0869  6549.4778
#> 11-1  4838.613646   -367.6176    945.0541 10423.3480 -4106.2351  3454.7792
#> 12-0   181.986173  -6869.6045   5816.6579  8303.2638 -7003.8936  6124.6476
#> 12-1  5795.997868   3828.1476  -1142.6105   556.8930  6579.4713  2650.2364
#> 13-0  6530.794092  10431.3814   2137.7183  2686.5900  3763.8531 -3042.1644
#> 13-1  -890.184620  -2078.8768   9201.8358   920.7028 -2390.2566 -1506.4115
#> 14-0  7552.462067  -9124.1043  12492.5132  5670.1058 -5877.5989  3995.0186
#> 14-1  1478.065874   3903.2793  -5375.7732  4279.9679 -6036.5378  5801.7498
#> 15-0  6591.134800   5400.5706   4285.9159 10481.0865  4860.5079  2430.6974
#> 15-1  9576.733900  13940.7445   -328.3050  7820.7343  5462.8578 -3791.5263
#> 16-0  2120.559069  14750.7616  -7594.7195 19742.0992  9488.5403 -3908.5440
#> 16-1  4116.268945  14959.0430  -1588.8176  1372.2165  8177.4077 -2023.5991
#> 2-0   3610.888574   9958.2112  -3238.3756  3804.7170  4705.6331  1829.2226
#> 2-1   8818.719579  12969.4963   2667.4887  5172.8546  7920.0896 -6936.0115
#> 3-0  20514.321244   4816.7688   2049.7715  5524.7853  7281.0816  -633.6892
#> 3-1   4816.768779  49480.7889 -13974.3869  7130.5586 15671.1632   616.1173
#> 4-0   2049.771467 -13974.3869  33435.1865 -4071.9437  5013.9936  3179.2130
#> 4-1   5524.785273   7130.5586  -4071.9437 29568.5189 -1298.8883  -715.5308
#> 5-0   7281.081619  15671.1632   5013.9936 -1298.8883 26602.1703 -1502.0720
#> 5-1   -633.689152    616.1173   3179.2130  -715.5308 -1502.0720 28682.0821
#> 6-0   8410.367154  -7235.3281  10638.6083  1412.3901  1220.3936  7157.5704
#> 6-1   4610.524004  13696.0912    947.2321  1016.7736 10639.6054  3934.4794
#> 7-0   6930.879966   6923.4559   3570.0124 17211.7397  2378.7351 -1433.6210
#> 7-1   4008.191209   6827.8932   3495.6879 -2896.6729  7518.0493  9340.7072
#> 8-0   4893.098916   4647.5336  -2379.2196  8325.9714 -1883.3327  2709.3207
#> 8-1    480.993831  -3335.8919  -6736.2957  2548.7842 -4997.2916  3621.1965
#> 9-0   5591.241496   2276.0979  -7484.1609   506.9447  -290.1018 -6042.9260
#> 9-1   3715.813629   -277.6833   2315.2284  9657.1757 -4358.0701  1537.7183
#>             6-0         6-1         7-0         7-1        8-0         8-1
#> 1-0  -5306.4164  9833.00010  2209.43772  1303.66798  3339.5334 -6071.36775
#> 1-1  15785.5663 -7702.55882  3567.89952 -2192.55438 13710.8170  8032.93613
#> 10-0 -5891.6309  2456.64332  2938.92538  6130.18322   600.2053 -8195.40106
#> 10-1  4070.7022  1012.89465  -887.02019  1132.17382  6887.7449  1154.19181
#> 11-0 -4919.0874 15209.90431  2934.14741  8248.61324 -3736.5059 -7000.86265
#> 11-1  7854.9838  2590.29944  9755.91272 -2890.38705  3941.6367  8477.59070
#> 12-0  5571.9288 -8837.47158  6514.54668 -4296.47338  7224.4602  5378.75720
#> 12-1   785.3576  5076.64852 -5846.10182  5139.01396  -948.6867  3312.80982
#> 13-0 -1932.9801   -65.93226  7058.07232  4904.85114  6724.3688 -3362.43024
#> 13-1  6165.8060 -2333.03370  3063.61300 -4832.64292  3600.9188   251.86809
#> 14-0 13143.9014 -5116.87396  5923.51648  -663.14087  5910.3920 -1703.57295
#> 14-1 -3021.1455  1006.65338   903.07075 -5048.51394  7512.9271  2018.83624
#> 15-0  3276.4780  2207.34721  6525.23641  4377.65915  5716.8863 -1085.65658
#> 15-1  7499.1027   905.59230  6549.24359  1396.16270 10884.1660  -800.95538
#> 16-0 -5724.3191 -1747.59015  9790.25738 -1526.90745  1828.3795  8166.51525
#> 16-1 -2989.4589  8934.52659  4760.09643  -532.43348 -1072.1907 -8490.76013
#> 2-0   3446.3486   562.56745   176.68689  2878.18581  6578.4623 -1381.84893
#> 2-1   2775.0360  8840.89624  6935.66288  7043.65900 -1238.6548  -583.82883
#> 3-0   8410.3672  4610.52400  6930.87997  4008.19121  4893.0989   480.99383
#> 3-1  -7235.3281 13696.09124  6923.45590  6827.89324  4647.5336 -3335.89186
#> 4-0  10638.6083   947.23209  3570.01238  3495.68794 -2379.2196 -6736.29568
#> 4-1   1412.3901  1016.77363 17211.73968 -2896.67287  8325.9714  2548.78423
#> 5-0   1220.3936 10639.60543  2378.73505  7518.04932 -1883.3327 -4997.29157
#> 5-1   7157.5704  3934.47937 -1433.62099  9340.70718  2709.3207  3621.19653
#> 6-0  29406.9151  3991.66743  1561.47503   278.65775  6081.2996  1441.22637
#> 6-1   3991.6674 23625.85424    96.39912  7948.94043 -6012.6421 -8774.49794
#> 7-0   1561.4750    96.39912 28075.26526 -3242.44596  4287.4374 -2481.21729
#> 7-1    278.6577  7948.94043 -3242.44596 17992.36313 -2417.7658   -80.24206
#> 8-0   6081.2996 -6012.64210  4287.43739 -2417.76583 24609.0558  4186.91585
#> 8-1   1441.2264 -8774.49794 -2481.21729   -80.24206  4186.9159 23676.90488
#> 9-0  -6861.3126 -6007.87578  1972.79492 -3586.15570  3603.0807  5227.06530
#> 9-1   1270.5010  -839.31871  5665.08638  2975.00729  4241.7754  3647.54462
#>               9-0        9-1
#> 1-0    811.531302 -3165.0316
#> 1-1  -2334.233923  7953.4090
#> 10-0    74.693044   842.5924
#> 10-1 -4040.629176  -467.3622
#> 11-0     7.188923   615.9905
#> 11-1  -103.123092  5773.9386
#> 12-0  2075.582106  2173.5072
#> 12-1   584.913221  -419.1209
#> 13-0  1410.511356  7008.1887
#> 13-1 -3058.218685  5319.2480
#> 14-0 -2523.291698  7928.1106
#> 14-1   354.337700  4679.9841
#> 15-0 -2514.101536  9080.7330
#> 15-1  -615.349984  5622.6708
#> 16-0 -1024.651404  5746.0260
#> 16-1  -752.060974   321.2172
#> 2-0  -2549.227867  2746.5849
#> 2-1   1037.483396  2253.4917
#> 3-0   5591.241496  3715.8136
#> 3-1   2276.097946  -277.6833
#> 4-0  -7484.160857  2315.2284
#> 4-1    506.944710  9657.1757
#> 5-0   -290.101797 -4358.0701
#> 5-1  -6042.926016  1537.7183
#> 6-0  -6861.312588  1270.5010
#> 6-1  -6007.875784  -839.3187
#> 7-0   1972.794921  5665.0864
#> 7-1  -3586.155698  2975.0073
#> 8-0   3603.080702  4241.7754
#> 8-1   5227.065298  3647.5446
#> 9-0  17955.299204 -4650.0345
#> 9-1  -4650.034473 15154.3452
#> 
#> $stimcats
#>    stim cat
#> 1   1-0   0
#> 2   3-1   1
#> 3  15-1   1
#> 4  16-1   1
#> 5  13-1   1
#> 6  10-1   1
#> 7   6-1   1
#> 8   7-1   1
#> 9   4-1   1
#> 10  1-1   1
#> 11  8-1   1
#> 12 14-0   0
#> 13 11-0   0
#> 14  8-0   0
#> 15  5-1   1
#> 16  5-0   0
#> 17  2-0   0
#> 18 15-0   0
#> 19 11-1   1
#> 20 12-0   0
#> 21  9-0   0
#> 22  6-0   0
#> 23  3-0   0
#> 24 16-0   0
#> 25 13-0   0
#> 26 10-0   0
#> 27  7-0   0
#> 28  4-0   0
#> 29  2-1   1
#> 30 14-1   1
#> 31  9-1   1
#> 32 12-1   1
#>