Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

aat_stimulusscores(
  ds,
  subjvar,
  stimvar,
  pullvar,
  targetvar = NULL,
  rtvar,
  iters = 5
)

Arguments

ds
the data.frame to use

subjvar
Name of the subject-identifying variable

stimvar
Name of the stimulus-identifying variable

pullvar
Name of the movement-direction identifying variable

targetvar
Optional. Name of the stimulus-category identifying variable

rtvar
Name of the reaction-time identifying variable

iters
If there are missing values (which is almost inevitable) then multiple imputation will be used to complete the covariance matrix - this argument sets the number of multiple imputations to be used.

Value

Exports a list containing a data.frame with stimulus-specific bias scores, indicated in the column names, a covariance matrix of that same data, and a data.frame indicating to which stimulus category each stimulus belongs.

Examples

ds<-aat_simulate(biasfx_jitter=40,nstims=16)
ds$stim<-paste0(ds$stim,"-",ds$is_target)
aat_stimulusscores(ds,"subj","stim","is_pull","is_target","rt")

#> $data
#>          0-1-0      0-10-0      0-11-0     0-12-0      0-13-0      0-14-0
#> X1  -80.112533   25.113696   19.534052   59.52155  285.022983   23.454422
#> 2  -147.648360  -40.540887    7.858474  -58.70277   29.549480  -75.312538
#> 3  -144.577202   10.127563  296.296052 -109.18966   93.010097  179.050728
#> 4   167.682420 -227.706028   23.952605  110.38167 -169.608281 -138.366204
#> 5  -181.377444 -148.850517 -105.431461 -264.64584  -48.530601  -77.389602
#> 6   -57.522820  -64.043768 -145.856233 -160.91257   -6.954944   24.599771
#> 7   362.598700  185.894583  -29.022621   74.87999  -90.218775  262.152706
#> 8  -147.375096  -24.120133  -35.361881 -218.65139 -207.996943 -196.827545
#> 9  -316.779739  -38.650011   24.375988  -99.82826 -131.494693 -129.890862
#> 10 -101.278522 -178.910894  -87.093650 -121.32180 -212.397230 -131.714283
#> 11  195.046802  -99.362017 -268.206260 -149.93357  -78.485945 -167.504617
#> 12 -204.616348 -103.650729  160.651275  131.28456   23.947677  182.216559
#> 13 -145.556260  -80.327200 -121.423404 -160.93244  -85.255624 -258.272071
#> 14  -77.681659  -62.497821   80.029167  101.13265 -119.804034 -134.689178
#> 15   25.541780 -181.918849 -121.904747 -243.17406  349.973633 -129.991772
#> 16 -125.722509  -31.141888   29.240116  -84.63970 -236.488078   -3.817449
#> 17  -19.890419  -89.524864   85.577512  -10.09300   -5.070609   22.489018
#> 18  144.968432  -48.741726   50.502644   82.35024   44.891130  -54.692806
#> 19 -104.324918  -75.922848 -212.696983 -105.98993 -239.593533 -151.587024
#> 20   17.159345  -48.075497 -182.097984  -36.32434 -155.163701  -40.034116
#> 21  -21.827328   -2.997035 -106.907508 -113.21679 -135.331549  -98.662499
#> 22  253.146088  105.395161 -296.093133  -79.33721 -334.345165   60.409662
#> 23 -259.947941 -207.931811  180.823916  138.12839   93.824061 -483.990654
#> 24  -86.620025 -102.665437 -147.460893 -141.27981 -193.331588   85.316132
#> 25  -20.543373  -16.904741  -76.660197  -31.38401    0.641825  -25.171743
#> 26  -47.310672 -165.848437   67.010334 -102.06116  -67.137583 -140.867619
#> 27  116.101434 -113.643732   78.820979  -75.59000   57.315372   73.014211
#> 28 -211.268965   28.471060   33.129364 -201.22315 -371.748287 -592.305235
#> 29   -5.457715  250.750921 -409.295152 -309.96877   63.749732    1.414232
#> 30  -10.034691   43.137195 -128.172082 -168.53934   70.615301 -147.787303
#> 31 -139.734697  -18.152887   24.291074 -314.66295  -80.245581  123.821760
#> 32   66.047395   66.822278   54.492010  -45.32446  -32.830742  -73.454113
#> 33   57.037387  155.186367 -207.470816  -22.99241  -13.255691  -40.138653
#> 34  136.764087 -144.764434 -429.469202 -598.90134 -297.533736 -818.366031
#> 35   26.637135 -238.326375   41.486201  224.09487 -188.791720   36.507103
#> 36    9.151939  -91.212860  -30.307464  -76.15003   66.334266  -66.385021
#> 37 -131.110851 -277.521475  112.759743  107.83470  216.368875   21.365565
#> 38   95.103253   20.439675  153.012545   18.58099  181.560758  152.381152
#> 39 -219.567830 -129.386104  -51.794333 -310.79912 -187.847016  -62.414885
#> 40  155.735445  -40.590118  119.449999  -29.50564   98.428392   49.212131
#>           0-15-0        0-16-0       0-2-0       0-3-0       0-4-0       0-5-0
#> X1 -162.29021485   68.16216889  -81.299370   57.943539  124.973452 -198.916169
#> 2   -48.86923761  -13.31987242 -186.925375 -114.531451 -314.414051 -109.579850
#> 3    78.31085329 -275.97536860 -266.059071   32.282844 -397.187428   28.256096
#> 4   -59.10262564  -99.65875909   61.237359 -144.140544  -43.114528   32.848889
#> 5   -79.35821771 -104.20638509 -154.331723 -187.716950  -28.821443    4.168837
#> 6  -120.65890858  145.01541140 -233.482478    4.398394   83.959048  -10.620799
#> 7  -104.92137995   85.09336381  -88.248152   14.260134  -65.594187  233.447815
#> 8  -228.92967559  -16.25472566   -6.891942  329.903459  155.196469  -35.134147
#> 9  -272.30507681   13.09142824 -111.705081   -7.732346 -164.373606   -8.418657
#> 10  -17.18029763 -143.33210239  -19.237330  -42.815176 -170.214826  -77.143232
#> 11 -164.78029866 -300.76968308  -89.730664  -89.053372  -85.966852 -168.461285
#> 12   95.30378406   70.41540131 -294.573042 -222.221937  253.551497   13.325675
#> 13 -217.96042404   95.54771855 -321.856527  -17.769843 -165.125399  -19.596292
#> 14  101.70548664    0.08243101   36.248573  -51.990432  187.695963 -145.358306
#> 15  -10.82541175  -11.44409138   31.679645   -9.199121  -38.603317 -345.331265
#> 16 -136.64084893  -39.89354602  -45.711350  -46.446225    1.891475 -128.409911
#> 17   89.09821011    6.03505409   28.270029  110.801584 -110.974920   13.740933
#> 18  -43.32211729   96.90441062   17.665299 -267.947045  182.796028  -15.403867
#> 19   40.04745572 -282.30058025  -85.865259  -72.364429 -207.264358 -152.197346
#> 20  -90.12441409  175.20133276  -65.214910 -317.315107  -29.596094   15.741734
#> 21 -185.56148726  -78.46653447   95.829870  -79.987300  -51.023084 -171.602398
#> 22   16.61732864 -115.04943733  183.322736  300.663611 -234.617748  -13.911528
#> 23   40.55646331  165.42464656 -106.963966  248.349708   -1.477172  -92.916969
#> 24  -97.90759450   -1.28022947  -80.627956  185.622894  134.664554  100.524913
#> 25 -200.70153555 -144.97806529  -28.854556  -29.682406  -51.739316   41.542465
#> 26   19.07901263  -34.22040217 -133.063177 -120.848407  -97.148630 -149.358521
#> 27  -73.44918923  202.94488012  138.963219  210.206343    7.205836  -28.909683
#> 28    0.07483052 -215.91184586   11.311788  -42.384767   74.611649 -193.963661
#> 29  200.29760761 -159.68115835 -274.254615  -64.280520   17.019791  -82.835694
#> 30 -119.14785081   -1.04281429   33.698647    8.171027 -181.355034  -25.371528
#> 31 -297.96296078   24.95632611   44.090789 -170.049062   25.371908   62.913443
#> 32   66.50453775  114.74351620  -36.115194  177.492420  -39.197292    2.054839
#> 33  -38.18928682   59.39861904 -194.514318 -213.136602  -32.582678   24.536525
#> 34 -915.13470145  278.52217756 -262.750874   98.761799 -269.373934 -168.794438
#> 35  283.09334285   -7.77818586   97.579629   45.609394   58.428913 -155.787750
#> 36  -54.65904379  -11.96181076   21.806744 -100.099030  -69.646537  109.353925
#> 37   62.91225176  -30.17853867   96.363264  -48.729375   -6.439311   35.379712
#> 38  146.32069042  142.45996092   52.583761 -110.863997   41.992772  175.596393
#> 39   27.72153392  -25.69380176 -151.644781  -95.324214  100.380555 -225.750788
#> 40   22.31024512   74.05332477   87.257443   34.576932   31.756857  -20.410288
#>          0-6-0        0-7-0        0-8-0        0-9-0        1-1-1      1-10-1
#> X1  -91.802088  242.3420240   59.1636779   47.8823760 -124.9408987  -87.075866
#> 2   -97.944193  -79.2783408    0.1052105    0.4517835   99.2512164   22.111360
#> 3     5.856516  -59.0184380   -7.0793053 -157.3832338    0.9101115 -269.060389
#> 4   -39.949163 -344.1390010   73.2876194 -112.9093580   39.5542336   25.088492
#> 5  -101.653796 -218.1347371  -68.6644330  -88.9565049   52.7492510   -3.429502
#> 6    43.807941  144.2386580   17.9892204 -122.5010625   92.4729593  -87.705286
#> 7   155.925422 -289.0845636  223.8079324  -86.4464806 -176.5203498  -41.176502
#> 8    48.301732  300.6887904 -193.2661655 -212.5809892  -69.5936872   25.577077
#> 9   183.307583   23.3126340  -28.9862208 -303.1656247  -10.0648031    7.286420
#> 10  -15.507624 -101.8070719 -296.4554986   20.7384395 -175.9676284  103.392013
#> 11  -35.237667 -154.6834100  -40.3665034  -50.0671904 -227.1240809  148.482738
#> 12   17.694631  285.8779802 -251.1723631  219.4010207  107.3498447 -268.934901
#> 13 -104.923880 -262.3212000 -241.6867381 -118.2750720  -78.4161281 -397.663335
#> 14   41.555984  -47.2692836 -156.0169218  -62.1293666  -44.8208139  104.290663
#> 15   23.006220 -223.0800446   70.8409841 -106.5732211   51.4717079   31.787630
#> 16  -79.232966  -15.8341733  -71.9081578 -103.6217661 -203.3298525   22.857966
#> 17  -26.907273  125.6903942  -56.7873541  141.8039070   22.2157253   93.200706
#> 18   68.580237   29.7440183   -3.6400381 -144.8395140   42.3767962   29.816060
#> 19  -89.286180 -113.0353204  -61.6176200   76.8613167  -62.4294852    4.774242
#> 20  112.020970 -191.1364067  193.6820191 -168.2497241 -135.7048177  115.979968
#> 21  -67.612088 -149.2863870 -136.7152785    6.1968657 -180.8735628   44.482186
#> 22 -172.382297  -42.9875303 -359.2919620 -108.8298682  -41.1037732  -49.397434
#> 23   49.367013  -88.9743827   98.3062317  149.5863555  140.8115521 -404.115620
#> 24 -175.749106   15.8909005  -48.2281383   -9.1329489   52.1412432 -103.446809
#> 25   -9.675925    2.3478072 -142.0500537 -136.3984440   26.4116454  -32.077325
#> 26   20.721779 -124.5085122 -170.0535717  -40.4465160 -174.6689441 -178.107757
#> 27  133.583792  143.3300629  -42.3325937   22.8355972   -1.5029756   85.963462
#> 28  109.881369 -331.3961583 -132.4036481  -81.7790767 -310.5961638 -290.043355
#> 29 -296.065595 -183.9692744  -25.0459628  -59.6504329   99.7930593  204.491279
#> 30  -90.384577    0.5206916  -84.3641371  -79.6786401   -8.7226995   -5.834192
#> 31 -135.344490   12.5169140   40.3602637   28.2540864   83.5560174    3.557801
#> 32   -7.559147  -28.7550780  -72.0412144   75.1227314 -278.1809082  100.699594
#> 33  -61.012288    4.8115626 -219.0976284 -140.4996499  -15.3015472   67.903020
#> 34 -398.752526 -641.0370529    9.9400327 -547.8758653 -306.1126606 -168.327565
#> 35  -70.801349   70.0849772  -50.2590038  184.8006245   98.8453819 -352.794050
#> 36  -53.500677  -27.9438317 -138.8736368  108.4813289  117.5210963 -112.312465
#> 37  148.760488   63.8647399   70.5703141   82.6750579  -35.0734116  -10.405588
#> 38  -52.399164   16.8349214   29.0883714   13.0034290   58.4770595  174.229674
#> 39 -265.343848 -276.7575351 -180.7387311 -322.5514739    5.0989574 -117.426701
#> 40 -116.471274    7.1926891  153.3826045 -149.5378139  142.4921312  -34.695294
#>         1-11-1      1-12-1      1-13-1      1-14-1      1-15-1       1-16-1
#> X1  -63.534668   77.856017  291.872007  -42.735733   30.615613 -135.9396892
#> 2    74.087147  -47.607052  122.402488  133.775138   16.496700  -43.4689066
#> 3  -250.495527   28.589687  -70.595961  -26.071919   59.824821  222.4986061
#> 4  -178.123012  -18.954142  135.273965   46.374968  105.492967   43.2348951
#> 5   -97.632041  -28.090984   83.870865 -110.855247  -10.429875  -62.6920544
#> 6  -109.338435  350.704629 -148.394866  169.369165 -230.244059 -150.8022110
#> 7   104.467320 -185.371519  171.386797  -23.752417   -8.811972   37.2020508
#> 8    66.282841 -273.387067   42.805146 -177.189604   86.089132 -168.7323859
#> 9   -47.755458  -27.777825   60.596658  -41.706829   96.459384  -69.9764353
#> 10    6.360077   69.159543   -2.783625  161.572412  167.840720  163.4617239
#> 11 -137.293463 -222.645282 -144.365158 -110.172515  -35.584377 -118.6591523
#> 12  -67.212920 -192.379539 -190.428911 -102.341882  -74.615561 -344.6864779
#> 13   46.003246 -137.756520 -203.108514  129.722932 -247.530693 -210.1777226
#> 14   56.195431   17.150813 -104.923597  -17.995092 -164.343330   50.4955215
#> 15  -23.922336 -107.117584 -301.834377  226.502632 -315.011404  -18.4441629
#> 16  -36.685408  108.162773   85.511223 -123.191383  -72.299233  -94.0492203
#> 17  -87.636837  -48.261922  -62.040375   -5.101721   87.527404  -20.3777483
#> 18   36.099837  -96.686474  381.235160   16.749233 -222.781342   37.2158485
#> 19 -112.076617   23.150845 -183.873543  -69.388850  -84.534661 -135.4037480
#> 20 -198.588163   28.468284  243.667298   77.970575  108.493798   38.0703509
#> 21 -106.761624  208.326688   55.275907 -145.223208   63.530234  -53.1615708
#> 22 -148.645797  -65.762060  -52.156385  175.822117  150.537795 -200.9311908
#> 23  241.215557 -111.781065  176.944867   84.710836  222.795153   92.9299219
#> 24  -80.991919   52.924290  199.554364  -23.783454   57.657102    0.4575231
#> 25  -24.940956  -18.668832   67.002910   25.536170  -51.761671  116.8366477
#> 26  -99.856248   85.344149 -105.669350 -270.986992 -174.738942  -95.1309461
#> 27   72.570987  138.908033    8.277048  -17.093330   94.437014  -24.7123738
#> 28 -279.283752   31.971214  -43.492111  -45.956012  -48.253627 -101.5554878
#> 29 -160.258349 -142.151246   24.078381   76.344070  -15.618216  -37.7789184
#> 30  127.501092   24.730272   51.029663  137.958839   67.161339  396.9229720
#> 31  -56.571504   37.615010 -118.970595   96.653641  -58.575259 -131.8042014
#> 32 -108.569754 -343.918640  -48.734406  -46.838067   89.231610  -34.2285267
#> 33   76.204145 -171.224135  -58.385746   87.718346  -69.702796  -80.0737184
#> 34  136.737261  249.979360  322.741927 -773.610124  143.627389   -4.7347472
#> 35 -254.680608  -36.750817  -27.362352   90.034045  186.021939   56.3443598
#> 36   77.788627  -22.100261  -86.039160  204.585649  -97.060175   55.6656849
#> 37  205.847517  -83.869649  123.738329    5.912301   59.262148   38.5288197
#> 38  158.521130  -15.041940  138.119821    5.280265  -47.324453   53.7802827
#> 39 -326.927792   -7.269034  -81.884049   74.899025  -80.728174 -119.2429096
#> 40  -11.307141  -74.556689   -2.607806  288.566947    7.167885  138.7079193
#>           1-2-1       1-3-1      1-4-1      1-5-1       1-6-1       1-7-1
#> X1  221.9211757   35.635933  125.16754   49.12287   83.258149  -38.954338
#> 2   142.8522647 -126.756402   93.15858  100.49418  -54.080644  -66.864427
#> 3   123.7813941  -31.790887 -122.86343  -18.21432  295.280653 -187.884251
#> 4  -204.3856824  -68.232455 -132.38532  -75.50166 -121.603540  -41.211702
#> 5   -40.5249569 -171.052819 -138.50781    5.01172 -122.855700  -38.475597
#> 6   -45.6792082 -180.468253  -95.77844   15.07258  -94.763711  225.289994
#> 7   -20.3957928 -183.523611  -94.44296  110.12320  359.790946   60.273177
#> 8     1.1119652   -1.444593  -33.72189  206.07779  151.027231  -62.705217
#> 9  -180.1062458   80.913056   35.86871   83.88308   52.773512   66.944587
#> 10    1.0141861  -28.424235  122.10814  186.81312  -30.623261  -13.248067
#> 11  -22.5727360   10.047208   14.84503  -81.41707  -28.093620   43.922906
#> 12  118.0190330   93.771568   64.69537 -206.45167   27.235427  184.511327
#> 13 -294.5939875 -256.895255  -89.65915 -329.23877  -35.029741 -110.499055
#> 14    0.9610943 -117.622948   67.15848  215.70919 -101.583125  -70.662409
#> 15  -95.8809036  -36.106980  -84.03402   12.51126   58.300028   45.879676
#> 16   35.4828885  -93.204611  -15.51138 -168.88363  -30.210117  -83.204195
#> 17  -58.5325969  -96.595521  -32.17336  -21.12151   -1.933790   48.090120
#> 18  -50.7475320   69.858646   52.55439 -230.21130    3.518305    1.782710
#> 19 -179.3076768  -28.047145  -44.83793 -102.33965  -54.177736 -194.887905
#> 20   56.2981665  -60.487959 -170.80911   57.52277   55.898223 -242.328617
#> 21  -17.6552535 -226.344831 -124.57493  -97.69778 -158.213750 -243.204755
#> 22  -30.9198020 -323.446185  116.72474 -169.16119    1.901875 -229.449140
#> 23   21.4996767  127.072323 -115.99573   38.88712  120.540543 -112.171752
#> 24 -115.0200420   55.837462   24.78751  113.08913  181.363513  -73.508634
#> 25  218.5001889   79.277129   40.77658  -37.18698  -99.813835   72.507623
#> 26 -250.2658382   19.393442  -24.30158  -89.22940  -55.066901 -122.105563
#> 27  137.6037792  -11.472787  -31.30687 -218.38641  -94.400329  128.718698
#> 28   18.1667886  280.222500 -202.62985  -31.90363 -279.978686 -213.725331
#> 29   66.2286009   67.549789  -39.89197  133.58273  135.969995 -108.781287
#> 30   79.5357837  237.409403   89.77604  332.88616   22.580388   20.467071
#> 31  -95.5105062  176.598789  -26.99403  -37.28405  -53.753116   -8.355322
#> 32  109.8504377  -88.837116  -99.89022   87.66641   36.579555  121.024504
#> 33   33.2124252  -26.506340 -212.71277  -76.13843    4.064744  -44.902709
#> 34  441.9007731 -414.512205   77.29651 -182.42655 -875.402932  111.658881
#> 35  -14.3376581  122.741927  -15.62728  405.54482   -7.737924  100.602382
#> 36   48.3791611 -153.101065   44.39240  -14.23609  141.566005  113.949653
#> 37   59.7846274   -9.394147  122.12867  139.44249  -71.132312   54.636761
#> 38   40.5714862  183.578347   86.73180   58.23916    5.033174  -96.423296
#> 39 -188.3292827 -252.671652  -32.87625   98.34521   51.495899   26.914465
#> 40   66.9888924  -66.729375  -70.64952   28.80431  293.702830  -41.985544
#>          1-8-1      1-9-1
#> X1   66.071548  -92.50228
#> 2   -27.090170  -77.22582
#> 3   -68.904328   53.49093
#> 4  -133.849478 -111.39161
#> 5   101.995110 -125.40115
#> 6   -54.932913   43.15904
#> 7   114.028139  -78.55251
#> 8  -175.364867  184.15630
#> 9   -64.770220   92.88280
#> 10   12.983091  107.60852
#> 11 -205.255721 -122.40846
#> 12 -136.925261  -33.45390
#> 13  136.728367   10.70535
#> 14   76.883913  351.74972
#> 15 -233.120858 -116.51472
#> 16  -17.483157 -239.46846
#> 17   77.479823  147.59944
#> 18   49.912633  101.61353
#> 19 -212.920710   42.19851
#> 20   19.586791   51.89962
#> 21  -71.355863 -119.18265
#> 22   28.948688   11.52293
#> 23   89.420442  -58.49181
#> 24 -121.037045 -271.43008
#> 25  195.006193  -36.15071
#> 26  142.699218 -153.57066
#> 27 -147.617622  117.87676
#> 28  -70.532127  -28.99915
#> 29  271.607024  376.40584
#> 30   -9.519962   -2.81344
#> 31   33.011802   71.50680
#> 32 -166.203887 -103.10489
#> 33  -24.201100  -17.68393
#> 34 -367.091481 -691.03596
#> 35 -244.443377  272.44113
#> 36   13.236766  119.15304
#> 37   32.555218   53.18068
#> 38  110.818744   53.68815
#> 39 -105.427350  -40.32366
#> 40  -83.897010   39.32259
#> 
#> $covmat
#>              0-1-0     0-10-0     0-11-0     0-12-0     0-13-0     0-14-0
#> 0-1-0  21655.01681  4537.6210 -6726.3228  2581.3722   450.9536  5687.1087
#> 0-10-0  4537.62105 12647.9659 -4647.9601 -2287.5655  -445.3133  5861.0967
#> 0-11-0 -6726.32285 -4647.9601 24006.9036 14279.3425  8765.8907  9840.5233
#> 0-12-0  2581.37217 -2287.5655 14279.3425 24209.6949  6689.8281 13422.0602
#> 0-13-0   450.95357  -445.3133  8765.8907  6689.8281 26075.2126 11337.6544
#> 0-14-0  5687.10870  5861.0967  9840.5233 13422.0602 11337.6544 40040.1809
#> 0-15-0 -1618.94482   724.3775 12368.2293 17636.4616  8082.1524 19062.5828
#> 0-16-0  2912.23949  -954.5503  2045.5684  1429.3575  4682.2833 -1711.9460
#> 0-2-0   6510.72800 -2134.0712  4197.3941  6460.8513  -189.0455  4142.8723
#> 0-3-0   1400.35481   451.1564   289.4523 -1588.0373 -2382.1518 -4233.8033
#> 0-4-0  -1916.03253  -948.0637  4620.5973  6383.1681  2422.6855  5155.9563
#> 0-5-0   4200.93518  3542.1445  3759.0424  4928.8439  1473.5479 11525.4517
#> 0-6-0  -1517.18762 -1648.8601 10177.9396 10574.1368  3869.9771  4744.0259
#> 0-7-0  -5450.32490  1064.7434 12381.2067 13563.8296 10553.6756 20562.3919
#> 0-8-0   4447.37614  -750.1734  3403.5327  2743.8537  8226.0836  2694.3716
#> 0-9-0  -2230.01137 -2386.7795  9269.4921 13833.1255  7241.0464 11352.3656
#> 1-1-1  -2877.61523 -2037.5011  4996.7194  5389.6666  9468.1396 10801.2987
#> 1-10-1  7802.02988  5313.0502 -6241.7435 -2150.0403  2968.1535  8707.2079
#> 1-11-1  1863.80294  -858.0901  2051.4411  2806.4994  8364.0536 -3574.8851
#> 1-12-1 -2229.07628 -3300.5286 -2327.4557 -5866.2796 -3671.7502 -4164.5812
#> 1-13-1  4006.93085   836.6583  -269.1303  2762.5469  -157.1723 -3941.6399
#> 1-14-1   -12.87621  1960.7137  5062.5628  9987.4786  9545.5993 16159.6401
#> 1-15-1   732.69710  -974.1186   650.2253  2416.4491 -5314.2891 -2584.8631
#> 1-16-1  3570.33872  -746.8590  3820.8805  2749.4115  5318.5978  -119.1129
#> 1-2-1   3572.01348  3018.5610 -1238.8824 -1906.2557  4439.0897 -3904.6967
#> 1-3-1  -5898.61320   637.9107  8749.1421  6920.4815  6113.3340  3353.2189
#> 1-4-1    360.95003 -1461.9116  -162.5839  1130.5629  2190.3442  2216.5562
#> 1-5-1  -2971.26873  -121.8732  2208.1136  3846.8010  2158.3265  3929.1450
#> 1-6-1    281.09516  6092.5075 10959.1354 12907.2445 11198.9364 25660.0352
#> 1-7-1   1853.99630 -3225.6879  1375.2461   716.7083  5293.7461  2963.2282
#> 1-8-1  -1693.69164  5316.0156  2428.5415  4472.6081  5691.7063  7722.3134
#> 1-9-1  -3103.22435  3520.8811  6920.8914 10872.3674  4988.8212 14750.8185
#>            0-15-0     0-16-0      0-2-0      0-3-0        0-4-0        0-5-0
#> 0-1-0  -1618.9448  2912.2395  6510.7280  1400.3548 -1916.032534  4200.935182
#> 0-10-0   724.3775  -954.5503 -2134.0712   451.1564  -948.063688  3542.144489
#> 0-11-0 12368.2293  2045.5684  4197.3941   289.4523  4620.597262  3759.042419
#> 0-12-0 17636.4616  1429.3575  6460.8513 -1588.0373  6383.168076  4928.843869
#> 0-13-0  8082.1524  4682.2833  -189.0455 -2382.1518  2422.685539  1473.547915
#> 0-14-0 19062.5828 -1711.9460  4142.8723 -4233.8033  5155.956252 11525.451705
#> 0-15-0 35067.0144 -7357.9890  5229.2813 -2378.9141  6568.953293  1237.146397
#> 0-16-0 -7357.9890 17426.4942  -682.2818  2141.5387  4979.587667  3281.885752
#> 0-2-0   5229.2813  -682.2818 16367.0863  4206.2358  2968.691533   525.737460
#> 0-3-0  -2378.9141  2141.5387  4206.2358 21333.2431 -2192.750318  -681.777249
#> 0-4-0   6568.9533  4979.5877  2968.6915 -2192.7503 19692.815512    -9.959702
#> 0-5-0   1237.1464  3281.8858   525.7375  -681.7772    -9.959702 13089.286745
#> 0-6-0   6088.9384   767.9881  4025.0089 -1018.4699  3016.227516  3109.358950
#> 0-7-0  12527.5241  2258.2339  5026.6618  5203.8162 10566.704661  4818.873779
#> 0-8-0  -1702.3974  5171.8746  1035.0374 -2020.1734   514.953663  2415.769808
#> 0-9-0  17367.5938 -2214.6305  4892.6998   175.5647  4643.300876  2289.540808
#> 1-1-1   9449.4056  1692.9923   395.6290 -1420.2182  3537.096397  3915.507108
#> 1-10-1  1175.0105 -1682.2930  6070.9518 -4161.9694   689.308369  2094.329399
#> 1-11-1 -6483.8884  8710.5581  1296.9079  3158.6343  -548.191818  4552.216527
#> 1-12-1 -8778.8089  2154.7234   417.1715 -1361.5523 -2377.719524 -2975.964083
#> 1-13-1 -9069.6053  7310.8776  1621.3191   251.0276  1536.735805  4152.175998
#> 1-14-1 19182.2813 -2660.6525  6092.9389 -2123.8543  2060.878144  3399.282068
#> 1-15-1 -1423.9958   629.0778  4018.0181  8188.5629 -4543.126653  2236.233244
#> 1-16-1  2433.7780  -366.5914  4736.2557   653.1936 -5356.907624  2837.160102
#> 1-2-1  -7783.7802  4694.9632  -901.8359  3284.3925 -1516.185350   199.662038
#> 1-3-1  10140.3788 -3254.8318  3604.7464 -2713.2340  6828.300889  1328.228439
#> 1-4-1  -1733.6134   539.8009  2163.1581  1872.6552  -322.747787   150.154826
#> 1-5-1   9435.1777 -1599.5718  4116.8059  3564.6192  2473.687544    54.823531
#> 1-6-1  20973.1701 -3874.0176  1649.8535  2329.6790  4060.216065  7697.978559
#> 1-7-1  -3281.0309  5856.3227 -1223.9351   992.3327  3824.339125  1950.502969
#> 1-8-1   7242.9010  -659.0311 -1282.2912 -4243.5406  1109.438138  5640.704940
#> 1-9-1  21909.4479 -4580.2062  5613.9523 -1325.6867  6334.811240  2977.788142
#>             0-6-0     0-7-0      0-8-0      0-9-0      1-1-1      1-10-1
#> 0-1-0  -1517.1876 -5450.325  4447.3761 -2230.0114 -2877.6152  7802.02988
#> 0-10-0 -1648.8601  1064.743  -750.1734 -2386.7795 -2037.5011  5313.05015
#> 0-11-0 10177.9396 12381.207  3403.5327  9269.4921  4996.7194 -6241.74351
#> 0-12-0 10574.1368 13563.830  2743.8537 13833.1255  5389.6666 -2150.04030
#> 0-13-0  3869.9771 10553.676  8226.0836  7241.0464  9468.1396  2968.15345
#> 0-14-0  4744.0259 20562.392  2694.3716 11352.3656 10801.2987  8707.20788
#> 0-15-0  6088.9384 12527.524 -1702.3974 17367.5938  9449.4056  1175.01049
#> 0-16-0   767.9881  2258.234  5171.8746 -2214.6305  1692.9923 -1682.29303
#> 0-2-0   4025.0089  5026.662  1035.0374  4892.6998   395.6290  6070.95180
#> 0-3-0  -1018.4699  5203.816 -2020.1734   175.5647 -1420.2182 -4161.96943
#> 0-4-0   3016.2275 10566.705   514.9537  4643.3009  3537.0964   689.30837
#> 0-5-0   3109.3590  4818.874  2415.7698  2289.5408  3915.5071  2094.32940
#> 0-6-0  14753.9496  7825.461  3180.3585  4689.2245  -770.1389   485.68473
#> 0-7-0   7825.4611 32991.862 -3383.9983 13398.1840  9291.9899  2683.10536
#> 0-8-0   3180.3585 -3383.998 16454.8211  -782.5210  2231.1777  2394.22165
#> 0-9-0   4689.2245 13398.184  -782.5210 20886.5789  5326.4723 -1580.61702
#> 1-1-1   -770.1389  9291.990  2231.1777  5326.4723 15968.2863 -1538.69304
#> 1-10-1   485.6847  2683.105  2394.2217 -1580.6170 -1538.6930 22979.31899
#> 1-11-1  2631.9982  3303.317  2562.5335  1287.6438  2071.4661  1786.61026
#> 1-12-1 -2699.5477 -3612.697  1167.3552 -4229.2319 -1323.5439 -1705.68905
#> 1-13-1  -603.8960 -2751.694  8348.6976 -5259.0138 -2170.0008  2489.47530
#> 1-14-1  4335.8229  9672.524   734.0906  9149.5120 12532.4116  1949.65660
#> 1-15-1  -926.9368   562.144  1396.1771  1576.7952 -1657.0768  -229.74278
#> 1-16-1  1275.3456 -3036.668  5544.7900  -834.5066  2054.5809  2247.02611
#> 1-2-1  -2815.2583  1411.119  2840.1514 -1828.6755 -2550.0506  1512.71387
#> 1-3-1   6409.1077 10769.054  2623.7849  8680.0639  4049.3498  -373.77845
#> 1-4-1  -2014.4831  5094.800 -2482.5703  1357.1761  1019.4452  2057.48741
#> 1-5-1   1334.0499  5012.473  3491.5522  3260.9779  3148.7669  3105.60210
#> 1-6-1   7181.5000 15731.342  4182.1696  9414.7654  9840.9408  1389.17776
#> 1-7-1   1352.4161  5408.697  1224.4769  1065.1336  3031.3644    30.62377
#> 1-8-1   1850.6034  2645.860  -460.5009  3621.8356  3562.5102  2216.60239
#> 1-9-1   7195.0747 15576.310 -2690.1241  9865.8158  9849.3605  5422.90126
#>             1-11-1      1-12-1     1-13-1      1-14-1     1-15-1     1-16-1
#> 0-1-0   1863.80294  -2229.0763  4006.9309   -12.87621   732.6971  3570.3387
#> 0-10-0  -858.09010  -3300.5286   836.6583  1960.71367  -974.1186  -746.8590
#> 0-11-0  2051.44106  -2327.4557  -269.1303  5062.56276   650.2253  3820.8805
#> 0-12-0  2806.49944  -5866.2796  2762.5469  9987.47857  2416.4491  2749.4115
#> 0-13-0  8364.05362  -3671.7502  -157.1723  9545.59931 -5314.2891  5318.5978
#> 0-14-0 -3574.88507  -4164.5812 -3941.6399 16159.64015 -2584.8631  -119.1129
#> 0-15-0 -6483.88836  -8778.8089 -9069.6053 19182.28128 -1423.9958  2433.7780
#> 0-16-0  8710.55807   2154.7234  7310.8776 -2660.65255   629.0778  -366.5914
#> 0-2-0   1296.90794    417.1715  1621.3191  6092.93893  4018.0181  4736.2557
#> 0-3-0   3158.63429  -1361.5523   251.0276 -2123.85426  8188.5629   653.1936
#> 0-4-0   -548.19182  -2377.7195  1536.7358  2060.87814 -4543.1267 -5356.9076
#> 0-5-0   4552.21653  -2975.9641  4152.1760  3399.28207  2236.2332  2837.1601
#> 0-6-0   2631.99815  -2699.5477  -603.8960  4335.82287  -926.9368  1275.3456
#> 0-7-0   3303.31653  -3612.6969 -2751.6939  9672.52353   562.1440 -3036.6677
#> 0-8-0   2562.53346   1167.3552  8348.6976   734.09062  1396.1771  5544.7900
#> 0-9-0   1287.64383  -4229.2319 -5259.0138  9149.51196  1576.7952  -834.5066
#> 1-1-1   2071.46605  -1323.5439 -2170.0008 12532.41156 -1657.0768  2054.5809
#> 1-10-1  1786.61026  -1705.6890  2489.4753  1949.65660  -229.7428  2247.0261
#> 1-11-1 17716.60092  -1767.9114  5929.4333 -1540.13345   266.8285  4327.1528
#> 1-12-1 -1767.91139  18092.8051  3150.3606 -5106.08056  -259.2031  2024.3141
#> 1-13-1  5929.43329   3150.3606 22939.2549 -8322.79310  8316.8790  6215.4399
#> 1-14-1 -1540.13345  -5106.0806 -8322.7931 29691.69541 -3572.0521  4600.6319
#> 1-15-1   266.82848   -259.2031  8316.8790 -3572.05212 15661.1672  4966.3308
#> 1-16-1  4327.15283   2024.3141  6215.4399  4600.63186  4966.3308 16997.1333
#> 1-2-1   5285.05026   2239.5999  8602.5421 -7884.21963  6084.7624  4598.1589
#> 1-3-1    334.96167  -3631.0105   514.8150  5615.55498   755.9305  4661.8966
#> 1-4-1   4443.76589   1058.1918  2876.3488 -1279.10115  1192.1627   992.4651
#> 1-5-1    -44.88215  -2438.2821  2020.2039  6629.97254  6441.0910  9707.8342
#> 1-6-1  -1692.87485 -11991.9013 -4028.8949 20805.66644 -1006.5868  3855.3852
#> 1-7-1   3547.98758   -737.4469 -1865.3213  -657.54310 -1534.6763  -277.8376
#> 1-8-1   3186.28151  -1352.5658  2799.7335  7295.68906 -3012.7349  2091.3713
#> 1-9-1  -2389.00753  -6232.1736 -7872.0906 17967.82957 -1374.3005  2289.2959
#>             1-2-1      1-3-1      1-4-1       1-5-1       1-6-1       1-7-1
#> 0-1-0   3572.0135 -5898.6132   360.9500 -2971.26873    281.0952  1853.99630
#> 0-10-0  3018.5610   637.9107 -1461.9116  -121.87319   6092.5075 -3225.68786
#> 0-11-0 -1238.8824  8749.1421  -162.5839  2208.11364  10959.1354  1375.24605
#> 0-12-0 -1906.2557  6920.4815  1130.5629  3846.80095  12907.2445   716.70829
#> 0-13-0  4439.0897  6113.3340  2190.3442  2158.32646  11198.9364  5293.74608
#> 0-14-0 -3904.6967  3353.2189  2216.5562  3929.14505  25660.0352  2963.22818
#> 0-15-0 -7783.7802 10140.3788 -1733.6134  9435.17765  20973.1701 -3281.03092
#> 0-16-0  4694.9632 -3254.8318   539.8009 -1599.57177  -3874.0176  5856.32271
#> 0-2-0   -901.8359  3604.7464  2163.1581  4116.80590   1649.8535 -1223.93510
#> 0-3-0   3284.3925 -2713.2340  1872.6552  3564.61916   2329.6790   992.33272
#> 0-4-0  -1516.1853  6828.3009  -322.7478  2473.68754   4060.2161  3824.33912
#> 0-5-0    199.6620  1328.2284   150.1548    54.82353   7697.9786  1950.50297
#> 0-6-0  -2815.2583  6409.1077 -2014.4831  1334.04985   7181.5000  1352.41609
#> 0-7-0   1411.1193 10769.0545  5094.7996  5012.47332  15731.3421  5408.69655
#> 0-8-0   2840.1514  2623.7849 -2482.5703  3491.55217   4182.1696  1224.47690
#> 0-9-0  -1828.6755  8680.0639  1357.1761  3260.97794   9414.7654  1065.13363
#> 1-1-1  -2550.0506  4049.3498  1019.4452  3148.76693   9840.9408  3031.36444
#> 1-10-1  1512.7139  -373.7785  2057.4874  3105.60210   1389.1778    30.62377
#> 1-11-1  5285.0503   334.9617  4443.7659   -44.88215  -1692.8748  3547.98758
#> 1-12-1  2239.5999 -3631.0105  1058.1918 -2438.28210 -11991.9013  -737.44692
#> 1-13-1  8602.5421   514.8150  2876.3488  2020.20393  -4028.8949 -1865.32127
#> 1-14-1 -7884.2196  5615.5550 -1279.1012  6629.97254  20805.6664  -657.54310
#> 1-15-1  6084.7624   755.9305  1192.1627  6441.09103  -1006.5868 -1534.67632
#> 1-16-1  4598.1589  4661.8966   992.4651  9707.83420   3855.3852  -277.83761
#> 1-2-1  18883.4731  -535.1455  3552.0391  1356.82331  -7936.9208  3648.52446
#> 1-3-1   -535.1455 22226.2403   397.9222  7306.97728   7262.7476  -447.07064
#> 1-4-1   3552.0391   397.9222  8757.2082  2262.07356  -2066.3039  2981.58658
#> 1-5-1   1356.8233  7306.9773  2262.0736 23041.42170   8707.8737  1788.18285
#> 1-6-1  -7936.9208  7262.7476 -2066.3039  8707.87371  35633.2454 -2306.08737
#> 1-7-1   3648.5245  -447.0706  2981.5866  1788.18285  -2306.0874 13699.15831
#> 1-8-1  -2498.5072  2949.3821   758.9252   288.60226   8089.2325 -4037.67226
#> 1-9-1  -6074.6343  9088.1963   319.4109 10328.29038  16538.6896 -1020.84087
#>             1-8-1      1-9-1
#> 0-1-0  -1693.6916 -3103.2244
#> 0-10-0  5316.0156  3520.8811
#> 0-11-0  2428.5415  6920.8914
#> 0-12-0  4472.6081 10872.3674
#> 0-13-0  5691.7063  4988.8212
#> 0-14-0  7722.3134 14750.8185
#> 0-15-0  7242.9010 21909.4479
#> 0-16-0  -659.0311 -4580.2062
#> 0-2-0  -1282.2912  5613.9523
#> 0-3-0  -4243.5406 -1325.6867
#> 0-4-0   1109.4381  6334.8112
#> 0-5-0   5640.7049  2977.7881
#> 0-6-0   1850.6034  7195.0747
#> 0-7-0   2645.8595 15576.3102
#> 0-8-0   -460.5009 -2690.1241
#> 0-9-0   3621.8356  9865.8158
#> 1-1-1   3562.5102  9849.3605
#> 1-10-1  2216.6024  5422.9013
#> 1-11-1  3186.2815 -2389.0075
#> 1-12-1 -1352.5658 -6232.1736
#> 1-13-1  2799.7335 -7872.0906
#> 1-14-1  7295.6891 17967.8296
#> 1-15-1 -3012.7349 -1374.3005
#> 1-16-1  2091.3713  2289.2959
#> 1-2-1  -2498.5072 -6074.6343
#> 1-3-1   2949.3821  9088.1963
#> 1-4-1    758.9252   319.4109
#> 1-5-1    288.6023 10328.2904
#> 1-6-1   8089.2325 16538.6896
#> 1-7-1  -4037.6723 -1020.8409
#> 1-8-1  17757.9204  8962.3363
#> 1-9-1   8962.3363 31219.9431
#> 
#> $stimcats
#>      stim cat
#> 1   0-1-0   0
#> 2   1-3-1   1
#> 3  1-15-1   1
#> 4  1-16-1   1
#> 5  1-13-1   1
#> 6  1-10-1   1
#> 7   1-6-1   1
#> 8   1-7-1   1
#> 9   1-4-1   1
#> 10  1-1-1   1
#> 11  1-8-1   1
#> 12 0-14-0   0
#> 13 0-11-0   0
#> 14  0-8-0   0
#> 15  1-5-1   1
#> 16  0-5-0   0
#> 17  0-2-0   0
#> 18 0-15-0   0
#> 19 1-11-1   1
#> 20 0-12-0   0
#> 21  0-9-0   0
#> 22  0-6-0   0
#> 23  0-3-0   0
#> 24 0-16-0   0
#> 25 0-13-0   0
#> 26 0-10-0   0
#> 27  0-7-0   0
#> 28  0-4-0   0
#> 29  1-2-1   1
#> 30 1-14-1   1
#> 31  1-9-1   1
#> 32 1-12-1   1
#>