Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

Compute stimulus-specific bias scores Computes mean single-difference scores (push - pull) for each stimulus.

aat_stimulusscores(
  ds,
  subjvar,
  stimvar,
  pullvar,
  targetvar = NULL,
  rtvar,
  aggfunc = c("mean", "median"),
  iters = 5
)

Arguments

ds

the data.frame to use

subjvar

Name of the subject-identifying variable

stimvar

Name of the stimulus-identifying variable

pullvar

Name of the movement-direction identifying variable

targetvar

Optional. Name of the stimulus-category identifying variable

rtvar

Name of the reaction-time identifying variable

aggfunc

The function with which to aggregate the RTs before computing difference scores. Defaults to mean but can be changed to median.

iters

If there are missing values (which is almost inevitable) then multiple imputation will be used to complete the covariance matrix - this argument sets the number of multiple imputations to be used.

Value

Exports a list containing a data.frame with stimulus-specific bias scores, indicated in the column names, a covariance matrix of that same data, and a data.frame indicating to which stimulus category each stimulus belongs.

Examples

ds<-aat_simulate(biasfx_jitter=40,nstims=16)
ds$stim<-paste0(ds$stim,"-",ds$is_target)
aat_stimulusscores(ds,"subj","stim","is_pull","is_target","rt")
#> $data
#>          0-1-0      0-10-0      0-11-0      0-12-0     0-13-0     0-14-0
#> X1  -25.794669  250.058250  408.430744   34.171891  147.88947  192.73987
#> 2   142.347786  112.175551   25.519086   47.967209  127.78077  133.48141
#> 3    83.783356  115.139647  164.695074  -12.765392   88.19451   11.07300
#> 4   -15.808923   82.286489 -154.268097   46.580550    6.15276   41.78482
#> 5   187.948863  -14.219629   72.448970  149.516390  -98.87753 -215.09680
#> 6   -50.371464    5.728330  101.956461  -71.444230   16.41364  130.85832
#> 7   127.748777  262.194417  536.307703  -33.610281   51.37162   24.15792
#> 8    45.471332   34.547644   71.105979  121.486956  -59.97819  318.16302
#> 9    88.303667 -121.137192 -109.639513  -25.115826   16.85966  184.39806
#> 10 -141.990482 -146.418257   29.414507 -102.441961   66.36472  -38.29776
#> 11   38.586052  228.774818   -6.938402  -38.641923  196.10674  145.95241
#> 12   68.719341  170.991734 -337.838186 -106.800294  114.16928 -121.56797
#> 13  -17.016396  -62.310477   20.026116    5.861627   17.00488   38.99703
#> 14 -282.335284  205.188838   -4.256466  363.260329 -115.61868  -83.48265
#> 15   -2.886801  -31.954982 -155.796667  -17.737554   79.15463   75.06836
#> 16 -159.563201   59.653702  -26.098250  -21.831892 -107.27742   71.54553
#> 17  111.258883   25.672925  -41.389039   89.587751  143.14640  122.14422
#> 18 -159.934160   58.201048  108.013908  143.403856  137.97752  102.29732
#> 19  204.412924  -51.510152  132.205051   34.178832  -50.57147  -13.94649
#> 20  139.932924  -44.517829  -51.770774   42.436517   77.81057  108.74424
#> 21 -115.836711 -126.988021   -2.719093   30.705703 -148.20338    5.13583
#> 22  -11.640686 -136.147299   -4.982621  -43.197610  323.51102  -20.41807
#> 23   17.180144  149.676638   27.891826 -135.504652   17.06477 -209.33863
#> 24   77.336986   18.486593  167.781505  101.345746  133.25821   48.75119
#> 25  238.276443 -228.967069  116.809028   28.649767   69.90238   80.66350
#> 26  353.990831  138.979627  157.828209 -167.839170   61.55252 -133.18868
#> 27  155.549668    6.485805  112.459096  -20.536717   33.28150   23.19970
#> 28 -241.800349 -242.602691  333.549179  217.702855  189.87086 -109.34103
#> 29 -108.438844  -85.594726  116.818142 -270.490319 -172.54873  102.05584
#> 30   79.276255  132.326682  -19.261156 -143.208267  103.64173  281.78407
#> 31   -7.775719  -24.505864   47.504208   16.569368  130.09917   46.55241
#> 32  348.686257   55.725567 -318.936503   24.920074   52.75208  127.22921
#> 33   73.932029  135.981111  -37.833704  -57.301302  264.27302   20.14706
#> 34  -67.046163   86.900983  111.250178  -96.621723  133.91869  213.45703
#> 35 -127.306430 -124.563776 -136.401477   76.046020  109.21890 -127.35564
#> 36 -180.958332  164.430686 -319.115769  -58.167534   54.25169  317.50797
#>          0-15-0      0-16-0      0-2-0       0-3-0       0-4-0       0-5-0
#> X1  300.3671774  130.678130  -90.38662   42.626756   -2.918416 -144.715802
#> 2   130.6168151   92.135090   90.49659 -133.088708   85.232190   26.430181
#> 3    65.4687650   58.055869   50.95406   -9.108983   96.624270  104.798556
#> 4   124.2421300  112.345849   26.18928  166.645584  -78.061413   60.962144
#> 5     8.3747148  113.433098   73.22329 -127.485112  156.714096   30.425465
#> 6   -62.6402931  105.538422   77.25633   11.977294   48.178458  130.930964
#> 7   308.0340046  119.472122   12.37195  130.333783   20.699414  116.771761
#> 8    51.9330366  -28.190529  125.23609   77.079955  166.172005  162.178240
#> 9     0.5615297 -232.779568  -84.78001  -31.462224 -209.513560  -57.332084
#> 10  -27.2238953   13.452582   59.37896   22.779929  -43.338653  144.484379
#> 11  -54.4300070  -51.590269   60.96239  146.993408   71.769047  190.949561
#> 12   72.4799631  -23.577181  174.20197   73.413692 -219.492333  -47.835157
#> 13   22.3907317   46.135081  116.99764   29.136925  113.484442   24.746231
#> 14  194.6726408 -122.252489   95.29649  198.610798  -50.373513   -4.053058
#> 15  -14.9512256  -46.957262   52.04915   97.086300  -66.063572  -25.577515
#> 16  104.2668519   40.828018  -64.66236  -48.581397   -2.174245   74.998921
#> 17   -9.3348573  -40.664419  183.74257   66.310844  -22.555709  148.494612
#> 18   23.3850600    4.756358  -81.57936  -38.131945  -68.432652  171.768146
#> 19   72.2944620  -55.090384   -1.89828   57.195766  -62.226938  -92.475771
#> 20   17.1755157   40.923499   75.47640   15.053252   28.998576  -22.460881
#> 21   24.5537315  -66.855407  -34.66649  -92.023519   94.823755  -49.104362
#> 22   19.2520858   69.736116  294.35314  135.639811  114.156856  204.705813
#> 23  -90.3336506   17.322609  -86.13117 -217.076965 -118.876696 -144.063031
#> 24  -76.6692900   70.459070   88.96633  -43.043526  225.626236   55.264719
#> 25  195.1302223  287.183892 -187.16395   18.707339 -136.049474   36.817353
#> 26   52.0658646  182.006539   32.84962   52.580684 -212.908946   82.874626
#> 27   51.2207317  176.027741  144.29960  303.307778  171.696663   -1.195951
#> 28 -153.6556347  203.634323  115.53599 -348.784748   44.653715  -65.854741
#> 29  143.7313380  175.107712  491.85706 -231.140249  174.001477 -222.566084
#> 30  -20.3052378  172.979676  197.92783   53.552955  207.055104   15.521039
#> 31  138.5474573  100.508051   60.26407   78.639657  125.410547  162.667326
#> 32  246.9449828   74.817341  -80.79892  281.194164 -277.712400 -156.476962
#> 33  -11.0772554  103.771113  -23.67339  141.978059   62.983951  -11.744546
#> 34    8.0185443   16.866799  -83.78510  -20.594141  135.709591    8.716908
#> 35   23.0510090 -330.789717  142.21931   22.861952   19.887128  -67.455875
#> 36  -70.1975135   34.997909   48.23958  287.197530  -96.950052  110.378097
#>           0-6-0       0-7-0      0-8-0       0-9-0        1-1-1      1-10-1
#> X1   15.6554741  133.102219  -52.13295  -57.658888 -116.9457111   31.822482
#> 2    53.3199169   74.050724  177.65211   82.848486  160.3421749  190.390627
#> 3    13.1419314   -6.587839  -21.10452   55.612762  137.2013391    1.421965
#> 4     2.9446031  -12.961951   56.31640   -5.221158  -41.4045545  -14.225759
#> 5    82.2498428   76.742308  134.65141  -35.232540   47.9033281  126.401263
#> 6    22.6309068    8.286868  -52.47090  -98.930358   99.4255435  -13.924068
#> 7   105.4334771  116.806001   50.94697  -77.744264  142.6078378  111.740325
#> 8   236.8087772   70.269893  108.35387  128.803190   61.6804002   41.071279
#> 9    44.7214410 -220.200289   72.52407 -142.361493  -86.6286681  -39.066791
#> 10  193.0334805  140.508468 -344.65870 -116.176180   62.9495185  -63.748610
#> 11  183.4321921  106.494893  -74.89410  -70.754715  -29.6177819  119.339321
#> 12  -79.6196054    2.268722   15.80707   26.719776  213.6850418   -8.840560
#> 13   51.6328675   10.476824   89.69434   40.051914  138.4637338   33.776102
#> 14  215.9147935  -48.617947   74.19185  -81.518839  252.9277302 -210.378086
#> 15   73.4321013  -32.081190   32.34336   81.009634  -63.1398619  -38.047273
#> 16   -0.2847516  -12.919647  120.31714  -26.753679   77.4160557   50.712604
#> 17   -5.1743067  225.063812   12.33541  -56.000226    7.2815292  178.789211
#> 18  134.3173045   74.022536  -22.03461   94.102913   81.1184592   36.512021
#> 19   20.6544871  -34.778586   99.54796 -114.762561   25.4945671  146.558095
#> 20  -53.5413903   59.108711   39.28305   34.665965   -1.8820606  114.737491
#> 21   65.3958227    5.391312  -83.09663   25.146371  -81.2601758  138.142528
#> 22  167.0199454   50.141868   23.76474  299.322698   -4.0622724   20.439584
#> 23  227.7957075 -116.089059 -118.81576  -62.926682  312.4744712 -116.333589
#> 24  122.8055744  -22.372750   85.84943  303.997735    0.6199945  202.995386
#> 25  -18.8975405   20.679663  100.23307 -148.397898  -97.1339696  238.109095
#> 26 -171.3945523  130.845182  -73.38920  -79.785082  228.4119534 -258.298882
#> 27   68.1067331   34.596533  109.68055  141.360392  -33.8929892   -5.605687
#> 28   12.4428887  209.888676  -18.32142 -114.052317  101.8136629  156.336406
#> 29 -288.2790490 -229.491441  255.44306  -86.877656  361.0649287  176.879322
#> 30   14.9349624    0.104228  155.20841   16.057259   12.7514083 -127.799680
#> 31   98.1378235  -15.187959   33.43242   66.350729  164.5068538  -79.833804
#> 32  -27.1929955  175.836467 -114.10656   88.449751   -1.2468355  304.634649
#> 33 -105.4375594   59.794790   24.12503  178.519716   90.4682365  121.266698
#> 34  128.1309775  115.822952   93.97215  -30.029068   46.0924861   87.542309
#> 35  -11.8013967 -137.693025  -17.01446  -70.863513 -256.9502608   95.525269
#> 36   84.4389654    6.437827   90.41175  202.100207  141.6938306  103.909764
#>         1-11-1      1-12-1      1-13-1      1-14-1      1-15-1       1-16-1
#> X1  202.813692  169.037901  208.928081  128.770270  186.312768  182.4616744
#> 2   203.699583  101.523903  199.808363   92.869758    5.916510  100.3370612
#> 3   180.884101  102.000404  117.181196   20.546843   63.923439   89.1518148
#> 4    -9.496553  -79.063391   95.676145   52.745476  108.472278   57.7987114
#> 5    80.212030  114.540403   44.147765   51.562838 -133.842317  125.7488015
#> 6   112.949191  196.252388   35.407335   35.216552   78.923383   55.1087448
#> 7   217.015605   30.607436   71.446548  108.408924   49.937552   58.5994609
#> 8    23.236449   85.068918  159.535119  159.247828  270.562998  140.6330573
#> 9   105.990737 -125.283733   47.000453  128.162836   19.871203 -102.6198640
#> 10  -36.022900    2.955489 -190.685767  104.288448    5.705881   13.2515623
#> 11   24.253456  148.828223  -17.387897  188.949606 -154.671418   90.9475637
#> 12 -153.353723  123.483384  241.418682  140.125328  -10.517330  104.7682568
#> 13  112.865091   75.140313  133.037684  186.337297   41.252102   58.6321829
#> 14  -55.412142  -54.106194   86.923326  -36.766838   53.096542  -27.2295212
#> 15   12.866451  145.757922   69.764005   44.770729 -147.575321  -53.1699922
#> 16  -51.306207  114.110968  -44.208246  -33.025425  -27.664549  -26.5693450
#> 17   10.534859  197.284591   -1.732608  -77.187400  101.232198  190.0753081
#> 18  143.568122  -52.727541  232.842458  130.328631  -96.852733   -5.9277093
#> 19   64.062650   29.979051  135.939309   70.039155   55.042207   71.1086189
#> 20  219.008285   53.291623  166.308273   58.975230   78.917924  158.5996665
#> 21  271.498900   -1.054737  134.374092   21.743298  -30.222709   62.6332872
#> 22   42.539441  330.002456   92.184129  -83.863179   94.633967  -11.4273061
#> 23  253.233212  -92.239248   31.815925 -138.632052   -1.745724  -68.0572413
#> 24   50.540191   98.115767   47.666972  -86.875443   57.907652  -37.4835920
#> 25  -33.096637   60.032252  -52.311055   89.930714  201.622788  -85.0577862
#> 26  197.182543  112.191850  207.084609   68.828810  163.298176  -78.9096700
#> 27   -2.960008   30.672346  -19.179588    1.965374   53.313104  116.7478236
#> 28   15.248421 -388.952153  -77.282425   26.410384 -120.975789 -138.8205587
#> 29  -31.813965   97.587384 -252.412286  -25.194681   80.836067   48.1929167
#> 30  350.492346  118.666603   32.054866  254.074923  215.286742  104.0035097
#> 31  -90.958708  -82.848774 -160.477881  219.804245  292.486708   75.4133496
#> 32  300.975258  200.269917 -117.009339  -68.766537  -32.038576  -73.7762458
#> 33  236.947966   72.797227   34.526258   36.427139  124.115421   23.3063681
#> 34  245.610705  141.759580  -12.480079 -154.772549   76.252716   -0.1401565
#> 35   89.334024  206.511497   -8.703988  295.733720  -18.195848  200.7123875
#> 36  123.712173  525.639948  195.746158  168.547653   87.712443  188.7812460
#>         1-2-1       1-3-1        1-4-1       1-5-1       1-6-1       1-7-1
#> X1 307.807290  -35.554261   49.9727907 -244.912105 -102.561044   85.172567
#> 2   90.943181  136.672485  235.6024589  142.752915  177.791730   84.347907
#> 3  179.725036  164.301882   79.3277925  212.282660  149.159126  158.383729
#> 4  -50.885975  124.409731  159.1311111  209.264322   50.829692  -79.454934
#> 5  -40.657855   96.385661   53.0834127   97.555393   92.724258   39.137141
#> 6  122.718433  204.140442   80.6646399   68.683248  103.313240  120.373980
#> 7  186.293037  -14.452738   -0.1788008  210.755586  223.205923   53.751579
#> 8  103.917634  -18.014830   -0.3883747  140.690496   73.962296  229.673434
#> 9   69.575881   61.864276 -151.7445505   10.310952  -85.962425    6.092296
#> 10  13.131605    8.773810 -180.5831257  -83.609214 -209.643417  -19.869133
#> 11 295.897802   42.642000   11.4229262 -147.181298   34.132911   96.651220
#> 12 158.047089  -19.120038  137.3809579   76.401017  -25.815658  396.178189
#> 13  83.046024  -93.361840   11.1295874   83.019270   77.624143   34.052556
#> 14 -80.239589 -103.660400   31.6555534  -58.217141 -317.066999  219.034667
#> 15 -95.981906    9.513279  147.9990948   19.818746 -145.223298  -40.900275
#> 16  95.972673  -13.405344  -25.6610835  -52.670302   53.030015   44.458821
#> 17  28.401018  148.063674  126.7699388  -84.718023  131.707576  255.967136
#> 18 126.149000  141.016146  109.8816308    2.989782  188.051167   86.079097
#> 19  22.697238  205.401803   10.5545726  161.860379 -218.447238  -47.841260
#> 20  34.923685   70.660226   79.0064272   -3.242321   95.047272  -47.076544
#> 21 237.618876  237.416993  110.9870269 -115.073971   40.570560 -106.749077
#> 22 146.640791  -76.132732   68.1407382  -54.832891   -4.768456   61.560839
#> 23  72.773664   74.853832   65.8934523   83.419584 -266.235455   23.737612
#> 24  42.047424  -35.056645   48.6270778  266.697065  -34.895685  147.040353
#> 25  43.959360  211.557845  384.2708270  150.802320   64.992651   88.753764
#> 26 295.308855  156.486703   90.8736387  445.831076  182.888197    2.217156
#> 27  44.813004   80.084132  120.5930933   59.135686    1.028968  -52.037092
#> 28   4.361177  222.356198   94.0564143  -30.813837 -154.073250 -160.998683
#> 29 393.714659  267.400381  -11.3346370  197.932583 -126.358609  210.091873
#> 30  57.241755  -53.531147  174.0129119 -106.310784  120.302569  171.971996
#> 31 193.170332   76.372554   42.1998579  223.952603  206.318974  245.436461
#> 32  73.333256   43.713214  149.1986925  110.179878  -50.670679  354.890234
#> 33 101.786476   38.403732   54.2216766  134.904714  116.625934  141.248943
#> 34 129.743884   37.398210  215.5924804   26.681589   59.614425   -6.837452
#> 35 127.240996  172.987776   10.9640752  -46.681356  408.808013  -30.189134
#> 36 -63.553114   39.831785   80.6047943  103.061083  234.222410   74.896437
#>         1-8-1       1-9-1
#> X1  -45.59695   35.763318
#> 2   133.57796  167.299404
#> 3   196.33354  235.394789
#> 4    13.53634   91.324788
#> 5  -237.43847   90.360014
#> 6    49.81177  103.548637
#> 7   120.61789  134.836132
#> 8    37.24168   23.783723
#> 9     9.19433    8.702953
#> 10  -12.15234  252.995594
#> 11   36.73279   22.629972
#> 12  182.94196  205.737073
#> 13  -43.19751   31.908313
#> 14  -52.49133 -160.516454
#> 15   47.65885  -75.532494
#> 16  -72.96524   74.756194
#> 17   25.94293  104.828563
#> 18    6.01893   30.172895
#> 19  152.91442  137.930217
#> 20  109.99190  172.238472
#> 21  108.99285   36.045668
#> 22  -89.22358  -83.240130
#> 23  -95.53802  -88.449477
#> 24   30.80917  211.333809
#> 25   33.35967   16.662977
#> 26  -27.60011  144.527678
#> 27   74.28084  -53.692245
#> 28  218.54465 -209.531701
#> 29  -24.40239   55.393734
#> 30  101.92146  110.188107
#> 31  123.89189  -85.716330
#> 32 -137.32385   15.890969
#> 33  159.31722  124.147778
#> 34   76.35875  -59.517627
#> 35  120.26201  107.648994
#> 36  -31.59777   10.652346
#> 
#> $covmat
#>              0-1-0     0-10-0      0-11-0       0-12-0     0-13-0     0-14-0
#> 0-1-0  22449.91424  1999.9493    49.81853 -3959.354043  1625.8836  -507.8565
#> 0-10-0  1999.94930 17038.0545   943.70466 -1520.957518  1065.6112  2905.5945
#> 0-11-0    49.81853   943.7047 32016.75733  2043.222640   616.7134 -2285.1081
#> 0-12-0 -3959.35404 -1520.9575  2043.22264 13185.320662  -580.4478 -2075.9024
#> 0-13-0  1625.88356  1065.6112   616.71345  -580.447751 11879.0536  1007.0004
#> 0-14-0  -507.85646  2905.5945 -2285.10809 -2075.902359  1007.0004 17284.2425
#> 0-15-0  4008.77992  4052.6715  4146.70643   919.498999 -2582.7145  1206.8346
#> 0-16-0  5190.25798   435.2315  8756.76591 -3279.575510  1404.2950   752.1966
#> 0-2-0  -3199.93470 -2238.0576 -1480.06930 -2702.482393  -382.7221  -726.9000
#> 0-3-0   4777.66420  6863.3100 -9156.59461   306.609952  2385.9562  5401.7235
#> 0-4-0  -4105.98091 -1507.6153  8991.77873   617.121024   281.5493  2870.7369
#> 0-5-0  -1034.05463  1348.3636  1278.79311  1518.290538  3743.5108  2663.2343
#> 0-6-0  -4697.20228  1647.9117  1552.60811  5076.086111  1090.4207   720.6299
#> 0-7-0   3131.90163  2150.6237  4460.65147  2763.837510  4715.0955   901.8392
#> 0-8-0   1127.46522   312.7080   599.62643   629.284017 -2876.0211  4075.3368
#> 0-9-0    660.98045  1753.4566 -5295.22685   549.919021  4621.5464  3305.7185
#> 1-1-1  -2043.32841  5338.8682  2236.65293 -3545.585783 -3231.1233 -3502.0138
#> 1-10-1  3339.84334 -4303.8500    91.95534  1936.193427  1005.1661  3414.1816
#> 1-11-1  5192.52100  4464.8889  2918.46610 -2763.916933  1625.2161  3425.5859
#> 1-12-1  3220.56757  5264.5771 -9529.94189 -5356.458412  2157.3234  7054.2255
#> 1-13-1  2106.47978  6395.0779  -860.83868  2755.108855  1237.2999  1094.4262
#> 1-14-1  -644.57217   276.4272 -1598.83476    -3.571094   933.1104  2812.1640
#> 1-15-1  3261.53373  1372.1713  3009.09498 -2271.585837   606.0587  5370.6997
#> 1-16-1  -202.30038  3309.4775 -1564.36943     3.790426  -185.2878  3140.8823
#> 1-2-1   1474.82899  2675.6791  7026.83066 -6237.725465   638.2526   863.5330
#> 1-3-1   1337.16204 -5474.0298  2282.01880  -979.146352 -2274.5035 -2216.7753
#> 1-4-1   4452.59580   509.3400 -1504.19301   653.509104  2313.7683  1507.3714
#> 1-5-1   9807.51648  1077.9399  2141.38619 -3298.676022 -2006.8567 -3221.6784
#> 1-6-1   3748.70848  2186.9243 -1289.07590 -1386.460812  4205.4622  4428.7511
#> 1-7-1   3655.92935  6271.1166 -6275.51292  -829.488057   752.1122  3543.8767
#> 1-8-1   -793.01847  -746.9350  2806.09785  -185.460048  3295.2151   924.2413
#> 1-9-1   6061.26529  2334.6360  -638.73826 -3731.658153  -163.1781   122.4824
#>            0-15-0      0-16-0        0-2-0      0-3-0        0-4-0      0-5-0
#> 0-1-0   4008.7799  5190.25798 -3199.934697  4777.6642 -4105.980908 -1034.0546
#> 0-10-0  4052.6715   435.23153 -2238.057613  6863.3100 -1507.615314  1348.3636
#> 0-11-0  4146.7064  8756.76591 -1480.069302 -9156.5946  8991.778734  1278.7931
#> 0-12-0   919.4990 -3279.57551 -2702.482393   306.6100   617.121024  1518.2905
#> 0-13-0 -2582.7145  1404.29495  -382.722087  2385.9562   281.549300  3743.5108
#> 0-14-0  1206.8346   752.19659  -726.899958  5401.7235  2870.736929  2663.2343
#> 0-15-0 11180.4231  2259.89823 -2076.012019  4425.1055 -2738.942850 -2162.4887
#> 0-16-0  2259.8982 14234.25141   687.163050 -1571.4538  3045.433290   503.0995
#> 0-2-0  -2076.0120   687.16305 15517.003784 -1421.7429  7420.352680   149.5642
#> 0-3-0   4425.1055 -1571.45385 -1421.742889 19323.7454 -3809.887490  4379.9617
#> 0-4-0  -2738.9429  3045.43329  7420.352680 -3809.8875 16575.191806  2831.5960
#> 0-5-0  -2162.4887   503.09954   149.564169  4379.9617  2831.595967 11370.2927
#> 0-6-0  -2214.4188 -3907.09243 -3523.291909  2128.1006  2325.359531  5229.4335
#> 0-7-0    381.2237  4393.44482 -2418.975424  2280.7987  -268.408302  3967.7197
#> 0-8-0   1896.0438  2246.46444  4321.399695  -958.8353  5397.278550 -1567.5319
#> 0-9-0  -1968.2195  1253.72894  2815.730608  5649.8702  4986.040543  3356.8861
#> 1-1-1    169.5210  4547.46488  4643.630162 -4211.3181  -285.730898  -461.2005
#> 1-10-1  1258.6178  1545.65727  -613.877652 -2130.1465  2023.075405 -2381.5499
#> 1-11-1   -28.7267  1083.72343 -4841.832989  -817.4361   591.798295 -2901.0084
#> 1-12-1   793.6411 -1235.73030  3344.735513  9958.6605   474.337332  3224.8721
#> 1-13-1   190.9012 -2228.18511 -2988.429330  2615.2274 -2715.846335  1753.5136
#> 1-14-1   636.8377 -2278.24876   618.387221  2413.0572   150.775681  2135.0077
#> 1-15-1  3612.9395  4045.49952   905.286743  3804.0914  2282.579730  1742.8472
#> 1-16-1   870.1430 -1914.94349  3563.906945  3407.9615  4078.295214  1394.5943
#> 1-2-1   3309.8525  1691.94359  2510.964210 -3081.9795  2208.476670  -900.5102
#> 1-3-1  -1234.4440  1440.11775     7.459427 -6117.3096  -961.421799 -2207.3100
#> 1-4-1   1443.4575  5324.71236 -2222.362515   771.7411  -577.331768  -530.3609
#> 1-5-1   1816.0543  5662.74527   491.296321   333.8810 -1505.910836   852.9776
#> 1-6-1    508.0705   137.24909   145.613453  3703.4115  3795.647605  7305.5617
#> 1-7-1   4263.2237   490.26850  3544.460479  4783.7651 -1469.984275   900.1152
#> 1-8-1  -1259.6255  -520.44583   808.311361 -1306.6484  2073.749611   154.8999
#> 1-9-1    735.0597   -92.02117   727.558879   344.6112    -7.119327  1562.0400
#>             0-6-0      0-7-0      0-8-0      0-9-0      1-1-1      1-10-1
#> 0-1-0  -4697.2023  3131.9016  1127.4652   660.9804 -2043.3284  3339.84334
#> 0-10-0  1647.9117  2150.6237   312.7080  1753.4566  5338.8682 -4303.84998
#> 0-11-0  1552.6081  4460.6515   599.6264 -5295.2269  2236.6529    91.95534
#> 0-12-0  5076.0861  2763.8375   629.2840   549.9190 -3545.5858  1936.19343
#> 0-13-0  1090.4207  4715.0955 -2876.0211  4621.5464 -3231.1233  1005.16605
#> 0-14-0   720.6299   901.8392  4075.3368  3305.7185 -3502.0138  3414.18163
#> 0-15-0 -2214.4188   381.2237  1896.0438 -1968.2195   169.5210  1258.61777
#> 0-16-0 -3907.0924  4393.4448  2246.4644  1253.7289  4547.4649  1545.65727
#> 0-2-0  -3523.2919 -2418.9754  4321.3997  2815.7306  4643.6302  -613.87765
#> 0-3-0   2128.1006  2280.7987  -958.8353  5649.8702 -4211.3181 -2130.14647
#> 0-4-0   2325.3595  -268.4083  5397.2785  4986.0405  -285.7309  2023.07541
#> 0-5-0   5229.4335  3967.7197 -1567.5319  3356.8861  -461.2005 -2381.54986
#> 0-6-0  11916.8531  1584.2241 -3032.3125  2355.6550 -1461.1841 -2625.97956
#> 0-7-0   1584.2241 10353.8871 -3927.8653   843.4832 -1279.6263  2666.11146
#> 0-8-0  -3032.3125 -3927.8653 11167.0588  2151.9299  1678.7066  2678.52144
#> 0-9-0   2355.6550   843.4832  2151.9299 13451.1241  -484.8192  1979.39126
#> 1-1-1  -1461.1841 -1279.6263  1678.7066  -484.8192 15918.3098 -4579.19997
#> 1-10-1 -2625.9796  2666.1115  2678.5214  1979.3913 -4579.2000 14883.52486
#> 1-11-1  -539.0163  2111.8407 -1481.5199  1338.7581 -1337.4774  1210.23098
#> 1-12-1 -1255.7843   719.2876  1670.6292  6844.6827 -2454.1278  3039.84110
#> 1-13-1  1427.1548  1218.2534  1086.6026  3704.0447  -466.2642 -2823.68139
#> 1-14-1  -449.1677 -1199.2846   221.5211 -1567.0134 -3338.5655 -1983.26868
#> 1-15-1 -1534.8548  -417.0894  2495.3540  1477.4504  1141.2753 -2921.40513
#> 1-16-1  -340.4185   585.1701  1627.7275  1475.4595 -2125.8355  1360.32242
#> 1-2-1  -4299.3557  -577.2761  -861.0269 -1646.3457  2415.2700   231.60566
#> 1-3-1  -5034.2463 -1030.5867   -48.9962 -4529.6029  -287.1023  3866.91123
#> 1-4-1  -1844.6168  2573.0165  2880.5420  1618.7913 -1289.2789  3088.23317
#> 1-5-1  -4675.9641 -1992.0669  3578.4394  2456.0398  7553.2364  -500.10776
#> 1-6-1  -3308.7543  2157.0541  2851.8813  4047.7675 -4339.4301  2970.50007
#> 1-7-1  -1475.3672   281.5965  1540.3015  3029.6377  5435.3243   901.21272
#> 1-8-1  -2023.3137   207.9302   498.3141  -263.3140  -893.7194   946.50245
#> 1-9-1  -3820.3898   406.6533 -1249.2474   241.9451  -542.6983  1805.65741
#>             1-11-1     1-12-1     1-13-1       1-14-1     1-15-1       1-16-1
#> 0-1-0   5192.52100  3220.5676  2106.4798  -644.572170  3261.5337  -202.300376
#> 0-10-0  4464.88887  5264.5771  6395.0779   276.427166  1372.1713  3309.477488
#> 0-11-0  2918.46610 -9529.9419  -860.8387 -1598.834763  3009.0950 -1564.369430
#> 0-12-0 -2763.91693 -5356.4584  2755.1089    -3.571094 -2271.5858     3.790426
#> 0-13-0  1625.21608  2157.3234  1237.2999   933.110370   606.0587  -185.287816
#> 0-14-0  3425.58593  7054.2255  1094.4262  2812.164030  5370.6997  3140.882346
#> 0-15-0   -28.72670   793.6411   190.9012   636.837694  3612.9395   870.143010
#> 0-16-0  1083.72343 -1235.7303 -2228.1851 -2278.248762  4045.4995 -1914.943490
#> 0-2-0  -4841.83299  3344.7355 -2988.4293   618.387221   905.2867  3563.906945
#> 0-3-0   -817.43610  9958.6605  2615.2274  2413.057190  3804.0914  3407.961471
#> 0-4-0    591.79829   474.3373 -2715.8463   150.775681  2282.5797  4078.295214
#> 0-5-0  -2901.00842  3224.8721  1753.5136  2135.007661  1742.8472  1394.594291
#> 0-6-0   -539.01627 -1255.7843  1427.1548  -449.167657 -1534.8548  -340.418521
#> 0-7-0   2111.84065   719.2876  1218.2534 -1199.284621  -417.0894   585.170141
#> 0-8-0  -1481.51992  1670.6292  1086.6026   221.521097  2495.3540  1627.727528
#> 0-9-0   1338.75813  6844.6827  3704.0447 -1567.013420  1477.4504  1475.459539
#> 1-1-1  -1337.47743 -2454.1278  -466.2642 -3338.565529  1141.2753 -2125.835479
#> 1-10-1  1210.23098  3039.8411 -2823.6814 -1983.268681 -2921.4051  1360.322425
#> 1-11-1 15113.89630  2430.1461  3949.2700  -841.611503   543.1972   317.455303
#> 1-12-1  2430.14605 21374.2463  3603.0329   740.676334  2167.1137  6364.135384
#> 1-13-1  3949.26997  3603.0329 14248.8830  2405.976347   299.5921  2920.144869
#> 1-14-1  -841.61150   740.6763  2405.9763 11364.469251  1776.1622  4123.600379
#> 1-15-1   543.19716  2167.1137   299.5921  1776.162245 11666.8691  2232.119631
#> 1-16-1   317.45530  6364.1354  2920.1449  4123.600379  2232.1196  8175.185840
#> 1-2-1   2110.20203  1373.0827  -837.1555  1325.926438  2280.8222  1116.684549
#> 1-3-1    217.03296 -3309.6528 -2286.4635  -522.990490 -1240.7789  -242.483765
#> 1-4-1   2164.89624  1918.2456  2249.4581 -1719.775338  1417.8687  -329.827799
#> 1-5-1   -480.85736  -839.1360   700.4830 -1294.598492  4289.8384 -2696.416940
#> 1-6-1   4444.64444  8693.2562  3915.6941  7720.065565  3547.1967  6650.418320
#> 1-7-1  -2297.94412  5740.6526  -951.4554   157.133919  4188.5425  1805.239748
#> 1-8-1     50.77675 -4166.2160  1280.2058  2936.633751  1332.1555  1286.517943
#> 1-9-1   1776.65247  4296.5019  2668.6630  2365.119825  1065.0843  3621.604631
#>             1-2-1        1-3-1      1-4-1      1-5-1      1-6-1      1-7-1
#> 0-1-0   1474.8290  1337.162039  4452.5958  9807.5165  3748.7085  3655.9293
#> 0-10-0  2675.6791 -5474.029756   509.3400  1077.9399  2186.9243  6271.1166
#> 0-11-0  7026.8307  2282.018797 -1504.1930  2141.3862 -1289.0759 -6275.5129
#> 0-12-0 -6237.7255  -979.146352   653.5091 -3298.6760 -1386.4608  -829.4881
#> 0-13-0   638.2526 -2274.503488  2313.7683 -2006.8567  4205.4622   752.1122
#> 0-14-0   863.5330 -2216.775279  1507.3714 -3221.6784  4428.7511  3543.8767
#> 0-15-0  3309.8525 -1234.444008  1443.4575  1816.0543   508.0705  4263.2237
#> 0-16-0  1691.9436  1440.117750  5324.7124  5662.7453   137.2491   490.2685
#> 0-2-0   2510.9642     7.459427 -2222.3625   491.2963   145.6135  3544.4605
#> 0-3-0  -3081.9795 -6117.309632   771.7411   333.8810  3703.4115  4783.7651
#> 0-4-0   2208.4767  -961.421799  -577.3318 -1505.9108  3795.6476 -1469.9843
#> 0-5-0   -900.5102 -2207.309989  -530.3609   852.9776  7305.5617   900.1152
#> 0-6-0  -4299.3557 -5034.246266 -1844.6168 -4675.9641 -3308.7543 -1475.3672
#> 0-7-0   -577.2761 -1030.586706  2573.0165 -1992.0669  2157.0541   281.5965
#> 0-8-0   -861.0269   -48.996195  2880.5420  3578.4394  2851.8813  1540.3015
#> 0-9-0  -1646.3457 -4529.602924  1618.7913  2456.0398  4047.7675  3029.6377
#> 1-1-1   2415.2700  -287.102307 -1289.2789  7553.2364 -4339.4301  5435.3243
#> 1-10-1   231.6057  3866.911226  3088.2332  -500.1078  2970.5001   901.2127
#> 1-11-1  2110.2020   217.032959  2164.8962  -480.8574  4444.6444 -2297.9441
#> 1-12-1  1373.0827 -3309.652755  1918.2456  -839.1360  8693.2562  5740.6526
#> 1-13-1  -837.1555 -2286.463464  2249.4581   700.4830  3915.6941  -951.4554
#> 1-14-1  1325.9264  -522.990490 -1719.7753 -1294.5985  7720.0656   157.1339
#> 1-15-1  2280.8222 -1240.778920  1417.8687  4289.8384  3547.1967  4188.5425
#> 1-16-1  1116.6845  -242.483765  -329.8278 -2696.4169  6650.4183  1805.2397
#> 1-2-1  12630.9635  2369.345853 -1449.7229   863.2986  3499.6509  2786.9662
#> 1-3-1   2369.3459 10000.901598  2318.4696  3284.6495  2486.1174 -3461.0436
#> 1-4-1  -1449.7229  2318.469610  9946.1223  1885.4358  3477.1050  1095.0663
#> 1-5-1    863.2986  3284.649456  1885.4358 18803.3659  4690.4166  2198.5926
#> 1-6-1   3499.6509  2486.117355  3477.1050  4690.4166 24232.6328  1075.7046
#> 1-7-1   2786.9662 -3461.043594  1095.0663  2198.5926  1075.7046 15735.7160
#> 1-8-1   1225.3219  2839.864797  1631.6008  1043.7994  3364.0961 -1187.7553
#> 1-9-1   1932.6645   890.467505 -1466.7980  3893.0925  5536.0163  2218.4752
#>              1-8-1        1-9-1
#> 0-1-0   -793.01847  6061.265294
#> 0-10-0  -746.93502  2334.636016
#> 0-11-0  2806.09785  -638.738256
#> 0-12-0  -185.46005 -3731.658153
#> 0-13-0  3295.21507  -163.178106
#> 0-14-0   924.24131   122.482405
#> 0-15-0 -1259.62553   735.059733
#> 0-16-0  -520.44583   -92.021174
#> 0-2-0    808.31136   727.558879
#> 0-3-0  -1306.64841   344.611215
#> 0-4-0   2073.74961    -7.119327
#> 0-5-0    154.89992  1562.039985
#> 0-6-0  -2023.31373 -3820.389780
#> 0-7-0    207.93020   406.653276
#> 0-8-0    498.31410 -1249.247440
#> 0-9-0   -263.31396   241.945137
#> 1-1-1   -893.71943  -542.698265
#> 1-10-1   946.50245  1805.657407
#> 1-11-1    50.77675  1776.652466
#> 1-12-1 -4166.21604  4296.501894
#> 1-13-1  1280.20583  2668.663006
#> 1-14-1  2936.63375  2365.119825
#> 1-15-1  1332.15549  1065.084255
#> 1-16-1  1286.51794  3621.604631
#> 1-2-1   1225.32194  1932.664481
#> 1-3-1   2839.86480   890.467505
#> 1-4-1   1631.60080 -1466.797987
#> 1-5-1   1043.79939  3893.092495
#> 1-6-1   3364.09609  5536.016335
#> 1-7-1  -1187.75533  2218.475211
#> 1-8-1   9909.74233  2095.365801
#> 1-9-1   2095.36580 11822.169027
#> 
#> $stimcats
#>      stim cat
#> 1   0-1-0   0
#> 2  1-14-1   1
#> 3   0-2-0   0
#> 4   1-3-1   1
#> 5   1-4-1   1
#> 6  1-11-1   1
#> 7   0-8-0   0
#> 8   1-9-1   1
#> 9  1-16-1   1
#> 10 1-10-1   1
#> 11 0-14-0   0
#> 12  0-5-0   0
#> 13 1-15-1   1
#> 14  0-4-0   0
#> 15  1-5-1   1
#> 16  1-6-1   1
#> 17 1-13-1   1
#> 18 0-10-0   0
#> 19 0-11-0   0
#> 20 1-12-1   1
#> 21 0-16-0   0
#> 22  0-7-0   0
#> 23  1-2-1   1
#> 24  1-8-1   1
#> 25  0-6-0   0
#> 26 0-12-0   0
#> 27 0-13-0   0
#> 28  0-3-0   0
#> 29  0-9-0   0
#> 30  1-1-1   1
#> 31 0-15-0   0
#> 32  1-7-1   1
#>