Compute a minimally biased average of correlation values

This function computes a minimally biased average of correlation values. This is needed because simple averaging of correlations is negatively biased, and the often used z-transformation method of averaging correlations is positively biased. The algorithm was developed by Olkin & Pratt (1958).

cormean(
  r,
  n,
  wts = c("none", "n", "df"),
  type = c("OP5", "OPK", "OP2"),
  na.rm = F
)

Arguments

r
a vector containing correlation values

n
a single value or vector containing sample sizes

wts
Character. How should the correlations be weighted? none leads to no weighting, n weights by sample size, df weights by sample size minus one.

type
Character. Determines which averaging algorithm to use, with "OP5" being the most accurate.

na.rm
Logical. Should missing values be removed?

Value

An average correlation.

References

Olkin, I., & Pratt, J. (1958). Unbiased estimation of certain correlation coefficients. The Annals of Mathematical Statistics, 29. https://doi.org/10.1214/aoms/1177706717

Shieh, G. (2010). Estimation of the simple correlation coefficient. Behavior Research Methods, 42(4), 906-917. https://doi.org/10.3758/BRM.42.4.906

Examples

cormean(c(0,.3,.5),c(30,30,60))

#> [1] 0.2695279